refactoring svg: divido en icons e illustrations
