+++ src/components/ThemeSelector..jsx
@@ -1,6 +1,5 @@
 import useTheme, { LIGHT } from '../hooks/useTheme'
-import MoonSVG from './svg/MoonSVG'
-import SunSVG from './svg/SunSVG'
+import { MoonIcon, SunIcon } from './icons/icons'
 
 const ThemeSelector = () => {
     const { theme, handleChangeTheme } = useTheme()
@@ -11,7 +10,7 @@ const ThemeSelector = () => {
                 onClick={handleChangeTheme}
                 className={`rounded-full text-white bg-gray-100 border-gray p-2 dark:bg-slate-800 dark:border-slate-700`}
             >
-                {theme === LIGHT ? <MoonSVG /> : <SunSVG />}
+                {theme === LIGHT ? <MoonIcon /> : <SunIcon />}
             </button>
         </div>
     )
