+++ src/components/group/GroupHeader.jsx
@@ -2,8 +2,7 @@ import { totalAmountOfExpenses } from '../../logic/logic'
 import { useGroupStore } from '../../store/store'
 import { currencyFormat } from '../../utils/utils'
 import Avatar from '../common/Avatar'
-import MoneySVG from '../svg/MoneySVG'
-import PeopleSelfieSVG from '../svg/PeopleSelfieSVG'
+import { MoneyIllustration, PeopleSelfieIllustration } from '../illustrations/Illustrations'
 import GroupHeaderLoading from './GroupHeaderLoading'
 
 const GroupHeaderInfo = ({ groupName = 'viaje brasil', cantPersons = 1 }) => {
@@ -38,7 +37,7 @@ const GroupHeaderInfo = ({ groupName = 'viaje brasil', cantPersons = 1 }) => {
                     <h4 className='m-0'>Total Gastos</h4>
                     <div className='flex justify-start items-center'>
                         <div className='shrink-0'>
-                            <MoneySVG width={80} height={50} />
+                            <MoneyIllustration width={80} height={50} />
                         </div>
                         <div className='relative'>
                             <h3 className='absolute top-2 '>$</h3>
@@ -61,7 +60,7 @@ const GroupHeaderInfo = ({ groupName = 'viaje brasil', cantPersons = 1 }) => {
 const GroupHeaderImage = () => {
     return (
         <div className='hidden md:block md:absolute md:right-0 lg:right-80 '>
-            <PeopleSelfieSVG width={350} height={300} />
+            <PeopleSelfieIllustration width={350} height={300} />
         </div>
     )
 }
