+++ src/components/group/divisions/DivisionsList.jsx
@@ -2,13 +2,13 @@ import { calculateFinalResult } from '../../../logic/logic'
 import { useGroupStore } from '../../../store/store'
 import Card from '../../common/Card'
 import MoneyAmount from '../../common/MoneyAmount'
-import ArrowSVG from '../../svg/ArrowSVG'
-import PeopleCoffeSVG from '../../svg/PeopleCofeeSVG'
+import { ArrowIcon } from '../../icons/icons'
+import { PeopleDivisionIllustration } from './../../illustrations/Illustrations'
 
 const EmptyDivisionsList = () => {
     return (
         <div className='flex flex-col justify-center items-center p-10'>
-            <PeopleCoffeSVG width={160} height={180} />
+            <PeopleDivisionIllustration width={160} height={180} />
             <h4 className='text-center'>Aún no hay divisiones...</h4>
             <p className='text-center'> Aquí verás cuánto le corresponde pagar a cada integrante</p>
         </div>
@@ -24,7 +24,7 @@ const DivisionListItem = ({ division }) => {
                     <div className='flex flex-col gap-1'>
                         <h5 className='m-0 capitalize w-20 '>{personaFrom.name}</h5>
                         <div className='flex items-center gap-1'>
-                            <ArrowSVG />
+                            <ArrowIcon />
                             <h5 className='m-0 capitalize ml-auto text-gray-400'>{personaTo.name}</h5>
                         </div>
                     </div>
