+++ src/components/group/expenses/CashTypeDetail.jsx
@@ -1,11 +1,11 @@
 import Tag from '../../common/Tag'
-import CashSVG from '../../svg/CashSVG'
+import { CashIcon } from '../../icons/icons'
 
 const CashTypeDetail = () => {
     return (
         <div className='flex gap-2 items-center  '>
             <Tag>
-                <CashSVG width='w-5' height='h-5' /> Efectivo/Débito
+                <CashIcon width='w-5' height='h-5' /> Efectivo/Débito
             </Tag>
         </div>
     )
