+++ src/components/group/expenses/ExpensesListItem.jsx
@@ -2,8 +2,7 @@ import { CREDIT } from '../../../constants'
 import useModal from '../../../hooks/useModal'
 import { useGroupStore } from '../../../store/store'
 import MoneyAmount from '../../common/MoneyAmount'
-import CashSVG from '../../svg/CashSVG'
-import CreditCardSVG from '../../svg/CreditCardSVG'
+import { CashIcon, CreditCardIcon } from '../../icons/icons'
 import RemoveExpense from './RemoveExpense'
 import ViewDetailExpense from './ViewDetailExpense'
 
@@ -22,7 +21,7 @@ const ExpensesListItem = ({ expense }) => {
             className='flex justify-between items-center gap-4 hover:bg-gray-50 rounded-xl mx-2 px-2 dark:hover:bg-slate-600'
         >
             <div className='flex-1 flex items-center gap-10  cursor-pointer' onClick={openModal}>
-                {type === CREDIT ? <CreditCardSVG /> : <CashSVG />}
+                {type === CREDIT ? <CreditCardIcon /> : <CashIcon />}
                 <div className='flex flex-col justify-center items-start w-20 md:w-auto'>
                     <h4 className='m-0 capitalize'>{name}</h4>
                     <h5 className='m-0 text-xs capitalize text-gray-400'> {persona?.name}</h5>
