+++ src/components/group/persons/PersonsList.jsx
@@ -1,7 +1,7 @@
 import { cantExpensesInWhichEachPersonIsIncluded } from '../../../logic/logic'
 import { useGroupStore } from '../../../store/store'
-import AvatarSVG from '../../svg/AvatarSVG'
 import Card from '../../common/Card'
+import { AvatarIcon } from '../../icons/icons'
 
 const colors = ['bg-primary-300', 'bg-red-300', 'bg-yellow-500']
 
@@ -24,7 +24,7 @@ const PersonsList = () => {
                             className='hover:bg-gray-50 rounded-xl p-3 my-2 dark:hover:bg-slate-600  '
                         >
                             <div className='flex gap-5 items-center mb-2'>
-                                <AvatarSVG width={30} height={30} backgroundColor={avatarColor} />
+                                <AvatarIcon width={30} height={30} backgroundColor={avatarColor} />
 
                                 <div className='flex flex-col'>
                                     <h5 className='m-0 capitalize'>{name}</h5>
