+++ src/components/icons/Icons.jsx
@@ -1,0 +1,358 @@
+const fillColors = {
+    default: 'fill-white',
+    black: 'fill-black dark:fill-gray-200'
+}
+
+export const ArrowIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-6 h-6 bg-gray-100 rounded p-1 dark:bg-slate-600'
+        >
+            <path strokeLinecap='round' strokeLinejoin='round' d='M17.25 8.25L21 12m0 0l-3.75 3.75M21 12H3' />
+        </svg>
+    )
+}
+
+export const AvatarIcon = ({ width, height, backgroundColor }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            width={width}
+            height={height}
+            viewBox='0 0 532 532'
+            className={`rounded-full ${backgroundColor}`}
+        >
+            <circle cx='270.75986' cy='260.93427' r='86.34897' fill='#ffb6b6' />
+            <polygon
+                points='221.18982 360.05209 217.28876 320.6185 295.18982 306.05209 341.18982 418.05209 261.18982 510.05209 204.18982 398.05209 221.18982 360.05209'
+                fill='#ffb6b6'
+            />
+            <path
+                d='m216.0374,340.35736l17.03111,3.84802s-13.38821-42.45453-8.84396-46.50766c4.54427-4.05316,15.68007,2.33328,15.68007,2.33328l11.70201,13.1199,14.25394-14.51239s15.47495-19.2421,21.53397-24.6463-3.67319-25.46364-3.67319-25.46364c0,0,89.89185-24.23923,56.44299-67.83968,0,0-19.61093-34.18452-25.99734-23.04871-6.38641,11.1358-14.00162-6.55013-14.00162-6.55013l-23.25381,4.42198s-45.89429-27.06042-89.45331,30.82959c-43.55902,57.89003,28.57916,154.01572,28.57916,154.01572h-.00002Z'
+                fill='#2f2e41'
+            />
+            <path
+                d='m433.16003,472.95001c-47.19,38.26001-105.57001,59.04999-167.16003,59.04999-56.23999,0-109.81-17.33997-154.62-49.47998.08002-.84003.16003-1.66998.23004-2.5,1.19-13,2.25-25.64001,2.94995-36.12,2.71002-40.69,97.64001-67.81,97.64001-67.81,0,0,.42999.42999,1.29004,1.17999,5.23999,4.59998,26.50995,21.27997,63.81,25.94,33.25995,4.15997,44.20996-15.57001,47.51996-25.02002,1-2.88,1.30005-4.81,1.30005-4.81l97.63995,46.10999c6.37,9.10004,8.86005,28.70001,9.35004,50.73004.01996.90997.03998,1.81.04999,2.72998Z'
+                className='text-primary-500'
+            />
+        </svg>
+    )
+}
+
+export const CashIcon = ({ width = 'w-6', height = 'h-6' }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className={`${width} ${height}`}
+        >
+            <path
+                strokeLinecap='round'
+                strokeLinejoin='round'
+                d='M2.25 18.75a60.07 60.07 0 0115.797 2.101c.727.198 1.453-.342 1.453-1.096V18.75M3.75 4.5v.75A.75.75 0 013 6h-.75m0 0v-.375c0-.621.504-1.125 1.125-1.125H20.25M2.25 6v9m18-10.5v.75c0 .414.336.75.75.75h.75m-1.5-1.5h.375c.621 0 1.125.504 1.125 1.125v9.75c0 .621-.504 1.125-1.125 1.125h-.375m1.5-1.5H21a.75.75 0 00-.75.75v.75m0 0H3.75m0 0h-.375a1.125 1.125 0 01-1.125-1.125V15m1.5 1.5v-.75A.75.75 0 003 15h-.75M15 10.5a3 3 0 11-6 0 3 3 0 016 0zm3 0h.008v.008H18V10.5zm-12 0h.008v.008H6V10.5z'
+            />
+        </svg>
+    )
+}
+
+export const CheckIcon = ({ className }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            viewBox='0 0 24 24'
+            fill='currentColor'
+            className={`w-6 h-6 ${className}`}
+        >
+            <path
+                fillRule='evenodd'
+                d='M2.25 12c0-5.385 4.365-9.75 9.75-9.75s9.75 4.365 9.75 9.75-4.365 9.75-9.75 9.75S2.25 17.385 2.25 12zm13.36-1.814a.75.75 0 10-1.22-.872l-3.236 4.53L9.53 12.22a.75.75 0 00-1.06 1.06l2.25 2.25a.75.75 0 001.14-.094l3.75-5.25z'
+                clipRule='evenodd'
+            />
+        </svg>
+    )
+}
+
+export const ChevronDownIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-4 h-4'
+        >
+            <path strokeLinecap='round' strokeLinejoin='round' d='M19.5 8.25l-7.5 7.5-7.5-7.5' />
+        </svg>
+    )
+}
+
+export const ChevronLeftIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-5 h-5'
+        >
+            <path strokeLinecap='round' strokeLinejoin='round' d='M15.75 19.5L8.25 12l7.5-7.5' />
+        </svg>
+    )
+}
+
+export const ChevronRightIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-5 h-5'
+        >
+            <path strokeLinecap='round' strokeLinejoin='round' d='M8.25 4.5l7.5 7.5-7.5 7.5' />
+        </svg>
+    )
+}
+
+export const ChevronUpIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-4 h-4'
+        >
+            <path strokeLinecap='round' strokeLinejoin='round' d='M4.5 15.75l7.5-7.5 7.5 7.5' />
+        </svg>
+    )
+}
+
+export const CloseIcon = ({ color = 'default', width = 25, height = 25 }) => {
+    const fillClassNames = fillColors[color]
+    return (
+        <svg
+            part='svg'
+            xmlns='http://www.w3.org/2000/svg'
+            aria-hidden='true'
+            aria-labelledby='ic-close'
+            focusable='false'
+            viewBox='0 0 24 24'
+            width={width}
+            height={height}
+            className={`h-icon icon-gray icon-static-view-box h-icon--no-custom-width referral-banner__close-icon cursor-pointer ${fillClassNames}`}
+            // dataV21f970be=''
+        >
+            <g>
+                <path
+                    // fill={color}
+                    fillRule='evenodd'
+                    clipRule='evenodd'
+                    d='M19 6.41L17.59 5L12 10.59L6.41 5L5 6.41L10.59 12L5 17.59L6.41 19L12 13.41L17.59 19L19 17.59L13.41 12L19 6.41Z'
+                ></path>
+            </g>
+        </svg>
+    )
+}
+
+export const CreditCardIcon = ({ width = 'w-6', height = 'h-6' }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className={`${width} ${height}`}
+        >
+            <path
+                strokeLinecap='round'
+                strokeLinejoin='round'
+                d='M2.25 8.25h19.5M2.25 9h19.5m-16.5 5.25h6m-6 2.25h3m-3.75 3h15a2.25 2.25 0 002.25-2.25V6.75A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25v10.5A2.25 2.25 0 004.5 19.5z'
+            />
+        </svg>
+    )
+}
+
+export const LoadingIcon = () => {
+    return (
+        <svg
+            className='w-5 h-5 mr-3 -ml-1 text-white animate-spin'
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+        >
+            <circle
+                className='opacity-25'
+                cx='12'
+                cy='12'
+                r='10'
+                stroke='currentColor'
+                strokeWidth='4'
+            ></circle>
+            <path
+                className='opacity-75'
+                fill='currentColor'
+                d='M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z'
+            ></path>
+        </svg>
+    )
+}
+
+export const MenuIcon = ({ width, height }) => {
+    return (
+        <svg
+            part='svg'
+            xmlns='http://www.w3.org/2000/svg'
+            aria-hidden='true'
+            aria-labelledby='actionDots'
+            focusable='false'
+            viewBox='0 0 24 24'
+            className='h-icon icon-primary cursor-pointer d-none d-md-flex fill-black dark:fill-gray-200'
+            width={width}
+            height={height}
+            // dataV68e5bb38=''
+            // block=''
+        >
+            <g>
+                <g>
+                    <circle cx='12.1091' cy='4' r='2'></circle>
+                    <circle cx='12.1091' cy='20' r='2'></circle>
+                    <circle cx='12.1091' cy='12' r='2'></circle>
+                </g>
+            </g>
+        </svg>
+    )
+}
+
+export const MoonIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-5 h-5 text-secondary fill-secondary'
+        >
+            <path
+                strokeLinecap='round'
+                strokeLinejoin='round'
+                d='M21.752 15.002A9.718 9.718 0 0118 15.75c-5.385 0-9.75-4.365-9.75-9.75 0-1.33.266-2.597.748-3.752A9.753 9.753 0 003 11.25C3 16.635 7.365 21 12.75 21a9.753 9.753 0 009.002-5.998z'
+            />
+        </svg>
+    )
+}
+
+export const SunIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-5 h-5 fill-yellow-200 text-yellow-200'
+        >
+            <path
+                strokeLinecap='round'
+                strokeLinejoin='round'
+                d='M12 3v2.25m6.364.386l-1.591 1.591M21 12h-2.25m-.386 6.364l-1.591-1.591M12 18.75V21m-4.773-4.227l-1.591 1.591M5.25 12H3m4.227-4.773L5.636 5.636M15.75 12a3.75 3.75 0 11-7.5 0 3.75 3.75 0 017.5 0z'
+            />
+        </svg>
+    )
+}
+
+export const SorterArrowsIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-4 h-4 '
+        >
+            <path
+                strokeLinecap='round'
+                strokeLinejoin='round'
+                d='M3 7.5L7.5 3m0 0L12 7.5M7.5 3v13.5m13.5 0L16.5 21m0 0L12 16.5m4.5 4.5V7.5'
+            />
+        </svg>
+    )
+}
+
+export const TooltipTriangleIcon = () => {
+    return (
+        <svg
+            className='absolute left-0 -ml-2 bottom-0 top-0 h-full'
+            width='9px'
+            height='16px'
+            viewBox='0 0 9 16'
+            version='1.1'
+            xmlns='http://www.w3.org/2000/svg'
+        >
+            <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
+                <g id='Tooltips-' transform='translate(-874.000000, -1029.000000)' fill='#FFFFFF'>
+                    <g id='Group-3-Copy-16' transform='translate(850.000000, 975.000000)'>
+                        <g id='Group-2' transform='translate(24.000000, 0.000000)'>
+                            <polygon
+                                id='Triangle'
+                                transform='translate(4.500000, 62.000000) rotate(-90.000000) translate(-4.500000, -62.000000) '
+                                points='4.5 57.5 12.5 66.5 -3.5 66.5'
+                            ></polygon>
+                        </g>
+                    </g>
+                </g>
+            </g>
+        </svg>
+    )
+}
+
+export const UserIcon = ({ fillColor }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            viewBox='0 0 24 24'
+            fill='currentColor'
+            className={`w-6 h-6 ${fillColor}`}
+        >
+            <path
+                fillRule='evenodd'
+                d='M18.685 19.096A9.723 9.723 0 0021.75 12c0-5.385-4.365-9.75-9.75-9.75S2.25 6.615 2.25 12a9.723 9.723 0 003.065 7.097A9.716 9.716 0 0012 21.75a9.716 9.716 0 006.685-2.653zm-12.54-1.285A7.486 7.486 0 0112 15a7.486 7.486 0 015.855 2.812A8.224 8.224 0 0112 20.25a8.224 8.224 0 01-5.855-2.438zM15.75 9a3.75 3.75 0 11-7.5 0 3.75 3.75 0 017.5 0z'
+                clipRule='evenodd'
+            />
+        </svg>
+    )
+}
+
+export const ErrorIcon = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            fill='none'
+            viewBox='0 0 24 24'
+            strokeWidth={1.5}
+            stroke='currentColor'
+            className='w-6 h-6'
+        >
+            <path
+                strokeLinecap='round'
+                strokeLinejoin='round'
+                d='M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126zM12 15.75h.007v.008H12v-.008z'
+            />
+        </svg>
+    )
+}
