+++ src/components/illustrations/Illustrations.jsx
@@ -1,0 +1,1530 @@
+export const MoneyIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 740.67538 597.17519'
+        >
+            <path
+                d='M651.61162,679.71942c-5.12158-7.17175-12.05469-12.41919-20.771-14.38311-7.60058-1.71259-15.416-1.42072-23.14209-1.36115l-97.69238.7533c-8.3667.06451-16.38867,7.31311-16,16,.38574,8.61358,7.03174,16.06915,16,16q42.92578-.331,85.85108-.662c6.74658-.052,13.49707-.17371,20.24414-.15155,1.51513.00494,3.02392.08484,4.53759.102.3667.00415.63135.00091.81836-.007.04688.01428.08643.02734.14746.04407.78956.21679,1.59961.36621,2.394.57214-.27832.0083.51513.5033.91357.671.05469.108.17139.29938.36621.59515.272.51,1.07715,2.05866,1.11035,2.00989a71.76183,71.76183,0,0,1,2.98682,9.39289c1.85938,7.04718,2.26807,8.52026,2.53565,15.13195.33837,8.352,7.11718,16.39783,16,16,8.3789-.3753,16.36279-7.04236,16-16C663.3001,709.33283,660.65068,692.37738,651.61162,679.71942Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M515.74883,715.82494c-1.835-11.11963-3.63477-22.24637-5.50537-33.35952-1.898-11.27417-5.77442-21.59812-14.54786-28.9975-2.34423-1.97688-4.7749-3.82268-7.21582-5.66276-.85839-.6468-1.71533-1.29456-2.56347-1.95566-.05078-.03952-.083-.0643-.127-.09861-.09326-.0802-.20507-.17551-.35546-.30272-3.65674-3.0948-7.19-6.32723-10.55567-9.76318q-4.48315-4.57729-8.57471-9.545-1.02246-1.24071-2.01855-2.50366c-.11426-.14481-.72949-.93992-1.13477-1.45875-.377-.51927-1.06591-1.457-1.18164-1.62073q-1.1638-1.64569-2.28613-3.322a156.12618,156.12618,0,0,1-13.187-24.11549q-.54713-1.25119-1.06836-2.51707c-.09326-.25754-.21387-.58834-.374-1.02876-.74121-2.03833-1.481-4.07418-2.15869-6.13665q-2.15992-6.57108-3.71387-13.3362-.76832-3.34049-1.3833-6.7149-.31715-1.73529-.59326-3.47777c-.03028-.19285-.21-1.44617-.31592-2.15436-.07813-.71162-.22364-1.97155-.24317-2.16618q-.17577-1.75646-.312-3.51634-.30688-3.96012-.415-7.93346-.18456-7.06137.25293-14.12171c.166-2.64466.40235-5.28245.69336-7.91529.01758-.16178.03125-.29046.04395-.40642.01758-.11756.03808-.25112.0625-.41735.21533-1.455.44189-2.908.69531-4.35644a159.29393,159.29393,0,0,1,3.73926-16.31344q1.08765-3.80858,2.36279-7.55538.56836-1.66325,1.17285-3.31377c-.02637.071,1.18213-3.01574,1.19824-3.11945,4.79493-11.32906,12.6377-21.657,12.64844-34.52643.00733-8.71332-7.36426-17.06006-16-16.65739a16.62541,16.62541,0,0,0-16,16.65739c0,.19565-.001.36258-.00146.52461-.08057.27127-.14795.49761-.2085.69777-.93506,2.032-2.08057,3.96285-3.10693,5.94438-2.81055,5.42345-5.20948,11.02482-7.48828,16.70842a159.82519,159.82519,0,0,0-8.5293,28.62825,182.098,182.098,0,0,0-.96777,65.04019c6.20361,37.8128,26.66162,74.009,55.103,98.33928,2.46826,2.11134,4.99658,4.15171,7.54883,6.15089,1.05615.82714,2.12109,1.63909,3.19141,2.44577.50586.38139,2.60156,2.034,1.67968,1.23331,1.98243,1.69183,2.8169,2.5433,3.32618,3.78944.07373.2876.17431.67762.31054,1.21088.35889,1.40538.69971,2.80751.98291,4.23259,2.47559,12.45143,4.19629,25.11159,6.26416,37.64359,1.42139,8.61064,12.022,14.21487,19.68213,11.6342C513.49,733.31151,517.27178,725.05119,515.74883,715.82494Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M653.23,748.12012a25.37427,25.37427,0,0,1-25.43994-25.25V704.61865c-19.10889,2.13916-84.54321,1.64356-103.29981-1.37011v18.8916a25.4409,25.4409,0,0,1-50.88037,0v-37.0918a163.30721,163.30721,0,0,1-84-142.38818c0-49.83008,22.583-96.31446,61.95874-127.5337a164.07476,164.07476,0,0,1,89.32251-35.17333c4.427-.34717,8.88086-.52295,13.23877-.52295,15.3523,0,84.61866,2.37158,101.99781,7.37158L695.75562,347.48a4.16,4.16,0,0,1,4.55-.90674,4.10044,4.10044,0,0,1,2.58448,3.84668v6.252l7.71631-7.64209a4.1505,4.1505,0,0,1,4.54638-.90821,4.10219,4.10219,0,0,1,2.5874,3.84815v69.86474a162.75915,162.75915,0,0,1-39.07031,267.86621v33.16944A25.37427,25.37427,0,0,1,653.23,748.12012ZM629.79,702.35059v20.51953a23.44072,23.44072,0,0,0,46.87988,0V688.4375l.571-.271a162.45628,162.45628,0,0,0,66.53076-58.18164,160.32526,160.32526,0,0,0-27.69922-206.96l-.33228-.29834V351.96973a2.10314,2.10314,0,0,0-1.34179-1.99561,2.1492,2.1492,0,0,0-2.38135.47315l-11.127,11.0205V350.41992a2.10314,2.10314,0,0,0-1.3418-1.9956,2.15435,2.15435,0,0,0-2.38525.47656l-40.47632,40.16357-.5752-.18017c-16.156-5.05664-86.51269-7.4541-101.98169-7.4541-4.30591,0-8.70727.17382-13.08154.51709a162.07488,162.07488,0,0,0-88.23682,34.74609c-38.89453,30.83789-61.20166,76.751-61.20166,125.9668A161.31464,161.31464,0,0,0,475.094,683.585l.51587.28565v38.26953a23.441,23.441,0,0,0,46.88037,0V700.85449l1.18482.22266c16.85571,3.17139,87.83179,3.67139,104.98486,1.42139Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M517.66016,270.53,406.91992,303.81a13.8823,13.8823,0,0,0,7.99024,26.59l96.90966-29.13,2.03028-.61005a67.51666,67.51666,0,0,1,8.29-29.86,6.16489,6.16489,0,0,0-2-.52A6.367,6.367,0,0,0,517.66016,270.53Z'
+                transform='translate(-229.66231 -151.4124)'
+                className='fill-primary'
+            />
+            <path
+                d='M412.995,337.49524a7.01287,7.01287,0,0,1-6.63781-4.80127l-6.92481-20.896a6.99947,6.99947,0,0,1,4.44324-8.84668l90.13122-29.87012a6.99966,6.99966,0,0,1,8.84668,4.44189l6.92481,20.896a6.97565,6.97565,0,0,1-4.19312,8.75928q-.12487.04688-.25219.08838l-90.12915,29.86914A7.003,7.003,0,0,1,412.995,337.49524Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M581.79,233.42a70.03128,70.03128,0,0,0-69.97022,67.85c-.02.71-.02978,1.43-.02978,2.15a70,70,0,1,0,70-70Zm0,138a68.071,68.071,0,0,1-68-68q0-1.395.06006-2.76A67.998,67.998,0,1,1,581.79,371.42Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M640.79,303.42a59,59,0,0,1-118,0c0-1.87.08984-3.71.25976-5.53A59,59,0,0,1,640.79,303.42Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M599.35986,307.34a14.04753,14.04753,0,0,0-5.46972-4.33,53.13872,53.13872,0,0,0-8.22022-2.7V284.25a23.39587,23.39587,0,0,1,11.68994,5.33l3.26026-6.95a21.56068,21.56068,0,0,0-6.36035-3.93,31.25931,31.25931,0,0,0-8.06983-2.07v-8.14h-7.5498v8.22a20.67243,20.67243,0,0,0-8.25,2.77,16.3988,16.3988,0,0,0-5.59034,5.48,13.77113,13.77113,0,0,0-1.98974,7.29,12.46275,12.46275,0,0,0,2.29,7.8,14.51022,14.51022,0,0,0,5.6997,4.52,52.69206,52.69206,0,0,0,8.36036,2.77v15.39A24.98812,24.98812,0,0,1,565.25,317.18l-3.25977,6.96a24.58042,24.58042,0,0,0,7.43995,4.22,34.16142,34.16142,0,0,0,9.21,1.92v8.07h7.5498v-8.22a19.5385,19.5385,0,0,0,11.21-5.14,13.58959,13.58959,0,0,0,4.18018-10.1A12.0549,12.0549,0,0,0,599.35986,307.34Zm-20.1997-8.73a14.49857,14.49857,0,0,1-5.03028-2.7,5.32046,5.32046,0,0,1-1.77978-4.18,6.72557,6.72557,0,0,1,1.81982-4.81,9.20221,9.20221,0,0,1,4.99024-2.59Zm11.25,21.28a8.5575,8.5575,0,0,1-4.74024,2.47V309.12a13.23964,13.23964,0,0,1,4.74024,2.37,4.65775,4.65775,0,0,1,1.62988,3.7A6.8039,6.8039,0,0,1,590.41016,319.89Z'
+                transform='translate(-229.66231 -151.4124)'
+                className='fill-primary'
+            />
+            <path
+                d='M416.209,422.18421a7.02318,7.02318,0,0,1-6.775-5.21582l-3.77014-14.27a7.0039,7.0039,0,0,1,4.97509-8.55127l118.94068-31.42969a7.01537,7.01537,0,0,1,8.551,4.98535l3.19385,12.08984-.59839.04639a160.16254,160.16254,0,0,0-84.10754,31.8711l-.17346.084-38.46021,10.15967A6.96258,6.96258,0,0,1,416.209,422.18421Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M396.51011,492.6705a98.108,98.108,0,0,1-39.12217-14.79216A95.527,95.527,0,0,1,341.214,464.65c-4.82786-4.92434-9.4465-10.52122-11.83781-17.0792a25.299,25.299,0,0,1,.33124-18.7693,24.33352,24.33352,0,0,1,13.16876-13.12553c5.98961-2.38861,13.22619-2.64412,19.05541-.02821a17.08926,17.08926,0,0,1,6.96433,5.60584,11.43493,11.43493,0,0,1,1.8103,3.75816,7.67372,7.67372,0,0,1,.09036,3.75911,9.01634,9.01634,0,0,1-4.77388,5.723,24.06819,24.06819,0,0,1-8.1521,2.48626c-2.86013.39-6.96868.74485-9.268-1.551a4.66934,4.66934,0,0,1-1.09261-1.61079,3.1144,3.1144,0,0,1-.17238-1.45807c-.03354.21724.13872-.47236.12891-.44486-.03636.10194-.19384.3504.03-.03135a3.39683,3.39683,0,0,1,.23311-.363c-.076.09827-.24291.26227.05515-.01912a1.96186,1.96186,0,0,1,.91427-.56441c.13841-.04032.27883-.07242.41763-.11033.35277-.09637-.35161.00526.0093-.00631.13927-.00446.28021-.001.419-.00726.354-.01591-.34927-.09655-.01-.01126.13173.03312.26222.06547.392.10635.28208.08887-.19974-.20693-.00859.01446.224.25945-.23521-.25235.0194-.00119a4.6244,4.6244,0,0,1,.35767.39431,2.35924,2.35924,0,0,0,3.315,0,2.39935,2.39935,0,0,0,0-3.315c-2.3114-2.86535-6.87707-2.15735-9.21528.26436-2.982,3.08848-1.80111,7.98318,1.09954,10.678,3.2837,3.05076,7.9344,3.26284,12.15834,2.75114a29.817,29.817,0,0,0,10.26918-2.902,14.39829,14.39829,0,0,0,7.024-7.58817,13.17381,13.17381,0,0,0-1.05641-10.76889c-3.7293-7.01118-11.45152-10.60575-19.08437-11.22837a29.81221,29.81221,0,0,0-21.81879,7.16278,30.15422,30.15422,0,0,0-9.88822,20.075c-.66626,7.996,2.43608,15.76128,6.95922,22.21432a84.62746,84.62746,0,0,0,16.77944,17.3533,102.47323,102.47323,0,0,0,42.812,20.19071q2.79237.57685,5.61474.98758a2.42141,2.42141,0,0,0,2.88342-1.63715,2.36125,2.36125,0,0,0-1.63716-2.88343Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M619.8954,390.79425q-30.6631-.43278-61.133,1.55634c-6.9433.45338-12.95072,2.90742-12.95072,7.03206,0,3.44855,5.95964,7.48858,12.95072,7.032q30.45443-1.9887,61.133-1.55633C636.56931,405.09367,636.55948,391.02941,619.8954,390.79425Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#ccc'
+            />
+            <path
+                d='M589.68169,404.16175c-2.20923,0-54.10571-.08252-54.10571-7.251,0-7.168,51.89648-7.25049,54.10571-7.25049s54.10571.08252,54.10571,7.25049C643.7874,404.07923,591.89092,404.16175,589.68169,404.16175Zm-52.07642-7.251c1.42017,2.17334,20.49634,5.251,52.07642,5.251s50.656-3.07764,52.07641-5.251c-1.42016-2.17285-20.49633-5.25049-52.07641-5.25049S539.02544,394.73792,537.60527,396.91077Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M754.33555,592.61209c11.48831-.319,25.78612-.71776,36.801-8.76407a28.26188,28.26188,0,0,0,11.12052-21.10332c.22483-6.35459-2.071-11.90149-6.46488-15.61408-5.75306-4.86143-14.154-6.00169-23.20738-3.34108l9.38-68.548-6.88568-.94343L764.052,554.885l5.75009-2.63856c6.66637-3.05766,15.81731-4.61365,21.50378.19174a12.21541,12.21541,0,0,1,4.00576,10.06213,21.36089,21.36089,0,0,1-8.275,15.7346c-8.5719,6.26126-19.96858,7.06894-32.89451,7.43036Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#3f3d56'
+            />
+            <rect x='468.20272' y='329.7408' width='37.43223' height='6.95017' fill='#3f3d56' />
+            <path
+                d='M433.14751,454.15a7.01243,7.01243,0,0,1-6.904-5.91748L423.95867,433.651a7.00763,7.00763,0,0,1,5.83215-7.99951L551.32231,406.609a7.0074,7.0074,0,0,1,7.999,5.832l2.28467,14.582a7.00726,7.00726,0,0,1-5.83179,7.99952l-121.5315,19.042A7.09551,7.09551,0,0,1,433.14751,454.15Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <circle cx='153.18938' cy='80.85472' r='51' className='fill-primary' />
+            <path
+                d='M411.727,245.46306a12.09515,12.09515,0,0,0,4.36685-2.67614,8.13345,8.13345,0,0,0,2.25475-6.483,5.47156,5.47156,0,0,0-2.51209-4.16412c-1.8457-1.13505-4.28459-1.09217-6.74692.0493l-.2842-19.90787-1.99969.02846.334,23.40381,1.52231-.99844c1.76529-1.15588,4.30177-1.99407,6.12663-.87157a3.51373,3.51373,0,0,1,1.57291,2.68953,6.14619,6.14619,0,0,1-1.67626,4.83366c-2.16794,2.15111-5.37684,2.87189-9.03924,3.53112l.35442,1.96814A32.96256,32.96256,0,0,0,411.727,245.46306Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <rect
+                x='418.83797'
+                y='211.67905'
+                width='10.77167'
+                height='1.99975'
+                transform='translate(-256.7253 -85.49311) rotate(-8.61162)'
+                fill='#2f2e41'
+            />
+            <rect
+                x='385.22114'
+                y='216.76898'
+                width='10.77167'
+                height='1.99975'
+                transform='translate(-257.86644 -90.46937) rotate(-8.61162)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M416.55474,453.25843a9.03041,9.03041,0,0,1-8.43445-5.855L372.9572,353.2345a46.53156,46.53156,0,0,1,27.28418-59.80225L428.83306,282.756a8.95769,8.95769,0,0,1,6.87805.23877,8.83186,8.83186,0,0,1,4.67541,4.98584L505.523,410.96448a9.0015,9.0015,0,0,1-5.25659,11.63818L419.7,452.68665A8.97612,8.97612,0,0,1,416.55474,453.25843Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M420.15839,285.99487l64.629,86.42566L441.05579,286.902a8.079,8.079,0,0,0-10.349-4.846Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M522.49036,351.63l-115.39525-7.43312a13.87944,13.87944,0,1,1,1.78425-27.70148l115.3953,7.43258a6.49946,6.49946,0,0,1,6.06882,6.90407l-.9485,14.72919a6.50056,6.50056,0,0,1-6.90462,6.06876Z'
+                transform='translate(-229.66231 -151.4124)'
+                className='fill-primary'
+            />
+            <path
+                d='M491.65508,356.39514q-.22852,0-.45825-.01464h-.00025l-94.7561-6.10352a6.99919,6.99919,0,0,1-6.53565-7.43506l1.62036-25.1582a6.99967,6.99967,0,0,1,7.43592-6.53516l94.756,6.103a7.00755,7.00755,0,0,1,6.5354,7.43506l-1.62012,25.15869a7.01886,7.01886,0,0,1-4.54614,6.11084A6.926,6.926,0,0,1,491.65508,356.39514Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M382.26706,198.35294c6.70923-5.77031,15.92258-.60839,23.61673-1.512,7.36114-.86455,13.17369-7.46753,14.65857-14.48237,1.73235-8.184-2.51553-16.38319-8.77663-21.48577-6.85745-5.58858-15.95214-7.19293-24.57042-6.159-9.87767,1.185-18.91123,5.78935-27.017,11.35854a121.85037,121.85037,0,0,0-21.44164,18.16917c-5.75111,6.35166-10.62377,13.9227-12.17974,22.46976-1.414,7.76751-.33431,16.375,4.2575,22.95514a24.20528,24.20528,0,0,0,9.47417,7.80529c3.93743,1.93344,8.13754,3.31565,11.95217,5.50169,5.76775,3.30532,11.37351,10.15883,9.56677,17.29094a9.79281,9.79281,0,0,1-2.226,4.19c-1.29065,1.43275-3.6187-.46055-2.32465-1.89708,2.27282-2.52305,2.18331-5.88792.97206-8.90023a16.27233,16.27233,0,0,0-7.23346-7.95427c-3.993-2.28294-8.40146-3.6921-12.494-5.77156a27.02254,27.02254,0,0,1-9.91143-8.21385c-4.89822-6.807-6.36338-15.66585-5.2525-23.861,1.20214-8.86841,5.75869-16.9272,11.47065-23.68045,6.21621-7.34944,13.8519-13.57054,21.65246-19.15936,8.37006-5.99685,17.61734-11.06678,27.85008-12.90049,8.87016-1.58954,18.48474-.61091,26.24132,4.22294,7.24026,4.51209,12.79232,12.27377,13.34449,20.9467a22.23537,22.23537,0,0,1-10.37627,19.94206,19.101,19.101,0,0,1-11.41,2.64333c-4.2616-.27279-8.52728-1.57778-12.81609-1.21548a8.94077,8.94077,0,0,0-5.26807,2.1285c-1.46588,1.26074-3.21292-1.1807-1.75908-2.43109Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M969.147,748.5876H230.853a1.19069,1.19069,0,0,1,0-2.38137H969.147a1.19068,1.19068,0,0,1,0,2.38137Z'
+                transform='translate(-229.66231 -151.4124)'
+                fill='#3f3d56'
+            />
+        </svg>
+    )
+}
+
+export const NoDataIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 647.63626 632.17383'
+        >
+            <path
+                d='M687.3279,276.08691H512.81813a15.01828,15.01828,0,0,0-15,15v387.85l-2,.61005-42.81006,13.11a8.00676,8.00676,0,0,1-9.98974-5.31L315.678,271.39691a8.00313,8.00313,0,0,1,5.31006-9.99l65.97022-20.2,191.25-58.54,65.96972-20.2a7.98927,7.98927,0,0,1,9.99024,5.3l32.5498,106.32Z'
+                transform='translate(-276.18187 -133.91309)'
+                fill='#f2f2f2'
+            />
+            <path
+                d='M725.408,274.08691l-39.23-128.14a16.99368,16.99368,0,0,0-21.23-11.28l-92.75,28.39L380.95827,221.60693l-92.75,28.4a17.0152,17.0152,0,0,0-11.28028,21.23l134.08008,437.93a17.02661,17.02661,0,0,0,16.26026,12.03,16.78926,16.78926,0,0,0,4.96972-.75l63.58008-19.46,2-.62v-2.09l-2,.61-64.16992,19.65a15.01489,15.01489,0,0,1-18.73-9.95l-134.06983-437.94a14.97935,14.97935,0,0,1,9.94971-18.73l92.75-28.4,191.24024-58.54,92.75-28.4a15.15551,15.15551,0,0,1,4.40966-.66,15.01461,15.01461,0,0,1,14.32032,10.61l39.0498,127.56.62012,2h2.08008Z'
+                transform='translate(-276.18187 -133.91309)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M398.86279,261.73389a9.0157,9.0157,0,0,1-8.61133-6.3667l-12.88037-42.07178a8.99884,8.99884,0,0,1,5.9712-11.24023l175.939-53.86377a9.00867,9.00867,0,0,1,11.24072,5.9707l12.88037,42.07227a9.01029,9.01029,0,0,1-5.9707,11.24072L401.49219,261.33887A8.976,8.976,0,0,1,398.86279,261.73389Z'
+                transform='translate(-276.18187 -133.91309)'
+                className='fill-primary'
+            />
+            <circle cx='190.15351' cy='24.95465' r='20' className='fill-primary' />
+            <circle cx='190.15351' cy='24.95465' r='12.66462' fill='#fff' />
+            <path
+                d='M878.81836,716.08691h-338a8.50981,8.50981,0,0,1-8.5-8.5v-405a8.50951,8.50951,0,0,1,8.5-8.5h338a8.50982,8.50982,0,0,1,8.5,8.5v405A8.51013,8.51013,0,0,1,878.81836,716.08691Z'
+                transform='translate(-276.18187 -133.91309)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M723.31813,274.08691h-210.5a17.02411,17.02411,0,0,0-17,17v407.8l2-.61v-407.19a15.01828,15.01828,0,0,1,15-15H723.93825Zm183.5,0h-394a17.02411,17.02411,0,0,0-17,17v458a17.0241,17.0241,0,0,0,17,17h394a17.0241,17.0241,0,0,0,17-17v-458A17.02411,17.02411,0,0,0,906.81813,274.08691Zm15,475a15.01828,15.01828,0,0,1-15,15h-394a15.01828,15.01828,0,0,1-15-15v-458a15.01828,15.01828,0,0,1,15-15h394a15.01828,15.01828,0,0,1,15,15Z'
+                transform='translate(-276.18187 -133.91309)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M801.81836,318.08691h-184a9.01015,9.01015,0,0,1-9-9v-44a9.01016,9.01016,0,0,1,9-9h184a9.01016,9.01016,0,0,1,9,9v44A9.01015,9.01015,0,0,1,801.81836,318.08691Z'
+                transform='translate(-276.18187 -133.91309)'
+                className='fill-primary'
+            />
+            <circle cx='433.63626' cy='105.17383' r='20' className='fill-primary' />
+            <circle cx='433.63626' cy='105.17383' r='12.18187' fill='#fff' />
+        </svg>
+    )
+}
+
+export const PeopleDivisionIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 949 568.6528'
+        >
+            <path
+                d='M493.95429,675.17362a31.51716,31.51716,0,0,1-30.43451-5.36117c-10.65985-8.94708-14.002-23.68733-16.72141-37.336l-8.04348-40.37019,16.83986,11.59529c12.11023,8.33865,24.49328,16.94444,32.87836,29.02256s12.04517,28.56585,5.308,41.63494'
+                transform='translate(-125.5 -165.6736)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M492.51482,724.05014c1.70442-12.415,3.457-24.98956,2.26145-37.53768-1.06178-11.14409-4.46167-22.02873-11.38341-30.95339a51.49146,51.49146,0,0,0-13.2109-11.97129c-1.32109-.83387-2.537,1.25978-1.22166,2.09a48.95178,48.95178,0,0,1,19.363,23.36288c4.2156,10.72234,4.89256,22.41059,4.16611,33.8008-.43931,6.88809-1.37037,13.731-2.30861,20.5652a1.25406,1.25406,0,0,0,.84528,1.48875,1.21763,1.21763,0,0,0,1.48874-.84529Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#f2f2f2'
+            />
+            <path
+                d='M476.09817,702.08138a23.20063,23.20063,0,0,1-20.2149,10.43338c-10.23314-.48578-18.7643-7.62786-26.444-14.40842l-22.7149-20.05553,15.0335-.71945c10.8112-.51738,21.90132-1,32.19861,2.334s19.794,11.35766,21.67693,22.01619'
+                transform='translate(-125.5 -165.6736)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M497.33163,731.187c-8.20362-14.51579-17.71911-30.64843-34.72169-35.80468a38.74557,38.74557,0,0,0-14.60449-1.50794c-1.55056.1337-1.16337,2.52385.38438,2.39039a35.99579,35.99579,0,0,1,23.30647,6.166c6.57154,4.47307,11.68821,10.69183,16.0189,17.28653,2.65269,4.03945,5.02879,8.25028,7.40535,12.45545.7595,1.34389,2.97945.3738,2.21108-.98579Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#f2f2f2'
+            />
+            <path
+                d='M717.21327,440.39728,538.36,588.63164a11.02654,11.02654,0,0,1-15.50846-1.45176L447.4565,496.21142a11.02655,11.02655,0,0,1,1.45176-15.50847L627.7615,332.46859A11.02656,11.02656,0,0,1,643.27,333.92035L718.665,424.88881A11.02655,11.02655,0,0,1,717.21327,440.39728Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M710.046,435.93377,535.30428,580.76044a5.68027,5.68027,0,0,1-7.98921-.74787l-71.98738-86.8569a5.68027,5.68027,0,0,1,.74788-7.98921L630.81725,340.33979a5.68026,5.68026,0,0,1,7.98921.74788l71.98737,86.85689A5.68026,5.68026,0,0,1,710.046,435.93377Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#fff'
+            />
+            <path
+                d='M718.10135,466.25042l-35.94022-3.36416,25.72328-20.24515,13.857,16.71931a4.152,4.152,0,0,1,.48183,4.66117,4.27907,4.27907,0,0,1-1.05143,1.30238A4.15966,4.15966,0,0,1,718.10135,466.25042Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M645.92753,401.87527l-88.39874,73.26526a5.34016,5.34016,0,1,1-6.81537-8.22314l88.39873-73.26525a5.34016,5.34016,0,1,1,6.81538,8.22313Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M602.66388,408.25491l-59.61775,49.41145a5.34016,5.34016,0,1,1-6.81538-8.22313l59.61775-49.41146a5.34016,5.34016,0,0,1,6.81538,8.22314Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <circle cx='384.78656' cy='326.94689' r='16.688' fill='#6c63ff' />
+            <circle cx='517.5324' cy='288.15569' r='5.34016' fill='#e6e6e6' />
+            <circle cx='531.92289' cy='276.22879' r='5.34016' fill='#e6e6e6' />
+            <circle cx='546.31338' cy='264.30189' r='5.34016' fill='#e6e6e6' />
+            <polygon
+                points='461.198 556.214 473.458 556.213 479.29 508.925 461.196 508.926 461.198 556.214'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M583.57113,717.885l24.14379-.001h.001a15.38605,15.38605,0,0,1,15.38648,15.38623v.5l-39.53052.00147Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='417.198 556.214 429.458 556.213 435.29 508.925 417.196 508.926 417.198 556.214'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M539.57113,717.885l24.14379-.001h.001a15.38605,15.38605,0,0,1,15.38648,15.38623v.5l-39.53052.00147Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M555.001,707.41872h-.00219l-10.42554-.00293a4.48165,4.48165,0,0,1-4.49487-4.33789l-.32837-9.1211c-2.6355-72.94043-5.36084-148.36425,11.51733-182.26465l.13233-.2666,55.88012-2.02246,11.21851,102.91114a44.62081,44.62081,0,0,1-.52,13.16406L603.148,703.62673a4.49876,4.49876,0,0,1-4.59717,3.6582l-11.71045-.37988a4.50188,4.50188,0,0,1-4.30176-4.94336l9.199-72.2793a24.70205,24.70205,0,0,0-.83472-9.28418l-11.44043-39.99414a1.50016,1.50016,0,0,0-2.92822.20605L559.45755,703.53786A4.51837,4.51837,0,0,1,555.001,707.41872Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <circle cx='436.18347' cy='216.28068' r='24.56103' fill='#ffb8b8' />
+            <path
+                d='M604.66043,516.99977h-50.823a4.50035,4.50035,0,0,1-4.45313-5.15039l2.51075-17.19629-14.7439-40.55371a29.8437,29.8437,0,0,1,24.00879-39.7666h0a29.73019,29.73019,0,0,1,31.02075,16.81543l3.23853,6.89062L609.085,511.67946a4.49963,4.49963,0,0,1-4.42456,5.32031Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#6c63ff'
+            />
+            <path
+                d='M497.37388,560.13581a10.74268,10.74268,0,0,0,7.44457-14.69442l39.60234-89.47207-22.65457-5.706L489.962,540.00569a10.80091,10.80091,0,0,0,7.41184,20.13012Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M539.05977,471.97084l-22.03344-10.31357a4.81689,4.81689,0,0,1-1.97427-7.02164l13.39047-20.22269a13.37737,13.37737,0,0,1,24.198,11.41337l-6.92415,23.16293a4.81687,4.81687,0,0,1-6.65666,2.9816Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#6c63ff'
+            />
+            <path
+                d='M662.5751,329.89742a10.7427,10.7427,0,0,1-8.80893,13.91942L597.1973,423.65142,580.69476,407.115l60.79449-73.275a10.80091,10.80091,0,0,1,21.08585-3.94258Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M610.27876,414.28061a4.81164,4.81164,0,0,1-1.23486,3.81607l-16.34924,17.8092a13.37737,13.37737,0,0,1-22.19518-14.93937L581.01184,399.109a4.81688,4.81688,0,0,1,7.04207-1.90008l20.142,13.64334A4.81143,4.81143,0,0,1,610.27876,414.28061Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#6c63ff'
+            />
+            <path
+                d='M593.95486,371.6736a25.25979,25.25979,0,0,0-.91846-6.57593c-3.957-14.62243-18.56543-23.39868-33.22509-21.19629a31.69966,31.69966,0,0,0-8.76856-1.22778h0a31.621,31.621,0,0,0-31.24609,26.767l-11.8418,76.233h64l10.334-66.52551a31.77725,31.77725,0,0,0,.269-7.47449Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M673.5,734.3264h-211a1,1,0,1,1,0-2h211a1,1,0,0,1,0,2Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M886.07028,243.34113a10.05579,10.05579,0,0,1,10.98633,10.81926l32.70724,14.395-15.60167,10.07083-28.66745-15.21368a10.11027,10.11027,0,0,1,.57555-20.07138Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#a0616a'
+            />
+            <circle cx='892.50184' cy='35.14965' r='24.56103' fill='#a0616a' />
+            <polygon
+                points='848.05 385.364 835.79 385.363 829.958 338.075 848.052 338.076 848.05 385.364'
+                fill='#a0616a'
+            />
+            <path
+                d='M976.67645,562.92134l-39.53052-.00147v-.5a15.38605,15.38605,0,0,1,15.38647-15.38623h.001l24.1438.001Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='899.05 385.364 886.79 385.363 880.958 338.075 899.052 338.076 899.05 385.364'
+                fill='#a0616a'
+            />
+            <path
+                d='M1027.67645,562.92134l-39.53052-.00147v-.5a15.38605,15.38605,0,0,1,15.38647-15.38623h.001l24.1438.001Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M1023.96878,540.96559a4.50658,4.50658,0,0,1-.54883-.03418l-17.9458-2.209a4.49289,4.49289,0,0,1-3.94434-4.23535L996.30032,416.7234a1.49993,1.49993,0,0,0-2.85742-.55371l-13.43433,44.78418L977.242,532.8904a4.498,4.498,0,0,1-5.0227,4.29492l-15.54053-1.8291a4.48512,4.48512,0,0,1-3.97315-4.55664l1.31934-67.71387,10.52881-106.292.47046.01758,64.02221,2.46777-.5769,177.20117a4.50146,4.50146,0,0,1-4.50073,4.48535Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M1026.62014,368.992c-.03906,0-.07812,0-.11743-.001l-64.479-1.68945a4.49979,4.49979,0,0,1-4.34107-5.10449l12.39014-91.19141a44.13237,44.13237,0,0,1,18.68018-30.792,38.19591,38.19591,0,0,1,31.10815-5.54688c22.5061,5.70313,36.00757,29.87012,30.09717,53.87207l-18.968,77.0293A4.52648,4.52648,0,0,1,1026.62014,368.992Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M942.02981,296.5779a21.82211,21.82211,0,0,1-11.96411-3.55469l-37.60229-24.66992a4.4992,4.4992,0,0,1-.79883-6.85645l5.66455-5.98047a4.48379,4.48379,0,0,1,4.93164-1.08691l42.00171,16.72559,38.90869-27.78614a16.15455,16.15455,0,1,1,16.44629,27.67774l-48.08814,23.35742A21.74021,21.74021,0,0,1,942.02981,296.5779Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M1026.51241,224.31132c4.19022.9894,9.24531,1.74863,12.23986-1.34484a12.5105,12.5105,0,0,0,1.41592-1.89006l3.9152-5.97741c2.9992-4.57894,6.05107-9.27733,7.3205-14.60185a23.74125,23.74125,0,0,0-17.90687-28.67359c-5.84568-1.30882-11.56585-3.56078-17.54216-3.15s-12.98468.27592-18-3c.31289,1.46046,2.00253,7.9134,2.00253,7.9134s-7.5414,6.91867-9.054,8.64384a14.1663,14.1663,0,0,1,19.723,10.67121c.4281,2.62585.17666,5.538,1.73668,7.69318,1.48376,2.0498,4.14514,2.768,6.38173,3.95152a15.56319,15.56319,0,0,1,6.84685,19.04477Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M997.02308,377.78843H793.43361a9.66388,9.66388,0,0,1-9.653-9.653V264.58566a9.66388,9.66388,0,0,1,9.653-9.653H997.02308a9.66387,9.66387,0,0,1,9.65295,9.653V368.13548A9.66387,9.66387,0,0,1,997.02308,377.78843Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M994.683,370.7681H795.77372a4.97829,4.97829,0,0,1-4.97273-4.97273v-98.8696a4.97829,4.97829,0,0,1,4.97273-4.97273H994.683a4.9783,4.9783,0,0,1,4.97274,4.97273v98.8696A4.9783,4.9783,0,0,1,994.683,370.7681Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#fff'
+            />
+            <path
+                d='M983.16361,395.73038l-22.37044-22.37015,28.68.725V393.1169a3.63886,3.63886,0,0,1-2.28169,3.41475,3.75017,3.75017,0,0,1-1.43786.2908A3.64556,3.64556,0,0,1,983.16361,395.73038Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M970.46457,311.92684H869.83988a4.68022,4.68022,0,1,1,0-9.36044H970.46457a4.68022,4.68022,0,0,1,0,9.36044Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M937.703,292.03591H869.83988a4.68022,4.68022,0,1,1,0-9.36043H937.703a4.68021,4.68021,0,1,1,0,9.36043Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <circle cx='702.68566' cy='131.62756' r='14.62568' fill='#6c63ff' />
+            <circle cx='813.95497' cy='179.69184' r='4.68022' fill='#e6e6e6' />
+            <circle cx='830.33573' cy='179.69184' r='4.68022' fill='#e6e6e6' />
+            <circle cx='846.71649' cy='179.69184' r='4.68022' fill='#e6e6e6' />
+            <path
+                d='M993.8706,365.474a10.05578,10.05578,0,0,1,12.8385-8.53983l20.78344-29.06936,6.665,17.33237-20.75813,24.94752a10.11027,10.11027,0,0,1-19.52885-4.6707Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#a0616a'
+            />
+            <path
+                d='M1017.67789,363.9055a4.53555,4.53555,0,0,1-.98107-.02466,4.48884,4.48884,0,0,1-3.20506-2.06l-4.39223-6.96833a4.47981,4.47981,0,0,1,.16818-5.04824l26.602-36.55347-17.29062-44.57587a16.06991,16.06991,0,0,1,8.81348-20.73946h0a16.154,16.154,0,0,1,22.07333,11.65569l10.72775,52.37339a21.8839,21.8839,0,0,1-6.65051,20.45445l-33.21,30.32536A4.4868,4.4868,0,0,1,1017.67789,363.9055Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M1073.5,565.3264h-211a1,1,0,0,1,0-2h211a1,1,0,0,1,0,2Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M382.11757,316.44985H178.5281a9.66388,9.66388,0,0,1-9.653-9.653V203.24708a9.66387,9.66387,0,0,1,9.653-9.653H382.11757a9.66387,9.66387,0,0,1,9.65295,9.653V306.7969A9.66388,9.66388,0,0,1,382.11757,316.44985Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M379.77746,309.42952H180.86821a4.97829,4.97829,0,0,1-4.97273-4.97273v-98.8696a4.9783,4.9783,0,0,1,4.97273-4.97274H379.77746a4.9783,4.9783,0,0,1,4.97273,4.97274v98.8696A4.97829,4.97829,0,0,1,379.77746,309.42952Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#fff'
+            />
+            <path
+                d='M368.2581,334.3918l-22.37044-22.37015,28.68.725v19.03167a3.63886,3.63886,0,0,1-2.28169,3.41475,3.7502,3.7502,0,0,1-1.43786.2908A3.64558,3.64558,0,0,1,368.2581,334.3918Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M355.55905,250.58826H254.93437a4.68022,4.68022,0,1,1,0-9.36044H355.55905a4.68022,4.68022,0,0,1,0,9.36044Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <path
+                d='M322.79753,230.69733H254.93437a4.68022,4.68022,0,1,1,0-9.36044h67.86316a4.68022,4.68022,0,1,1,0,9.36044Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+            <circle cx='87.78015' cy='70.28897' r='14.62568' fill='#6c63ff' />
+            <circle cx='199.04946' cy='118.35326' r='4.68022' fill='#e6e6e6' />
+            <circle cx='215.43022' cy='118.35326' r='4.68022' fill='#e6e6e6' />
+            <circle cx='231.81098' cy='118.35326' r='4.68022' fill='#e6e6e6' />
+            <path
+                d='M373.6744,183.645a10.05575,10.05575,0,0,0-5.31,14.47619l-23.4203,26.99016,18.40886,2.43876,19.418-26.0042A10.11027,10.11027,0,0,0,373.6744,183.645Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M217.29809,194.45055a10.0558,10.0558,0,0,1-7.75445,13.32757l-5.82806,35.25639L190.08807,230.42l7.62367-31.54613a10.11027,10.11027,0,0,1,19.58635-4.42336Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ffb8b8'
+            />
+            <polygon
+                points='128.399 385.364 140.659 385.363 146.491 338.075 128.397 338.076 128.399 385.364'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M250.77164,547.03462l24.1438-.001h.001a15.38605,15.38605,0,0,1,15.38647,15.38623v.5l-39.53052.00147Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='65.399 385.364 77.659 385.363 83.491 338.075 65.397 338.076 65.399 385.364'
+                fill='#ffb8b8'
+            />
+            <path
+                d='M187.77164,547.03462l24.1438-.001h.001a15.38605,15.38605,0,0,1,15.38647,15.38623v.5l-39.53052.00147Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <circle cx='125.83969' cy='37.94709' r='24.56103' fill='#ffb8b8' />
+            <path
+                d='M248.39557,535.32816l-2.03076-129.40625a1.49985,1.49985,0,0,0-2.95923-.32422l-33.291,126.57617a4.55506,4.55506,0,0,1-4.571,3.459l-18.29-.78223a4.50093,4.50093,0,0,1-4.272-5.06055l28.24878-162.6582.05786-.459,78.68213,3.72265.51318.02344L273.61115,535.59867a4.51892,4.51892,0,0,1-4.48731,4.166H252.89508A4.51825,4.51825,0,0,1,248.39557,535.32816Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M310.54034,381.35746l-114.17773-9.37012a4.49926,4.49926,0,0,1-3.86353-6.01465l17.98072-49.71191,1.03467-59.00293a16.18147,16.18147,0,0,1,16.667-15.89063l7.593-6.88281h27.81958l2.42236,3.17578,19.35718,2.97461a16.27225,16.27225,0,0,1,13.58007,13.833L315.368,376.27249a4.5006,4.5006,0,0,1-4.45386,5.10059C310.7901,381.37308,310.66534,381.3682,310.54034,381.35746Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M288.29425,246.35453l48.68359-14.26368L359.01056,203.083a4.49872,4.49872,0,0,1,5.35229-1.416l13.75635,5.87988a4.49956,4.49956,0,0,1,2.11914,6.40332l-16.46557,28.25586A25.19721,25.19721,0,0,1,352.119,252.62894l-57.26123,25.3125Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M194.89288,267.69144a20.86486,20.86486,0,0,1-8.60376-18.5752l2.13648-28.30761a4.51169,4.51169,0,0,1,4.48755-4.16114h.03979l16.13184.14356a4.5009,4.5009,0,0,1,4.376,5.36523l-2.99316,15.27149,21.36011,9.02539-17.98706,34.99316Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M265.62782,174.1697s-25.685-13.83022-39.685,2.49239,5.3339,41.8476,5.3339,41.8476l27.548,3.4744s1-18.72648,7.559-18.36293S293.31286,180.35469,265.62782,174.1697Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M337.5,564.3264h-211a1,1,0,0,1,0-2h211a1,1,0,0,1,0,2Z'
+                transform='translate(-125.5 -165.6736)'
+                fill='#ccc'
+            />
+        </svg>
+    )
+}
+
+export const PeopleFunIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 732.64277 638.92105'
+            // xmlns:xlink='http://www.w3.org/1999/xlink'
+        >
+            <path
+                d='M355.19705,352.95965a33.50275,33.50275,0,0,1-5.15926.12031,1.12527,1.12527,0,0,0-1.23851.97666l-3.29091,24.68526a10.02142,10.02142,0,0,1-2.12316,4.89034,10.65824,10.65824,0,0,0-2.39918,8.14585v41.93238s-.21231,5.591,4.08355,5.591a62.10276,62.10276,0,0,0,15.48491,0s3.43951-.6511,3.43951-5.81038V392.2027a10.09919,10.09919,0,0,0-2.12316-8.20955,10.276,10.276,0,0,1-2.05945-4.862l-3.4183-25.20189A1.12526,1.12526,0,0,0,355.19705,352.95965Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#fff'
+            />
+            <path
+                d='M352.797,440.49242a62.674,62.674,0,0,1-7.773-.48345,4.17988,4.17988,0,0,1-3.09921-1.26857c-1.74822-1.81561-1.651-4.92536-1.64594-5.057l-.00051-41.86223a11.40411,11.40411,0,0,1,2.56375-8.64261,9.30817,9.30817,0,0,0,1.96541-4.53555l3.29031-24.68a1.84294,1.84294,0,0,1,2.01689-1.5865,32.6571,32.6571,0,0,0,5.01158-.12095l.01969-.00172a1.8326,1.8326,0,0,1,1.94831,1.57923l3.41851,25.20318a9.5778,9.5778,0,0,0,1.91686,4.52242,10.77949,10.77949,0,0,1,2.27209,8.70689V433.4909c0,5.67973-3.97539,6.49837-4.01565,6.506A63.14407,63.14407,0,0,1,352.797,440.49242Zm-2.85541-86.70877a.42442.42442,0,0,0-.44026.36215l-3.29151,24.68967a10.73614,10.73614,0,0,1-2.27383,5.2412,9.99118,9.99118,0,0,0-2.24756,7.6142l.00535.08709v41.93238c-.00138.05218-.07654,2.67433,1.25389,4.05176a2.79106,2.79106,0,0,0,2.122.83178l.08829.00553a61.46578,61.46578,0,0,0,15.30824,0c.06721-.01659,2.82016-.64448,2.82016-5.10851l.01141-41.41507a9.37629,9.37629,0,0,0-1.97439-7.63424,11.01282,11.01282,0,0,1-2.21249-5.21114l-3.419-25.20629a.41994.41994,0,0,0-.43628-.359,34.36271,34.36271,0,0,1-5.25572.12164A.55569.55569,0,0,0,349.94156,353.78365Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#3f3d56'
+            />
+            <path
+                className='fill-primary'
+                d='M116.32377,261.98877h5.21588a7.82029,7.82029,0,0,1,7.82029,7.82029v19.62507a7.8203,7.8203,0,0,1-7.8203,7.8203h-5.21587a7.8203,7.8203,0,0,1-7.8203-7.8203V269.80907A7.8203,7.8203,0,0,1,116.32377,261.98877Z'
+            />
+            <polygon
+                points='20.432 622.437 37.437 622.437 45.527 556.846 20.429 556.847 20.432 622.437'
+                fill='#9f616a'
+            />
+            <path
+                d='M249.77314,747.42485l33.489-.00136h.00136A21.34295,21.34295,0,0,1,304.6053,768.765v.69352l-54.83115.002Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M243.42009,518.702s-7.36985,8.33074,4.83348,40.805l-2.14951,178.06312,39.54717-.77152,8.69291-86.85779L310.1713,523.19059Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M323.73484,543.3204l-.433-.06611c-50.59181-7.71594-88.76927-15.11294-89.149-15.18678l-.47425-.092,13.26225-32.257c-.10477-3.37068-2.97558-98.02325,2.52008-125.2179,5.32015-26.326,30.39753-30.64031,33.2552-31.04623l2.86946-9.50739,36.25147,18.86363,6.32248,21.76,1.28425,59.63878Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M345.56864,420.57049a11.83043,11.83043,0,0,0-8.69092,3.8009H275.945l-22.31936,4.7509,3.66448,28.26389,79.39621-17.07512a11.87089,11.87089,0,1,0,8.88228-19.74057Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#9f616a'
+            />
+            <circle
+                cx='320.3296'
+                cy='305.17684'
+                r='29.36627'
+                transform='translate(-334.77698 309.33118) rotate(-61.33684)'
+                fill='#9f616a'
+            />
+            <path
+                d='M292.71018,281.7l-1.42366-2.8654,7.16351-3.55915s7.9025-12.85821,22.20235-9.24477,20.73358,5.776,20.73358,5.776l7.14539,3.59532-3.58624,3.56819,6.43357,2.16264-4.29807,2.13548,5.00092,2.87444-2.6921,14.73534s-4.47141-11.17619-13.06756-6.90524-24.31981-2.20782-24.31981-2.20782l-13.6604,26.43719s-2.81814-9.26755-7.84272-6.79639C290.49886,311.40586,277.66236,290.96281,292.71018,281.7Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M257.68022,462.68439,250.712,441.957l-1.68553-69.0936,2.969-9.80314a27.38884,27.38884,0,1,1,51.7277,17.89639l-15.91357,40.779.80894,1.56994,34.47683-.27912L335.342,445.0706Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M602.627,454.16756a10.61126,10.61126,0,0,0,1.6953-16.18255l16.99418-33.6624-19.465,2.2578-13.15376,31.62024A10.66877,10.66877,0,0,0,602.627,454.16756Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#ffb7b7'
+            />
+            <polygon
+                points='505.24 562.075 489.924 562.074 482.638 502.999 505.242 503.001 505.24 562.075'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M742.82423,707.46023l-49.38381-.00183v-.62463A19.22259,19.22259,0,0,1,712.662,687.61252h.00122l30.162.00122Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='421.54 562.075 406.224 562.074 398.938 502.999 421.543 503.001 421.54 562.075'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M655.37672,707.46023l-49.38381-.00183v-.62463a19.22259,19.22259,0,0,1,19.22155-19.22125h.00122l30.16195.00122Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='420.285 312.572 405.294 347.108 392.928 545.614 424.159 544.365 432.689 449.736 459.002 365.454 474.129 459.415 480.376 541.866 506.218 542.985 516.516 312.572 420.285 312.572'
+                fill='#2f2e41'
+            />
+            <path
+                className='fill-primary'
+                d='M728.899,272.694l-43.44888,1.08685-25.50892,13.09525-1.3999,96.6139-15.90944,66.70389s99.40388,6.84133,113.14563-3.15267l-16.24025-76.20426,21.78179-82.355Z'
+                transform='translate(-233.67861 -130.53948)'
+            />
+            <path
+                className='fill-primary'
+                d='M666.65057,288.47751l-4.04459-2.73159s-16.757,5.02234-21.7098,27.36637c0,0-3.79,11.21018-5.453,18.91745-1.90739,8.8402-44.02617,90.04135-44.02617,90.04135l25.21707,7.23431,46.232-74.76592Z'
+                transform='translate(-233.67861 -130.53948)'
+            />
+            <path
+                d='M760.63836,466.55646a10.61127,10.61127,0,0,1,2.83289-16.0226l-7.05258-37.04349,18.08768,7.53789,3.92422,34.02148a10.66878,10.66878,0,0,1-17.79221,11.50672Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#ffb7b7'
+            />
+            <path
+                className='fill-primary'
+                d='M754.78155,289.63567l4.641-1.51034s14.72235,9.44849,13.32155,32.29195c0,0,.55178,11.82066.0249,19.68766-.60433,9.02342,7.4955,98.69077,7.4955,98.69077H754.03028L742.787,357.59445Z'
+                transform='translate(-233.67861 -130.53948)'
+            />
+            <path
+                d='M739.56476,232.16033a30.11942,30.11942,0,1,1-30.11944-30.11941h0a30.04447,30.04447,0,0,1,30.1194,29.96935Q739.56493,232.0853,739.56476,232.16033Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M679.426,191.7045c3.52832-7.5871,10.35635-6.08706,16.32792-3.51525,7.56208-1.67738,14.75411-6.69923,22.90181-3.72933,8.02585,11.67312,34.97217,8.23871,29.23939,27.24263-.00749,4.55375,8.56874,1.90317,7.0715,9.36116,4.54315,14.35282-16.39873,41.53492-28.421,35.81082,2.97327-5.44922,9.76538-17.82623-.53635-19.05385-22.16059,20.61963-2.28641-39.26444-30.13067-21.84919C686.65921,224.40872,674.03758,200.14386,679.426,191.7045Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M765.91087,469.55994a10.00931,10.00931,0,0,0,11.97985-9.59409L853.518,388.78344l-17.0164-13.574-69.11686,74.41809a10.06355,10.06355,0,0,0-1.47386,19.93241Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#9f616a'
+            />
+            <path
+                d='M842.80154,400.28953,828.08838,383.0467a4.488,4.488,0,0,1,1.03243-6.71709l19.15464-11.991a12.46411,12.46411,0,0,1,16.12523,19.01023l-14.80328,16.978a4.488,4.488,0,0,1-6.79586-.03728Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#cbcbcb'
+            />
+            <polygon
+                points='569.558 620.997 551.067 620.996 542.27 549.675 569.561 549.676 569.558 620.997'
+                fill='#9f616a'
+            />
+            <path
+                d='M807.95189,769.46052l-59.622-.00221v-.75412A23.20776,23.20776,0,0,1,771.53646,745.498h.00147l36.41507.00147Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='717.365 620.997 698.875 620.996 690.078 549.675 717.369 549.676 717.365 620.997'
+                fill='#9f616a'
+            />
+            <path
+                d='M955.75964,769.46052l-59.622-.00221v-.75412A23.20778,23.20778,0,0,1,919.3442,745.498h.00148l36.41506.00147Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M901.30653,438.0961l18.63858-68.863-24.17978-13.59139H862.01439l-13.12731,14.04444s-21.932,15.25939-16.6236,32.16627l8.56367,38.50893-13.09738,61.16123-65.99064,91.51533L746.118,682.93028l61.49188,16.12041,66.49439-130.93034L906.344,713.09512l59.97741-3.01648L945.65408,534.29672s-5.89164-53.54942-36.77341-86.98486Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M939.79877,499.02537a10.00932,10.00932,0,0,0-1.28244-15.2944l-14.46166-97.5642-20.757,6.55425,20.77824,93.96478a10.06355,10.06355,0,0,0,15.72286,12.33957Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#9f616a'
+            />
+            <path
+                d='M927.62875,401.4788,905.111,404.07517a4.488,4.488,0,0,1-4.99972-4.603l.72936-22.58654a12.46412,12.46412,0,0,1,24.77233-2.78279l5.83678,21.75592a4.488,4.488,0,0,1-3.821,5.62009Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#cbcbcb'
+            />
+            <path
+                d='M841.14625,336.05834V307.04351a38.87987,38.87987,0,1,1,77.75974,0v29.01483a5.22857,5.22857,0,0,1-5.22267,5.22267H846.36891A5.22856,5.22856,0,0,1,841.14625,336.05834Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <circle
+                cx='872.0367'
+                cy='312.10021'
+                r='28.50536'
+                transform='translate(-276.51631 325.98852) rotate(-28.66322)'
+                fill='#9f616a'
+            />
+            <path
+                d='M831.94821,309.3647A30.7905,30.7905,0,0,1,862.70392,278.609h5.80325a30.79026,30.79026,0,0,1,30.75544,30.75572v.58029H886.999l-4.18278-11.71247L881.9798,309.945h-6.33793l-2.11037-5.9095-.42219,5.9095h-41.1611Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M869.23138,344.54377a5.14537,5.14537,0,0,1-.40887-5.46295c6.1495-11.6966,14.75959-33.30925,3.331-46.63794l-.82143-.95771h33.17834v49.81256l-30.14,5.31788a5.3353,5.3353,0,0,1-.92428.0816A5.20168,5.20168,0,0,1,869.23138,344.54377Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M564.2779,371.06852a11.47832,11.47832,0,0,0-1.154,1.39088l-54.05933-1.32268L502.758,359.61258l-18.636,7.24458,9.18192,21.48621a8.723,8.723,0,0,0,9.06747,5.23226l61.077-7.37823a11.44584,11.44584,0,1,0,.82955-15.12888Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#ffb7b7'
+            />
+            <polygon
+                points='154.374 568.836 172.149 568.836 180.605 500.275 154.371 500.276 154.374 568.836'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M383.51868,693.57253l35.00552-.00142h.00142A22.30946,22.30946,0,0,1,440.83388,715.879v.72493l-57.31414.00213Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='315.778 568.441 332.124 561.458 312.967 495.086 288.843 505.393 315.778 568.441'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M543.007,695.42455l32.191-13.75238.0013-.00056A22.30946,22.30946,0,0,1,604.47739,693.423l.28477.66666-52.706,22.51643Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='240.773 255.381 242.058 276.21 204.487 280.231 165.173 284.438 176.744 252.81 240.773 255.381'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M475.77146,393.39429c-.42593-.16718-42.86293-16.67463-68.92958-3.17691l-.641.33223-5.20008-25.59573-2.96314-55.02963,19.281-9.11064,5.30869-11.89143,36.15932-2.13977,7.43062,10.19038,25.29891,8.29126L476.396,393.64132Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#e4e4e4'
+            />
+            <path
+                d='M477.10036,343.72286l4.47253-35.14241,10.184-3.09971.1717.06336c.23506.08625,5.76976,2.20472,7.28473,9.8672,1.45429,7.35474,12.40655,53.12305,12.517,53.58465l.10968.45788-25.73309,10.29579Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#e4e4e4'
+            />
+            <path
+                d='M536.21761,686.9309l34.25472-7.02661-54.749-151.07208-18.73763-51.52846S485.10124,435.98844,479.73181,420l-3.24037-21.45282-76.70714,8.49049s-21.42411,31.46334-14.91562,82.7467L381.925,682.53927l34.25472,1.75665,28.692-194.40283,50.35736,106.57023s15.87209,45.47732,27.22811,55.33454Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M430.54561,414.1922a11.47765,11.47765,0,0,0-1.80457.09942l-35.78547-40.54086,4.16272-12.45995-17.98316-8.74-9.51292,21.34173a8.72306,8.72306,0,0,0,2.32826,10.20661l46.93716,39.771a11.44583,11.44583,0,1,0,11.658-9.67794Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M375.8954,358.55694,389.88534,318.708c1.57222-4.69748,3.54241-7.55121,5.856-8.4824a3.61333,3.61333,0,0,1,2.90324-.02715l6.27315,7.92549,5.8222,24.4585L395.81216,377.6961Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#e4e4e4'
+            />
+            <path
+                d='M468.49581,278.53619h-61.416a4.77041,4.77041,0,0,1-4.765-4.765V247.29874a35.473,35.473,0,0,1,70.94606,0v26.47241A4.77041,4.77041,0,0,1,468.49581,278.53619Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <circle cx='211.39854' cy='121.37287' r='26.00758' fill='#ffb7b7' />
+            <path
+                d='M481.65291,249.946H444.09853l-.38519-5.39168-1.92546,5.39168h-5.78256l-.76315-10.68617L431.4259,249.946h-11.189v-.52945a28.09227,28.09227,0,0,1,28.0605-28.06075h5.29474a28.09249,28.09249,0,0,1,28.06076,28.06075Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M443.79118,283.4778a4.86705,4.86705,0,0,1-.84329-.07446l-27.499-4.85189V233.10369H445.72l-.74945.8738c-10.42713,12.16076-2.5715,31.87961,3.03915,42.5513a4.6945,4.6945,0,0,1-.373,4.98426A4.74591,4.74591,0,0,1,443.79118,283.4778Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M773.19705,392.72009a33.504,33.504,0,0,1-5.15926.12031,1.12527,1.12527,0,0,0-1.23851.97665l-3.29091,24.68526a10.02141,10.02141,0,0,1-2.12316,4.89035,10.65824,10.65824,0,0,0-2.39918,8.14585v41.93238s-.21231,5.591,4.08355,5.591a62.10276,62.10276,0,0,0,15.48491,0s3.43951-.6511,3.43951-5.81039V431.96314a10.09918,10.09918,0,0,0-2.12316-8.20955,10.276,10.276,0,0,1-2.05945-4.862l-3.4183-25.20189A1.12526,1.12526,0,0,0,773.19705,392.72009Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#fff'
+            />
+            <path
+                d='M770.797,480.25286a62.674,62.674,0,0,1-7.773-.48345,4.17992,4.17992,0,0,1-3.09921-1.26857c-1.74822-1.81561-1.65095-4.92536-1.64594-5.057l-.00051-41.86223a11.4041,11.4041,0,0,1,2.56375-8.64261,9.30817,9.30817,0,0,0,1.96541-4.53556l3.29031-24.68a1.84294,1.84294,0,0,1,2.01689-1.58649,32.65836,32.65836,0,0,0,5.01158-.12095l.01969-.00173a1.83261,1.83261,0,0,1,1.94831,1.57924l3.41851,25.20318a9.5778,9.5778,0,0,0,1.91686,4.52242,10.77949,10.77949,0,0,1,2.27209,8.70689v41.22535c0,5.67973-3.97539,6.49837-4.01565,6.506A63.14218,63.14218,0,0,1,770.797,480.25286Zm-2.85541-86.70877a.42442.42442,0,0,0-.44026.36215l-3.29151,24.68966a10.73613,10.73613,0,0,1-2.27383,5.24121,9.99118,9.99118,0,0,0-2.24756,7.6142l.00535.08708v41.93238c-.00138.05219-.07654,2.67434,1.25389,4.05177a2.79106,2.79106,0,0,0,2.12195.83178l.08829.00552a61.46482,61.46482,0,0,0,15.30824,0c.06721-.01658,2.82016-.64448,2.82016-5.1085l.01141-41.41507a9.37632,9.37632,0,0,0-1.97439-7.63425,11.01274,11.01274,0,0,1-2.21249-5.21113l-3.419-25.20629a.41994.41994,0,0,0-.43628-.359,34.364,34.364,0,0,1-5.25572.12164A.55569.55569,0,0,0,767.94156,393.54409Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#3f3d56'
+            />
+            <path
+                className='fill-primary'
+                d='M534.32377,301.74921h5.21588a7.82029,7.82029,0,0,1,7.82029,7.82029v19.62507a7.8203,7.8203,0,0,1-7.8203,7.8203h-5.21588a7.82029,7.82029,0,0,1-7.82029-7.82029V309.56951A7.8203,7.8203,0,0,1,534.32377,301.74921Z'
+            />
+            <path
+                d='M596.19705,358.72009a33.504,33.504,0,0,1-5.15926.12031,1.12527,1.12527,0,0,0-1.23851.97665l-3.29091,24.68526a10.02141,10.02141,0,0,1-2.12316,4.89035,10.65824,10.65824,0,0,0-2.39918,8.14585v41.93238s-.21231,5.591,4.08355,5.591a62.10276,62.10276,0,0,0,15.48491,0s3.43951-.6511,3.43951-5.81039V397.96314a10.09918,10.09918,0,0,0-2.12316-8.20955,10.276,10.276,0,0,1-2.05945-4.862l-3.4183-25.20189A1.12526,1.12526,0,0,0,596.19705,358.72009Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#fff'
+            />
+            <path
+                d='M593.797,446.25286a62.674,62.674,0,0,1-7.773-.48345,4.17992,4.17992,0,0,1-3.09921-1.26857c-1.74822-1.81561-1.65095-4.92536-1.64594-5.057l-.00051-41.86223a11.4041,11.4041,0,0,1,2.56375-8.64261,9.30817,9.30817,0,0,0,1.96541-4.53556l3.29031-24.68a1.84294,1.84294,0,0,1,2.01689-1.58649,32.65836,32.65836,0,0,0,5.01158-.12095l.01969-.00173a1.83261,1.83261,0,0,1,1.94831,1.57924l3.41851,25.20318a9.5778,9.5778,0,0,0,1.91686,4.52242,10.77949,10.77949,0,0,1,2.27209,8.70689v41.22535c0,5.67973-3.97539,6.49837-4.01565,6.506A63.14218,63.14218,0,0,1,593.797,446.25286Zm-2.85541-86.70877a.42442.42442,0,0,0-.44026.36215l-3.29151,24.68966a10.73613,10.73613,0,0,1-2.27383,5.24121,9.99118,9.99118,0,0,0-2.24756,7.6142l.00535.08708v41.93238c-.00138.05219-.07654,2.67434,1.25389,4.05177a2.79106,2.79106,0,0,0,2.12195.83178l.08829.00552a61.46482,61.46482,0,0,0,15.30824,0c.06721-.01658,2.82016-.64448,2.82016-5.1085l.01141-41.41507a9.37632,9.37632,0,0,0-1.97439-7.63425,11.01274,11.01274,0,0,1-2.21249-5.21113l-3.419-25.20629a.41994.41994,0,0,0-.43628-.359,34.364,34.364,0,0,1-5.25572.12164A.55569.55569,0,0,0,590.94156,359.54409Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#3f3d56'
+            />
+            <path
+                className='fill-primary'
+                d='M357.32377,267.74921h5.21588a7.82029,7.82029,0,0,1,7.82029,7.82029v19.62507a7.8203,7.8203,0,0,1-7.8203,7.8203h-5.21588a7.82029,7.82029,0,0,1-7.82029-7.82029V275.56951A7.8203,7.8203,0,0,1,357.32377,267.74921Z'
+            />
+            <path
+                d='M566.19705,322.72009a33.504,33.504,0,0,1-5.15926.12031,1.12527,1.12527,0,0,0-1.23851.97665l-3.29091,24.68526a10.02141,10.02141,0,0,1-2.12316,4.89035,10.65824,10.65824,0,0,0-2.39918,8.14585v41.93238s-.21231,5.591,4.08355,5.591a62.10276,62.10276,0,0,0,15.48491,0s3.43951-.6511,3.43951-5.81039V361.96314a10.09918,10.09918,0,0,0-2.12316-8.20955,10.276,10.276,0,0,1-2.05945-4.862l-3.4183-25.20189A1.12526,1.12526,0,0,0,566.19705,322.72009Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#fff'
+            />
+            <path
+                d='M563.797,410.25286a62.674,62.674,0,0,1-7.773-.48345,4.17992,4.17992,0,0,1-3.09921-1.26857c-1.74822-1.81561-1.65095-4.92536-1.64594-5.057l-.00051-41.86223a11.4041,11.4041,0,0,1,2.56375-8.64261,9.30817,9.30817,0,0,0,1.96541-4.53556l3.29031-24.68a1.84294,1.84294,0,0,1,2.01689-1.58649,32.65836,32.65836,0,0,0,5.01158-.12095l.01969-.00173a1.83261,1.83261,0,0,1,1.94831,1.57924l3.41851,25.20318a9.5778,9.5778,0,0,0,1.91686,4.52242,10.77949,10.77949,0,0,1,2.27209,8.70689v41.22535c0,5.67973-3.97539,6.49837-4.01565,6.506A63.14218,63.14218,0,0,1,563.797,410.25286Zm-2.85541-86.70877a.42442.42442,0,0,0-.44026.36215l-3.29151,24.68966a10.73613,10.73613,0,0,1-2.27383,5.24121,9.99118,9.99118,0,0,0-2.24756,7.6142l.00535.08708v41.93238c-.00138.05219-.07654,2.67434,1.25389,4.05177a2.79106,2.79106,0,0,0,2.12195.83178l.08829.00552a61.46482,61.46482,0,0,0,15.30824,0c.06721-.01658,2.82016-.64448,2.82016-5.1085l.01141-41.41507a9.37632,9.37632,0,0,0-1.97439-7.63425,11.01274,11.01274,0,0,1-2.21249-5.21113l-3.419-25.20629a.41994.41994,0,0,0-.43628-.359,34.364,34.364,0,0,1-5.25572.12164A.55569.55569,0,0,0,560.94156,323.54409Z'
+                transform='translate(-233.67861 -130.53948)'
+                fill='#3f3d56'
+            />
+            <path
+                className='fill-primary'
+                d='M327.32377,231.74921h5.21588a7.82029,7.82029,0,0,1,7.82029,7.82029v19.62507a7.8203,7.8203,0,0,1-7.8203,7.8203h-5.21588a7.82029,7.82029,0,0,1-7.82029-7.82029V239.56951A7.8203,7.8203,0,0,1,327.32377,231.74921Z'
+            />
+            <circle cx='56.18747' cy='46' r='25' fill='#f0f0f0' />
+            <circle cx='632.18747' cy='25' r='25' fill='#f0f0f0' />
+            <circle cx='346.18747' cy='95' r='18' fill='#f0f0f0' />
+        </svg>
+    )
+}
+
+export const PeopleSelfieIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 766.80194 500.94794'
+        >
+            <path
+                id='ae90ee98-50a6-4343-bcab-c7717b4ab754-174'
+                data-name='Path 133'
+                d='M880.50872,613.59857a158.39405,158.39405,0,0,1-7.4,43.785c-.1.329-.211.653-.319.982h-27.613c.029-.295.059-.624.088-.982,1.841-21.166-8.677-148.453-21.369-170.483C825.00771,488.68756,882.64971,546.34454,880.50872,613.59857Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#f0f0f0'
+            />
+            <path
+                id='ace95f5d-9239-4670-9a5e-89bd21d3a3cd-175'
+                data-name='Path 134'
+                d='M878.43667,657.38354c-.231.329-.471.658-.717.982h-20.716c.157-.28.339-.609.55-.982,3.422-6.176,13.551-24.642,22.953-43.785,10.1-20.572,19.374-41.924,18.593-49.652C899.34167,565.68853,906.33364,618.82256,878.43667,657.38354Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#f0f0f0'
+            />
+            <path
+                id='b1e24b1a-80e9-492b-ae8f-155b8492cb31-176'
+                data-name='Path 133'
+                d='M947.93193,633.82353a86.83454,86.83454,0,0,1-4.0568,24.00366c-.05483.18035-.11568.358-.17488.53834H928.56233c.0159-.16171.03234-.3421.04825-.53834,1.00927-11.60357-4.75688-81.38443-11.71483-93.46166C917.50535,565.34522,949.10565,596.95375,947.93193,633.82353Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#f0f0f0'
+            />
+            <path
+                id='b092796d-fbce-4787-85db-a435af6317d1-177'
+                data-name='Path 134'
+                d='M946.796,657.82719c-.12663.18035-.25821.36074-.39307.53834H935.04606c.08607-.15351.18584-.33387.30151-.53834,1.876-3.3858,7.42888-13.50918,12.58322-24.00366,5.537-11.27792,10.62115-22.98344,10.193-27.22C958.25646,607.55845,962.08958,636.68742,946.796,657.82719Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#f0f0f0'
+            />
+            <path
+                id='f749498b-12ad-4a63-9639-cd9a5a319912-178'
+                data-name='Path 133'
+                d='M804.93193,633.82353a86.83454,86.83454,0,0,1-4.0568,24.00366c-.05483.18035-.11568.358-.17488.53834H785.56233c.0159-.16171.03234-.3421.04825-.53834,1.00927-11.60357-4.75688-81.38443-11.71483-93.46166C774.50535,565.34522,806.10565,596.95375,804.93193,633.82353Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#f0f0f0'
+            />
+            <path
+                id='a1477d31-24e2-4198-a2a2-41d3c8f4997f-179'
+                data-name='Path 134'
+                d='M803.796,657.82719c-.12663.18035-.25821.36074-.39307.53834H792.04606c.08607-.15351.18584-.33387.30151-.53834,1.876-3.3858,7.42888-13.50918,12.58322-24.00366,5.537-11.27792,10.62115-22.98344,10.193-27.22C815.25646,607.55845,819.08958,636.68742,803.796,657.82719Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#f0f0f0'
+            />
+            <polygon
+                points='314.574 442.1 329.13 442.1 336.055 385.955 314.572 385.956 314.574 442.1'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M527.87085,638.71326l28.66618-.00116h.00116a18.26932,18.26932,0,0,1,18.26834,18.26805v.59365l-46.93482.00174Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='235.763 442.1 250.319 442.1 257.243 385.955 235.76 385.956 235.763 442.1'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M449.05923,638.71326l28.66618-.00116h.00116a18.26932,18.26932,0,0,1,18.26834,18.26805v.59365l-46.93481.00174Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='229.832 187.638 229.832 335.879 233.585 431.579 264.547 431.579 274.867 251.438 295.508 325.559 308.644 429.702 346.173 427.826 320.841 187.638 229.832 187.638'
+                fill='#2f2e41'
+            />
+            <circle cx='265.82446' cy='31.89109' r='22.51761' fill='#ffb8b8' />
+            <polygon
+                points='256.468 60.275 250.473 66.606 244.844 193.268 311.458 191.391 295.508 80.68 281.435 67.193 274.294 60.275 256.468 60.275'
+                className='fill-primary'
+            />
+            <path
+                d='M462.79141,275.47713l5.0565-8.20674,1.51113,17.58907s12.197,39.40581,11.25881,74.12045,4.69116,116.34093-2.8147,117.27917-39.40581,2.8147-39.40581-3.75294-1.87647-181.07907-1.87647-181.07907Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M496.93315,267.27039l12.76994,10.0832,28.147,10.32057s0,120.09387,5.6294,148.24087,15.95,30.96171,2.81471,31.89994-12.197-3.75293-12.197-3.75293,9.38233-46.91168-2.8147-56.294-35.65288-63.79988-33.77641-93.82336S495.05668,267.27039,496.93315,267.27039Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M461.53113,240.30161s1.91318-3.35983,1.4684-14.5839,16.6468-4.87588,30.92861-5.9466S505.18674,238.51,505.18674,238.51s1.03034-1.17758,3.2363-10.73827-2.80275-19.0883-2.80275-19.0883c-1.47514-10.0465-9.42511-6.82659-9.42511-6.82659,3.82861,1.49044,3.09328,4.67734,3.09328,4.67734-3.24078-5.682-11.19076-2.46206-11.19076-2.46206-10.603-6.65362-21.349,2.10836-21.349,2.10836-12.80892,2.907-5.297,13.09344-5.297,13.09344C448.57021,225.78368,461.53113,240.30161,461.53113,240.30161Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M438.42881,458.93962a9.49863,9.49863,0,0,0-1.68433-14.46727l4.05627-21.32359-12.35709-5.62547-5.30529,30.19349a9.55009,9.55009,0,0,0,15.29044,11.22284Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M441.212,291.4271l-2.49124-1.33569S425.26207,295.18,426.2003,298.933s-7.50587,135.10561-4.69117,139.79677,25.33231.93824,25.33231.93824Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M548.31956,450.09621A9.49861,9.49861,0,0,1,547.986,435.535l-6.96762-20.55726,11.45989-7.28112,9.43188,29.169a9.55009,9.55009,0,0,1-13.59056,13.23051Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M537.32138,432.1739s22.82029.60126,24.95884-4.43423-24.39848-133.95644-23.98853-137.8031-13.62341-7.02415-13.62341-7.02415l-2.28247,1.66752Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M638.18107,437.97478l-14.12821-69.2292L617.84655,317.613,604.097,326.50775l1.89939,56.34823L627.942,441.23817a7.64968,7.64968,0,1,0,10.23905-3.26339Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M506.04064,443.3277l3.91075-70.54781-1.29534-51.4916,18.17731,6.80078,3.03452,56.02546-13.22247,60.95292a7.64968,7.64968,0,1,1-10.60477-1.73975Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#ffb7b7'
+            />
+            <polygon
+                points='327.101 447.07 336.389 447.07 340.808 411.244 327.1 411.245 327.101 447.07'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M541.7415,645.40318l18.29158-.00074h.00074a11.65746,11.65746,0,0,1,11.65683,11.65664v.3788l-29.9486.00111Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='362.281 442.96 371.46 444.38 381.305 409.651 367.757 407.556 362.281 442.96'
+                fill='#ffb7b7'
+            />
+            <path
+                d='M577.41309,640.96639l18.0766,2.79617.00073.00011a11.65745,11.65745,0,0,1,9.73738,13.302l-.05792.37435-29.59659-4.57823Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M541.38251,637.739l-.04814-.25982L507.71,455.68343a68.67452,68.67452,0,0,1,6.06847-42.94446l20.35148-40.70268,60.72681,8.22705,7.16679,51.821,25.53977,98.85618-.02529.08047L594.81606,634.96145l-21.4455,1.4106,18.99678-92.78153L557.58325,464.585l2.45346,171.74047Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M596.27034,386.89527l-.40079-.628c-.072-.11324-7.44846-11.22672-27.52144-3.80643-20.49131,7.57635-36.369-7.05672-36.52708-7.20531l-.0786-.07415-4.64063-41.362-5.47828-33.57784,17.43861-3.05945,8.57351-5.14374,34.09245-.607,8.97143,7.05614,14.72935,1.60518,1.34838,30.59541-.00618.03334Z'
+                transform='translate(-217.00961 -201.36553)'
+                className='fill-primary'
+            />
+            <path
+                d='M601.15642,333.64753V307.44946l4.142-7.55911.16842-.00517a12.41671,12.41671,0,0,1,8.57883,3.89782c4.16612,4.23294,6.05339,11.28535,5.60948,20.96122l-.00819.177Z'
+                transform='translate(-217.00961 -201.36553)'
+                className='fill-primary'
+            />
+            <path
+                d='M528.9619,330.7999l-21.111-2.5335-.01652-.24315c-.72025-10.52745,1.1173-18.34207,5.46146-23.22627a13.39066,13.39066,0,0,1,8.65069-4.71033l.12186-.00575,5.71481,5.19518Z'
+                transform='translate(-217.00961 -201.36553)'
+                className='fill-primary'
+            />
+            <path
+                d='M590.07389,280.87513H534.52208a4.31491,4.31491,0,0,1-4.31-4.31006V256.45149a32.086,32.086,0,1,1,64.17191,0v20.11358A4.31491,4.31491,0,0,1,590.07389,280.87513Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <circle cx='346.578' cy='55.78733' r='23.5243' fill='#ffb7b7' />
+            <path
+                d='M585.94786,257.88817H554.89008a2.39792,2.39792,0,0,1-2.376-2.098l-.4022-3.21664a1.43658,1.43658,0,0,0-2.72745-.42932l-2.036,4.36243a2.40372,2.40372,0,0,1-2.16976,1.3815H540.461a2.394,2.394,0,0,1-2.38371-2.62269l1.86016-19.42985a2.41285,2.41285,0,0,1,1.3684-1.95019c14.22351-6.64,28.64717-6.63157,42.87068.02525a2.398,2.398,0,0,1,1.35484,1.838l2.78662,19.4046a2.3938,2.3938,0,0,1-2.37016,2.73493Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M598.72673,440.175a8.59205,8.59205,0,0,0,4.60352-12.34441l20.135-22.95337-15.71777-2.16839-16.71083,22.12936a8.63859,8.63859,0,0,0,7.69009,15.33681Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#9f616a'
+            />
+            <polygon
+                points='406 442.234 391.957 442.233 385.276 388.067 406.003 388.068 406 442.234'
+                fill='#9f616a'
+            />
+            <path
+                d='M626.591,657.21212l-45.28071-.00167v-.57273a17.62547,17.62547,0,0,1,17.62451-17.62424h.00112l27.65592.00112Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='479.511 430.482 466.099 434.646 443.656 384.897 463.451 378.752 479.511 430.482'
+                fill='#9f616a'
+            />
+            <path
+                d='M703.97745,643.786,660.733,657.21212l-.16984-.547a17.62544,17.62544,0,0,1,11.60538-22.05793l.00107-.00034,26.41225-8.20014Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                id='f1066638-9b86-4717-bcc8-db5f2da468bd-180'
+                data-name='Path 973'
+                d='M612.25556,393.87135l-20.798,23.81941,16.28228,1.46511Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#e6e6e6'
+            />
+            <path
+                id='af262a15-c86f-47ab-a56c-b4ab6ae66702-181'
+                data-name='Path 975'
+                d='M604.129,424.57384s-7.22426,3.6117-8.1274,19.86651S598.711,503.1402,598.711,503.1402s-3.6117,18.96341,0,36.12048-3.6117,79.46267.90315,79.46267,27.99307,2.70944,28.89621,0,1.8063-43.34475,1.8063-43.34475,7.22425-20.76965,0-33.40848c0,0,25.22372,44.0448,41.53849,77.659,3.57061,7.35673,30.70248-.90314,26.18763-9.03052s-15.35082-45.151-15.35082-45.151-8.1274-27.09078-22.57508-38.82991l7.22426-57.79327s15.35165-38.82991,6.32283-44.2479S604.129,424.57384,604.129,424.57384Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <circle
+                id='fbcb1e7b-18f5-4394-a5dd-21324d3d96d7'
+                data-name='Ellipse 182'
+                cx='414.20933'
+                cy='52.53703'
+                r='23.47823'
+                fill='#a0616a'
+            />
+            <path
+                id='f91d5893-36b1-4b4e-ad68-e940b5b31ca4-182'
+                data-name='Path 976'
+                d='M657.4066,279.18705,628.219,296.34414s-11.176,6.29379-15.35082,17.15709c-4.44992,11.57762-1.86353,27.87858,0,31.60563,3.6117,7.22426-1.51493,25.28452-1.51493,25.28452l-4.51484,41.53847s-16.254,15.3508-3.6117,17.15709,35.21735-.90315,48.76275,0,28.89619,2.70944,24.38134-6.32283S666.438,407.41242,671.856,389.3556c4.2397-14.1324,4.056-66.97336,3.77576-89.2887a17.96455,17.96455,0,0,0-8.72465-15.18079Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#e5e5e5'
+            />
+            <path
+                id='ab5b7163-34aa-4542-8f5b-346ef47d8115-183'
+                data-name='Path 980'
+                d='M644.31661,307.633l2.70857,55.08386-26.25344,53.16567-4.45248-1.69435,27.09079-49.66586Z'
+                transform='translate(-217.00961 -201.36553)'
+                opacity='0.1'
+                // style='isolation:isolate'
+            />
+            <path
+                id='a0747de6-ab5b-4f6b-918c-fa029658d911-184'
+                data-name='Path 982'
+                d='M674.11256,374.45511v-6.32283L643.41348,418.703Z'
+                transform='translate(-217.00961 -201.36553)'
+                opacity='0.1'
+                // style='isolation:isolate'
+            />
+            <path
+                id='fa11fa6f-8d0b-4738-bd9e-1b384a1b66c7-185'
+                data-name='Path 983'
+                d='M614.55571,232.15286l-3.87829-1.55167s8.11715-8.93485,19.40684-8.15817l-3.17595-3.49637s7.76256-3.10758,14.82022,5.04973c3.71,4.28843,8.00179,9.32961,10.68046,15.00733h4.15683l-1.7345,3.8202,6.07162,3.8202-6.23227-.68354a21.13815,21.13815,0,0,1-.58957,9.89354l.1709,3.01958s-7.22767-11.18375-7.22767-12.73542v3.88513s-3.88171-3.49637-3.88171-5.82727l-2.11729,2.71968-1.05866-4.27218-13.0558,4.27218,2.11559-3.49806-8.11715,1.16547,3.17595-4.27219s-9.17409,5.04973-9.527,9.32361-4.93951,9.71153-4.93951,9.71153l-2.11729-3.88512S604.32637,237.9801,614.55571,232.15286Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M629.74558,442.7301a8.592,8.592,0,0,0,2.33926-12.96552l15.741-26.16289-15.85307.656-12.51746,24.74416a8.63859,8.63859,0,0,0,10.29029,13.72826Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#9f616a'
+            />
+            <path
+                id='b6ea18e9-d072-4df0-8b87-7f4e4805b6d9-186'
+                data-name='Path 981'
+                d='M665.53487,290.92788l6.77226-1.35428s20.31769,14.89967,13.99655,44.699S651.98947,409.223,651.98947,409.223s-6.32283,8.1274-8.1274,9.93284-5.418,0-3.6117,2.70941-2.70941,4.51484-2.70941,4.51484-19.86653,0-18.06024-7.22425S651.9895,360.456,651.9895,360.456l-4.51484-47.8596S643.86121,289.12161,665.53487,290.92788Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#e5e5e5'
+            />
+            <path
+                d='M982.81154,659.36553h-611a1,1,0,1,1,0-2h611a1,1,0,1,1,0,2Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#cacaca'
+            />
+            <path
+                d='M233.11656,243.13545v.34417H223.688a5.457,5.457,0,0,0-5.457,5.457v19.9756a5.457,5.457,0,0,0,5.457,5.457h51.7256a5.457,5.457,0,0,0,5.457-5.457V248.93661a5.457,5.457,0,0,0-5.457-5.457h-35.5857v-.34417Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M370.21772,326.10326c.219,29.48-18.61722,25.02175-41.58273,25.02175s-39.58048,4.391-41.58274-25.02175c-3.18165-46.73773,9.094-81.73773,41.58274-81.73773C362.243,244.36553,369.8706,279.36553,370.21772,326.10326Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M244.44075,263.57394a11.45606,11.45606,0,0,1-7.07667,16.078L235.26856,320.309l-17.03821-12.54026,4.60208-36.68608a11.51816,11.51816,0,0,1,21.60832-7.50876Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#a0616a'
+            />
+            <path
+                d='M398.08363,493.34215a11.45606,11.45606,0,0,0-5.00417-16.83867L396.1,435.90466l-18.47655,10.30417-.03567,36.97359a11.51816,11.51816,0,0,0,20.49588,10.15973Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#a0616a'
+            />
+            <polygon
+                points='95.985 487.409 109.952 487.408 116.596 433.535 95.982 433.536 95.985 487.409'
+                fill='#a0616a'
+            />
+            <path
+                d='M309.43154,684.21448l27.50589-.00111h.00111a17.52859,17.52859,0,0,1,17.52909,17.52881v.56962l-45.03526.00167Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='184.846 487.409 198.813 487.408 205.458 433.535 184.844 433.536 184.846 487.409'
+                fill='#a0616a'
+            />
+            <path
+                d='M398.29327,684.21448l27.50589-.00111h.00112a17.52859,17.52859,0,0,1,17.52908,17.52881v.56962l-45.03526.00167Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M393.884,470.97338l-.6069-.11014c-5.08742-.92119-10.23658-1.8524-14.76384-2.82588l-.44808-.09679-.00166-.45838c-.0776-20.65007-.15548-41.2968-.2698-62.13045a131.32591,131.32591,0,0,1-13.13061-70.39337l.02559-.25589.21-.15019a13.50463,13.50463,0,0,1,13.60985-1.05248,15.08909,15.08909,0,0,1,7.74586,11.45483c4.117,22.438,8.37918,45.67359,13.13284,68.56991l.01752.08344-.00806.08566c-1.76117,18.67307-3.58214,37.98252-5.45123,56.6656Z'
+                transform='translate(-217.00961 -201.36553)'
+                className='fill-primary'
+            />
+            <path
+                d='M384.7789,669.57388l-.111-.52512c-6.534-30.90669-12.84135-56.07484-18.94119-80.41413-7.30975-29.16888-14.23677-56.80912-21.23-92.57766.96041,45.10619,3.19107,77.5126,5.75523,114.76194,1.15789,16.81956,2.355,34.21208,3.56962,54.51615l.03727.61969-.62025-.01558c-21.85941-.55961-44.25536-1.91025-65.91424-3.21527l-.573-.03449.03894-.573c4.10058-60.4416,5.38558-99.7169,6.32374-128.39515,1.97951-60.5039,2.28379-69.80372,24.82131-121.00892.85416-1.94251,5.951-7.68217,11.55468-6.8077l44.34214,6.90116.0776.3905c1.25162,6.30261,7.46078,25.42737,16.05468,51.89721C408.33607,521.679,436.09924,607.192,444.77242,658.16466l.09818.57742-.57964.08233c-12.59742,1.79009-21.66389,3.65473-30.43191,5.45706-8.29574,1.70666-16.87351,3.47005-28.54974,5.21231Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M346.23983,428.801c-11.835.00112-23.533-4.10531-31.91883-12.10122a9.46491,9.46491,0,0,1-4.66938-7.35619,9.5971,9.5971,0,0,1,2.2543-7.78675c-8.22175-9.95178-16.71052-20.23732-24.83632-30.71088l-.22446-.28926.26006-.4962c5.31633-10.1576,10.81372-20.66009,16.09306-31.038a5.42314,5.42314,0,0,1,4.42239-3.34878,6.305,6.305,0,0,1,4.86881,1.08807l.01057-.25255c.22056-5.25346.4414-10.50693.66808-15.48672l.0242-.5318.53236-.01223c10.15092-.22363,20.64729-.45615,30.95758-1.06694l.39246-.02336,5.11356,11.407A41.825,41.825,0,0,1,374.1849,335.666l.33766.18023-.03922.3805a232.24542,232.24542,0,0,0,1.33117,58.8818c-.15965,8.8993-.33126,17.99552-2.13693,25.246l-.05257.21139-.18079.12238A48.64723,48.64723,0,0,1,346.23983,428.801Z'
+                transform='translate(-217.00961 -201.36553)'
+                className='fill-primary'
+            />
+            <path
+                d='M302.67675,379.31247l-.68116-.31263c-7.95224-3.98961-15.92951-8.41534-23.644-12.69533l-1.06582-.208c-20.31214-4.07751-31.50107-6.32264-49.82619-17.06877l-.14074-.08344-.07565-.14464c-9.17577-17.52825-11.7352-32.85031-9.44362-56.53209l.04589-.47506.47506-.03782c1.98647-.16132,5.22427-.26924,8.35527-.37382,3.04616-.10124,6.19608-.20694,8.12719-.35824l.44808-.03561.13851.42722c2.40117,7.41516,4.441,14.333,6.60077,21.65694,2.21982,7.52863,4.51418,15.30982,7.30891,23.95546,10.23242,2.94048,37.66321,2.06156,46.19231.90006a15.26092,15.26092,0,0,1,13.32559,6.05895l.15826.198-.04088.25032c-1.87771,11.45928-4.03384,22.98643-6.11876,34.1342Z'
+                transform='translate(-217.00961 -201.36553)'
+                className='fill-primary'
+            />
+            <circle
+                cx='328.12367'
+                cy='288.18638'
+                r='27.98122'
+                transform='translate(-299.14414 236.50296) rotate(-61.33681)'
+                fill='#a0616a'
+            />
+            <path
+                d='M296.16628,286.51786v0h11.0671l4.88244-13.671.97626,13.671h5.28963l2.84813-7.97477.56963,7.97477H361.1037v0a29.62055,29.62055,0,0,0-29.62055-29.62054h-5.69633A29.62054,29.62054,0,0,0,296.16628,286.51786Z'
+                transform='translate(-217.00961 -201.36553)'
+                fill='#2f2e41'
+            />
+        </svg>
+    )
+}
+
+export const PeopleIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 838 638.00376'
+        >
+            <path
+                d='M689.13548,323.255a5.09028,5.09028,0,0,0,3.68913-8.8227c-.12836-.51014-.22044-.8765-.34879-1.38664q.06888-.16659.13859-.33286c4.6863-11.17449,20.5687-11.09739,25.21513.09369,4.12406,9.93267,9.37471,19.88226,10.66755,30.38441a40.44007,40.44007,0,0,1-.70989,13.908,162.132,162.132,0,0,0,14.74848-67.33847,156.48736,156.48736,0,0,0-.97069-17.45764q-.80412-7.128-2.23118-14.15449a164.01034,164.01034,0,0,0-32.52468-69.51166,43.64774,43.64774,0,0,1-18.153-18.83392,33.29411,33.29411,0,0,1-3.02794-9.09827c.88371.11592,3.3322-13.34322,2.66567-14.169,1.23157-1.86885,3.43587-2.79784,4.78094-4.62148,6.68954-9.06961,15.906-7.486,20.71739,4.83888,10.27825,5.18745,10.37774,13.79049,4.07105,22.06463-4.01234,5.26406-4.56364,12.38693-8.08418,18.02253.36227.46365.739.91282,1.10112,1.37648a165.06913,165.06913,0,0,1,17.22906,27.29491c-1.36865-10.69089.652-23.57059,4.09674-31.873,3.9213-9.45971,11.27119-17.4267,17.74364-25.6046,7.77447-9.82292,23.71674-5.536,25.08643,6.91609q.01989.18086.03883.36158-1.44205.8135-2.8232,1.727a6.90459,6.90459,0,0,0,2.78379,12.56506l.1408.0217a68.70125,68.70125,0,0,1-1.81092,10.27183c8.30335,32.11122-9.623,43.80679-35.21939,44.33218-.565.28972-1.11566.57943-1.6807.8548a168.48694,168.48694,0,0,1,9.0693,42.68053,159.80436,159.80436,0,0,1-.11592,25.80246l.04349-.30421a42.227,42.227,0,0,1,14.41527-24.38271c11.09331-9.11275,26.76607-12.46849,38.73379-19.79336,5.76092-3.526,13.14312,1.03057,12.13769,7.70959q-.02407.1599-.04894.31988a46.37722,46.37722,0,0,0-5.201,2.50624q-1.44205.8137-2.8232,1.727a6.90468,6.90468,0,0,0,2.7838,12.56519l.14091.02164c.10137.01448.18834.029.2896.04359a68.7599,68.7599,0,0,1-12.6476,19.80451c-5.19187,28.03184-27.491,30.69143-51.34422,22.52828h-.01453a168.43679,168.43679,0,0,1-11.31491,33.03181H690.14836c-.145-.4491-.27531-.91275-.40575-1.36192a45.9261,45.9261,0,0,0,11.18457-.66624c-2.999-3.67994-5.99787-7.38886-8.99686-11.06867a2.51118,2.51118,0,0,1-.18835-.21734c-1.52118-1.88328-3.05689-3.75226-4.57807-5.6356l-.00081-.00229a67.27051,67.27051,0,0,1,1.97111-17.13665Z'
+                transform='translate(-181 -130.99812)'
+                fill='#f2f2f2'
+            />
+            <path
+                d='M818.913,219.8102c-7.51576-27.40163,53.10245-50.89907,69.02069-28.83249,2.21348-3.1372,10.39357-5.04278,14.14681-4.08041s6.90971,3.36827,9.98934,5.69723c4.23443,3.25279,8.62274,6.62107,11.22108,11.29814,2.61772,4.65781,3.02183,11.0095-.50033,15.05138-2.79088,3.21426-11.02877,5.293-15.22466,6.13988-2.92551.59667-5.38921,1.07785-7.50637,1.42428a14.35164,14.35164,0,0,1-2.50224-9.48889,11.64833,11.64833,0,0,0-5.08119,8.5458c-.05779.71214,23.65486,87.96005,23.65486,87.96005a51.72043,51.72043,0,0,1-8.95,4.02262,112.54468,112.54468,0,0,1-13.95427-30.41067,97.45318,97.45318,0,0,0,.26949,21.65318,119.1445,119.1445,0,0,0,2.0402,11.991c-28.178,6.08221-70.56044,4.08053-88.99936-21.7878q-.00491-.16053-.00929-.3213a4.40184,4.40184,0,0,0-8.50333-1.53505c-.44736,1.1413-.85908,2.12916-1.2265,2.91492C791.42794,280.84323,798.37617,258.61265,818.913,219.8102Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M770.96283,679.20736v.89737A3.809,3.809,0,0,1,770.96283,679.20736Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M798.5576,524.13266a15.32366,15.32366,0,0,1,2.183-.424Q799.47155,523.91547,798.5576,524.13266Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M797.919,524.2954c.19365-.0543.40921-.1086.63857-.16274C798.34394,524.18828,798.13849,524.2292,797.919,524.2954Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M1001.7905,768.99812H898.50437a8.21329,8.21329,0,0,1-2.2114-10.1149c2.82162-6.47111-2.70755-12.38425-2.70755-12.38425a17.5668,17.5668,0,0,1-3.53986-16.19126c2.73658-9.55783-7.98759-36.535-7.98759-36.535s-5.78127,34.83426-12.86549,75.22543H770.96283a18.25939,18.25939,0,0,1,0-18.48907c5.72448-9.70643,0-11.56794,0-11.56794s-5.72448-10.86155,0-18.748,0-40.08836,0-40.08836c.21572,1.42376,1.84656-1.16486,4.35751-6.59238,3.82253-8.25772,9.66421-23.10776,15.56624-40.348,4.21947-12.30455,8.47344-25.84306,12.02848-39.07958h.00864q1.39788-5.216,2.649-10.33723c1.63084-6.69591,3.0287-13.202,4.09-19.31114,2.30386-13.15884,3.06319-24.46254,1.29431-31.8142-1.4669-6.1178-4.67677-9.49162-10.21641-8.91351,2.96829-.49184,7.31716-.95778,12.67563-1.39785.96642-.07768,1.96733-.15531,3.00279-.24163,39.96837-3.02006,125.72089-4.7544,125.72089-4.7544L968.40587,564.35l33.26382,59.58155S1000.76368,703.8597,1001.7905,768.99812Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M891.54674,314.01971,831.0329,301.18327l-12.01528,24.1852L778.55553,348.7939s26.08156,106.64135,25.93277,121.63984c-.37777,38.079-61.30588,123.44042-15.28485,116.87355s55.36918-34.07334,70.14317-29.724c13.3243,3.92253,29.59359,21.37211,40.48552,8.89216s14.72832,1.93129,21.61565,2.42934c3.27307.23669,1.30185-15.75881,9.53707.49707s51.90445-5.62623,37.10363-20.42705c-24.65-24.65-29.65972-206.36132-29.65972-206.36132l-27.4406-8.44326Z'
+                transform='translate(-181 -130.99812)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M777.41122,701.29557a14.526,14.526,0,0,0-1.49594-20.20536,13.9357,13.9357,0,0,0-1.866-1.351l10.11793-49.6564-23.39032-13.99921-6.50593,69.52164a14.48576,14.48576,0,0,0,3.33874,17.36689,13.86282,13.86282,0,0,0,19.556-1.38555Q777.29042,701.44274,777.41122,701.29557Z'
+                transform='translate(-181 -130.99812)'
+                fill='#a0616a'
+            />
+            <path
+                d='M850.0367,439.55249c-.14673-1.59635-2.1831-2.32114-5.00472-3.3221a30.95043,30.95043,0,0,1-7.99881-3.788c-6.24726-4.51282-3.788-9.23272-.38831-15.75606.66443-1.26845,1.35472-2.58862,2.00188-3.96061,1.92419-4.06416-1.139-20.82982-9.1206-49.83967-2.45922-8.93073-2.45058-15.23835-4.15042-19.42332-1.69126-4.18492-5.08235-6.24721-14.32376-6.67864a42.251,42.251,0,0,0-30.41634,10.81184c-.2243.19845-.45729.41417-.68165.62125q-.6342.59541-1.24254,1.21668a42.21984,42.21984,0,0,0-12.27006,29.02707l-.04318,1.85521-1.90692,88.63448-5.35848,77.77956-7.3862,107.14316,5.47065.26754L779.048,655.228l2.649.12945,6.28171-26.35218,7.55879-31.71065,2.77848-11.64881,8.62012-36.13721,4.021-16.88647,2.45922-10.31136L814.4172,518.1l4.88385-20.46737c.11217-.43144,11.34682-44.75732,24.29858-52.16944C848.43171,442.702,850.17474,441.097,850.0367,439.55249Z'
+                transform='translate(-181 -130.99812)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M995.35486,700.7676a14.526,14.526,0,0,1-.68275-20.24915,13.93416,13.93416,0,0,1,1.71013-1.54361L980.98979,630.6923l21.75153-16.43034,13.93485,68.4208a14.48576,14.48576,0,0,1-1.45425,17.625,13.86281,13.86281,0,0,1-19.59166.72275Q995.49076,700.901,995.35486,700.7676Z'
+                transform='translate(-181 -130.99812)'
+                fill='#a0616a'
+            />
+            <path
+                d='M1005.99554,595.99963S989.031,536.3119,994.20466,529.85028s-10.10621-7.7235-3.52169-19.75745-2.50376-14.04665-2.50376-14.04665l-4.93184-27.66868-8.81682-67.96132-.26734-2.02117-2.56448-19.77218a42.34076,42.34076,0,0,0-16.696-28.61684c-.23307-.17843-.47114-.34983-.70925-.52125q-1.62813-1.17213-3.32715-2.172a42.26284,42.26284,0,0,0-28.10644-5.33868c-18.28571,2.84786-12.59457,9.6669-15.56046,27.935-4.819,29.7033-6.0699,46.69679-3.71994,50.53631.79163,1.2929,1.62357,2.5298,2.41661,3.71735,4.08278,6.129,7.03442,10.55052,1.30465,15.7181a31.25081,31.25081,0,0,1-7.54427,4.6162c-2.69312,1.304-4.64259,2.23962-4.61646,3.84262.03034,1.55289,1.934,2.95523,7.0314,5.179,6.17691,2.70319,12.81133,12.92324,18.24528,23.36341a257.87976,257.87976,0,0,1,11.51823,25.89924L978.53972,634.1368l7.61783,21.422,2.62767-.41552,21.58-3.42133,5.41173-.85662Z'
+                transform='translate(-181 -130.99812)'
+                fill='#e6e6e6'
+            />
+            <circle cx='685.01598' cy='119.90786' r='49.37021' fill='#a0616a' />
+            <path
+                d='M791.9668,223.65966c-7.51577-27.40164,53.10245-50.89908,69.02069-28.8325,2.21348-3.1372,10.39357-5.04277,14.14681-4.08041s6.90971,3.36827,9.98934,5.69723c4.23443,3.2528,8.62274,6.62107,11.22108,11.29814,2.61772,4.65781,3.02183,11.0095-.50033,15.05138-2.79088,3.21427-11.02877,5.293-15.22466,6.13989-2.92551.59666-5.38921,1.07784-7.50637,1.42428a14.35166,14.35166,0,0,1-2.50224-9.48889,11.64829,11.64829,0,0,0-5.08119,8.54579c-.0578.71214,23.65486,87.96006,23.65486,87.96006a51.72045,51.72045,0,0,1-8.95,4.02261,112.54462,112.54462,0,0,1-13.95427-30.41066,97.45318,97.45318,0,0,0,.26949,21.65318,119.14215,119.14215,0,0,0,2.0402,11.991c-28.178,6.0822-70.56044,4.08053-88.99936-21.78781q-.00492-.16051-.00929-.32129a4.40184,4.40184,0,0,0-8.50333-1.535c-.44736,1.14129-.85908,2.12916-1.2265,2.91492C764.48175,284.69268,771.43,262.4621,791.9668,223.65966Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path d='M810.43534,197.849' transform='translate(-181 -130.99812)' fill='#2f2e41' />
+            <path
+                d='M547.1209,394.255a5.09028,5.09028,0,0,1-3.68913-8.8227c.12835-.51014.22043-.8765.34879-1.38664q-.06889-.16659-.13859-.33286c-4.68631-11.17449-20.5687-11.09739-25.21514.09369-4.12406,9.93267-9.37471,19.88226-10.66754,30.38441a40.43982,40.43982,0,0,0,.70989,13.908A162.13182,162.13182,0,0,1,493.7207,360.7605a156.48974,156.48974,0,0,1,.97068-17.45764q.80413-7.128,2.23119-14.15449a164.01022,164.01022,0,0,1,32.52468-69.51166,43.64785,43.64785,0,0,0,18.153-18.83392,33.29478,33.29478,0,0,0,3.02794-9.09827c-.88372.11592-3.3322-13.34322-2.66567-14.169-1.23157-1.86885-3.43587-2.79784-4.78094-4.62148-6.68954-9.06961-15.906-7.486-20.7174,4.83888-10.27825,5.18745-10.37774,13.79049-4.07105,22.06463,4.01234,5.26406,4.56364,12.38693,8.08419,18.02253-.36227.46365-.739.91282-1.10113,1.37648a165.06908,165.06908,0,0,0-17.229,27.29491c1.36864-10.69089-.652-23.57059-4.09675-31.873-3.9213-9.45971-11.27118-17.4267-17.74364-25.6046-7.77447-9.82292-23.71674-5.536-25.08643,6.91609q-.01988.18086-.03883.36158,1.44205.8135,2.8232,1.727a6.90459,6.90459,0,0,1-2.78379,12.56506l-.14079.0217a68.7,68.7,0,0,0,1.81092,10.27183c-8.30336,32.11122,9.623,43.80679,35.21939,44.33218.565.28972,1.11566.57943,1.6807.8548a168.48694,168.48694,0,0,0-9.0693,42.68053,159.80436,159.80436,0,0,0,.11591,25.80246l-.04348-.30421A42.227,42.227,0,0,0,476.379,349.88019c-11.0933-9.11275-26.76607-12.46849-38.73378-19.79336-5.76092-3.526-13.14312,1.03057-12.1377,7.70959q.02407.1599.04895.31988a46.37792,46.37792,0,0,1,5.201,2.50624q1.44205.8137,2.82321,1.727a6.90468,6.90468,0,0,1-2.7838,12.56519l-.14091.02164c-.10138.01448-.18835.029-.28961.04359A68.75993,68.75993,0,0,0,443.014,374.78445c5.19187,28.03184,27.491,30.69143,51.34421,22.52828h.01454a168.43742,168.43742,0,0,0,11.3149,33.03181H546.108c.145-.4491.27532-.91275.40576-1.36192a45.9261,45.9261,0,0,1-11.18457-.66624c2.999-3.67994,5.99787-7.38886,8.99686-11.06867a2.51291,2.51291,0,0,0,.18835-.21734c1.52117-1.88328,3.05689-3.75226,4.57807-5.6356l.00081-.00229a67.27051,67.27051,0,0,0-1.97111-17.13665Z'
+                transform='translate(-181 -130.99812)'
+                fill='#f2f2f2'
+            />
+            <path
+                d='M358.51608,190.97871c-.9174,2.13788-29.20421,1.90879-30.49973,3.81247-9.98865,14.6971,26.0185,77.22955,16.02986,91.94573l-1.39152,1.65842a65.75661,65.75661,0,0,0,9.35956-5.28031c.2669-6.02364.5338-12.06649.81966-18.09013a49.72123,49.72123,0,0,1,7.16746,12.60017c2.32566-1.63933,4.23341-4.67832,14.20146-10.65578A62.515,62.515,0,0,1,388.0165,255.5982Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M379.05672,311.42954a23.4277,23.4277,0,0,0-35.49309,13.858L300.57777,437.29108,226.69,395.34309a22.87524,22.87524,0,1,0-14.27186,22.7523c15.56013,16.60716,83.01892,84.69637,110.94639,58.09227,23.47441-22.362,57.3508-109.02911,65.50546-138.957A23.43014,23.43014,0,0,0,379.05672,311.42954Z'
+                transform='translate(-181 -130.99812)'
+                fill='#ffb6b6'
+            />
+            <polygon
+                points='117.341 349.883 130.906 411.195 230.098 420.167 239.214 366.781 117.341 349.883'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M449.85848,638.108c-.04529.87892-.09966,1.74882-.16309,2.62779-.95143,12.70387-2.03879,27.356-3.21674,43.41254-1.85755,25.4803-3.93262,54.49451-6.02577,84.84978H232.397l40.35026-121.28154s16.29252-8.97031,7.99781-24.03914,8.02627-3.06065,8.1781-24.5811,14.777-44.41548,14.777-44.41548l2.40126-32.58436,1.36826.02721,108.27315,1.80318,1.97535.03623,3.32547,27.9722A123.668,123.668,0,0,1,449.85848,638.108Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M360.25006,281.71589l49.37257-2.54338,22.93226,36.23574s43.14762,7.21373,40.15679,58.05785-52.49777,95.46007-52.49777,95.46007l2.11627-7.25542-3.67543,10.502.47132,5.89689,5.014,6.245.51621,6.4585-4.44235,10.21851v0a5.83891,5.83891,0,0,1-5.53624,8.90024l-71.33958-7.14938c-1.62421-4.468-2.92741-9.81387-9.43789-.91444l-6.72031-.65106-27.93639-2.7066-1.6443-5.98994a23.20469,23.20469,0,0,1,3.77511-20.01465v0a30.60736,30.60736,0,0,1,1.77136-22.98081l.13362-.2677,2.96669-2.47279s-32.45673-48.76167,4.80261-85.48825l14.95415-25.42206,16.73241-20.41434Z'
+                transform='translate(-181 -130.99812)'
+                className='fill-primary'
+            />
+            <circle cx='191.538' cy='82.50891' r='46.81885' fill='#ffb6b6' />
+            <path
+                d='M408.09722,280.95294a22.5482,22.5482,0,0,1-7.54872-12.39051,48.44628,48.44628,0,0,0,.66725,17.68978c-2.53532,1.90623-4.99432,3.71718-7.32,5.35651a49.721,49.721,0,0,0-7.16745-12.60017c-.28587,6.02364-.55277,12.06649-.81967,18.09013a65.7548,65.7548,0,0,1-9.35955,5.2803l1.39151-1.65841c9.98864-14.71618-15.85987-41.80368-5.87123-56.50078,2.80222-4.11753-9.85519-36.65683-19.50073-38.12466-8.57805-1.29623-8.21586,1.18185-16.527,2.9928a41.5889,41.5889,0,0,0-3.98408-10.31279,34.07467,34.07467,0,0,1-.324,10.86555c-2.95464.11437-5.27927-1.76829-8.42552-2.24934-20.07269-3.069-18.033-24.78091-12.44776-31.6625,5.60433-6.8624,24.20918-8.92116,32.78723-11.15154.78151-.76242,2.55429-.28586,3.33591-1.01024,16.43935-15.045,38.42593-17.53657,58.21178-10.86829,21.10012,7.11122,34.25569,20.87458,41.00764,42.0924l1.93652,16.97969c1.73734,5.45956-.44817,30.86221-1.93652,36.39484C440.41735,269.67189,428.096,258.93593,408.09722,280.95294Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M476.36329,502.76855l-1.84316-143.55284a23.17268,23.17268,0,0,0-23.16985-22.80657h-.00007a23.17271,23.17271,0,0,0-23.14009,24.40154l6.486,149.05754.12313,90.73906a20.22151,20.22151,0,1,0,24.8266,7.56106Z'
+                transform='translate(-181 -130.99812)'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M718,688.49187s-2.85,51.27-4.16,80.51H615.7c-2.02-26.61-2.7-46.16-2.7-46.16l-5.37,46.16h-99.2l-6.15-85.49,29.72-73.67s161.06-11.17,176,3C722.93,627.00194,718,688.49187,718,688.49187Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <circle cx='450.21597' cy='72.59881' r='53.60692' fill='#ffb6b6' />
+            <path
+                d='M693.70022,186.178c-5.50615,2.51069-9.07382,20.69444-13.4016,16.4671s-4.03893-9.7286-9.87364-11.2887c-4.71184-1.26523-9.78253.18078-14.08037,2.46133-.7674.40241-1.51139.83995-2.2365,1.27509-.47865-.31981-.955-.62073-1.46639-.89836a17.03314,17.03314,0,0,0-12.5366-1.41014,16.7126,16.7126,0,0,1,8.2586,6.0872c-1.4384.96394-2.89543,1.93037-4.38568,2.78668-12.45014,7.18478-30.16623-8.43822-43.27124-14.3074,8.2086,15.404,4.16668,21.78351-2.96071,37.71453l-.24613.90714c.40491-3.99277-11.95042-6.96835-11.23395-1.34772.5087,3.99065,3.96222,12.86082,5.89652,16.38422-13.77842-15.93173-23.19417-35.42778-21.20733-55.5396.4325-4.22486.30214-8.68257,2.48127-12.3495,2.18143-3.648,6.7293-6.208,10.75272-4.81688a35.542,35.542,0,0,1,53.50674-32.52446c5.14478,3.01886,9.43984,7.28849,14.15516,10.97139,4.71543,3.68287,10.15119,6.89311,16.13038,7.19715,3.79317.18294,7.921-.74312,11.17973,1.22146,3.77491,2.2796,4.42131,7.05234,4.98779,11.64491a17.01391,17.01391,0,0,0-9.16691,3.815,16.648,16.648,0,0,1,10.02242,2.14968,1.03627,1.03627,0,0,1,.16413.09331C687.16438,180.46141,686.95993,182.144,693.70022,186.178Z'
+                transform='translate(-181 -130.99812)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M508.6679,635.777c45.65079-53.43516,30.23132-121.4003,12.11288-190.00758l3.44756-101.35372a31.00393,31.00393,0,0,1,14.02907-24.90184L584.861,289.0678,603,262.84191l50,5,27.72649,29.322,30.69062,23.42985a50.84312,50.84312,0,0,1,18.79205,51.39009L700.83948,504.80436s12.28383,68.85522,4.78852,76.45969-26.80842.00387-11.49291,8.54042c18.37976,10.24451,35.09164,8.35414,25.79051,17.65527s2.47025,22.23225,2.47025,22.23225Z'
+                transform='translate(-181 -130.99812)'
+                className='fill-primary'
+            />
+            <path
+                d='M528.18594,713.43646l7.61052-54.22065-30.36735-8.66909-6.00663,57.61172a18.32876,18.32876,0,1,0,28.76346,5.278Z'
+                transform='translate(-181 -130.99812)'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M586,298.84191s-86.4384,28.13284-88.4995,48.71733S496.12293,380.74489,487,463.11505l.48061,221.88319s57.95172,12.40752,52.0085-.8585-3.86948-19.50742-3.86948-19.50742,12.42491-4.90311,9.31468-20.47947,12.48716-92.76777,12.48716-92.76777L567.266,412.94623Z'
+                transform='translate(-181 -130.99812)'
+                className='fill-primary'
+            />
+            <path
+                d='M692.58051,710.565l-4.05011-54.60215,30.86907-6.66559,2.228,57.88114a18.32876,18.32876,0,1,1-29.047,3.3866Z'
+                transform='translate(-181 -130.99812)'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M661.99007,293.07805s84.41465,33.72272,85.12583,54.39792-.79456,33.20472,2.92474,115.99505l-14.983,221.37726s-58.63881,8.593-51.84116-4.2562,5.1363-19.21278,5.1363-19.21278-12.07784-5.70478-7.95612-21.04453S674,546.94918,674,546.94918l-.77441-138.78628Z'
+                transform='translate(-181 -130.99812)'
+                className='fill-primary'
+            />
+        </svg>
+    )
+}
+
+export const PercentageIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 700.74679 597.77667'
+        >
+            <path
+                d='M868.01492,746.55858l.0152-.697a80.52864,80.52864,0,0,1,30.11215-60.64309,78.13219,78.13219,0,0,1,16.78623-10.11535l.572-.25474.09758.61813c1.16153,7.3802,3.0009,14.91288,3.85784,18.25612l5.74416-22.41411.63822-.18437a15.27246,15.27246,0,0,1,16.303,5.20879,15.53409,15.53409,0,0,1,1.61173,17.33177c-1.33891,2.49455-2.655,5.09733-3.92819,7.61425-4.37386,8.6493-8.89655,17.59309-15.48708,24.54163a48.6753,48.6753,0,0,1-39.25924,14.64528Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#f2f2f2'
+            />
+            <polygon
+                points='546.768 585.081 536.773 585.08 532.017 546.524 546.77 546.525 546.768 585.081'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M796.747,746.62045l-30.73878-.00116v-.38874a11.96505,11.96505,0,0,1,11.96439-11.9642h.00074l5.61483-4.2597,10.476,4.26035,2.68329.00009Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#2f2e41'
+            />
+            <polygon
+                points='650.238 556.92 642.341 563.048 614.944 535.505 626.599 526.46 650.238 556.92'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M906.53628,716.053l-24.28357,18.84624-.23835-.3071a11.965,11.965,0,0,1,2.11584-16.78725l.00058-.00045,1.82379-6.80774,10.888-3.05769,2.11979-1.64516Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M856.62661,521.48485s-.68218,10.43874-1.54259,25.36144c-.14111,2.44744.69106,5.10617.54259,7.76856-.09851,1.76641-1.17624,3.48366-1.27636,5.32545-.10567,1.94378,1.38234,3.66437,1.27636,5.67455-.10355,1.964-1.80109,4.21635-1.90349,6.22674-1.102,21.63419-5.09651,47.53327-1.09651,54.77326,1.97134,3.56812,32,53,32,53s-14,21-18,14c-5.28313-9.24547-46-52.5-46-60.76,0-8.25-9.41993-70.08-9.41993-70.08l-6.56834,36.59036-3.01173,7.24966.94634,4.25606-1.94634,5.74394-.10516,5.68434-1.89484,10.55565s5.69995,82.24,2.49,84.53c-3.22,2.3-18.81006,6.43-20.64991,2.3-1.83007-4.13-8.84008-85.83-8.84008-85.83s-1.91993-37.43994-4-72.24c-.11988-2.02-1.54-4.46-1.65-6.45-.13-2.31.77-5.30994.65-7.55-.13-2.56994-1.27-4.38-1.38-6.84-.85-19.01-.98-33.99.53992-36.27,4.59009-6.88,8.72009-12.38,8.72009-12.38s65.6-16.06,82.12,4.58Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M797.62661,324.61485l15.81737-.88448,11.92026,6.50684,33.14757,187.71555s-33.8852,15.66209-54.8852,9.66209-43.1148-5.66209-43.1148-5.66209l16.1148-179.33791,13-12Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M824.01455,328.59465l24.483,9.99826a16.66267,16.66267,0,0,1,15.58371,13.66737l-1.6245,48.35257-8.83011,23.002,6.29683,52.40017-3.29683,6.59983,2.84875,6.73651c-4.18164,6.78919-3.33044,15.68512-.84875,25.26349h-3l-6-1S836.417,518.201,833.7107,501.72978c-7.08409-43.11493-11.08409-24.11493-11.08409-24.11493l-3.4872-133.77245Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M790.62661,328.61485,767.7557,352.59291A16.66267,16.66267,0,0,0,752.172,366.26028l9.45462,74.35457-4.43717,28.53025,2.43717,6.46975-1.765,7.31259,2.765,6.68741-2,9,3,13,15-4-.1148-10.66209,4-145Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M714.354,451.87174l-31.15473,22.607a6.803,6.803,0,1,0,6.422,9.86626L731.98568,461.729Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M757.62661,356.61485c-5.92121,2.7311-20.54883,72.65156-24.8419,74.134s-2.65392,1.25967-3.2562,2.49728-.74795,6.7793-4.86925,7.42965-25.896,20.50182-25.896,20.50182l5.51957,18.3993s41.90147-23.93873,53.45495-37.195,17.774-81.42911,17.774-81.42911l1.18175-6.80129S763.54781,353.88375,757.62661,356.61485Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M851.58866,442.63989l-27.76457,26.66127a6.803,6.803,0,1,0,7.71282,8.89365l38.86738-28.20563Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#ffb6b6'
+            />
+            <path
+                d='M855.62661,362.61485c-5.49132,3.5164,15.37534,54.52026,11.32578,56.57677s-2.45637,1.61131-2.88345,2.91974.18767,6.81786-3.8057,8.02658-22.84379,23.85561-22.84379,23.85561l7.98773,17.46986s38.22762-29.45241,47.85648-44.16628,3.363-19.68228,3.363-19.68228c-11.94605-25.37237-23.11939-66.75079-40-64C856.62661,343.61485,861.11792,359.09846,855.62661,362.61485Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <circle cx='551.39233' cy='142.73745' r='26.80653' fill='#ffb6b6' />
+            <path
+                d='M828.16428,311.12189s-5.904,1.61882-9.33976,2.52007c.17761-2.55664,1.28486-4.95,1.84714-7.45109s.44025-5.44257-1.41851-7.20853-5.73-1.07792-6.02951,1.46313q-2.35871-5.04375-4.72718-10.1059c-4.50223,9.26712-13.70672,1.36989-23.74286,3.71145-5.11992,1.20608-11.53673.688-14.24163-3.82337a16.71144,16.71144,0,0,0,3.782-4.53721,15.26826,15.26826,0,0,1-4.8081.60916,10.07877,10.07877,0,0,1,4.45653-8.1348c1.44345-.94375-2.12937-4.30177-.86958-5.48383,2.58379-2.41255,4.31775-.77428,5.8442-3.95434a11.74567,11.74567,0,0,1,10.61238-6.48825c4.38693.10274,12.24765.0604,13.94968,4.102,5.42448-3.551,12.9755-2.5477,18.27776,1.17736,4.40717,3.08424,9.42281,12.40684,10.52973,17.6696C833.39351,290.45,830.95413,306.53528,828.16428,311.12189Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M496.12661,252.11485H324.47817A12.21074,12.21074,0,0,1,312.154,240.93419a12.00675,12.00675,0,0,1,11.97266-12.81934h172a12,12,0,0,1,0,24Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#6c63ff'
+            />
+            <path
+                d='M594.86293,271.61485H312.62661a1,1,0,0,1,0-2H594.86293a1,1,0,0,1,0,2Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M583.12661,252.61485h-259a12.5,12.5,0,0,1,0-25h259a12.5,12.5,0,0,1,0,25Zm-259-23a10.5,10.5,0,0,0,0,21h259a10.5,10.5,0,0,0,0-21Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M419.12661,346.11485H324.47817A12.21074,12.21074,0,0,1,312.154,334.93419a12.00675,12.00675,0,0,1,11.97266-12.81934h95a12,12,0,0,1,0,24Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#6c63ff'
+            />
+            <path
+                d='M594.86293,365.61485H312.62661a1,1,0,0,1,0-2H594.86293a1,1,0,0,1,0,2Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M583.12661,346.61485h-259a12.5,12.5,0,0,1,0-25h259a12.5,12.5,0,0,1,0,25Zm-259-23a10.5,10.5,0,0,0,0,21h259a10.5,10.5,0,0,0,0-21Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M546.12661,440.11485H324.47817A12.21074,12.21074,0,0,1,312.154,428.93419a12.00675,12.00675,0,0,1,11.97266-12.81934h222a12,12,0,0,1,0,24Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#6c63ff'
+            />
+            <path
+                d='M594.86293,459.61485H312.62661a1,1,0,0,1,0-2H594.86293a1,1,0,0,1,0,2Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M583.12661,440.61485h-259a12.5,12.5,0,0,1,0-25h259a12.5,12.5,0,0,1,0,25Zm-259-23a10.5,10.5,0,0,0,0,21h259a10.5,10.5,0,0,0,0-21Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#3f3d56'
+            />
+            <polygon
+                points='230.214 88.806 228.214 88.806 228.214 11.763 372.871 11.763 372.871 13.763 230.214 13.763 230.214 88.806'
+                fill='#ccc'
+            />
+            <circle cx='379.0541' cy='11.9459' r='11.9459' fill='#ccc' />
+            <polygon
+                points='160.786 182.806 162.786 182.806 162.786 138.763 18.129 138.763 18.129 140.763 160.786 140.763 160.786 182.806'
+                fill='#ccc'
+            />
+            <circle cx='11.9459' cy='138.9459' r='11.9459' fill='#ccc' />
+            <polygon
+                points='296.214 280.201 294.214 280.201 294.214 335.244 416.871 335.244 416.871 333.244 296.214 333.244 296.214 280.201'
+                fill='#ccc'
+            />
+            <circle cx='423.0541' cy='335.06048' r='11.9459' fill='#ccc' />
+            <path
+                d='M949.18271,748.581l-271.75.30734a1.19069,1.19069,0,1,1,0-2.38137l271.75-.30733a1.19068,1.19068,0,0,1,0,2.38136Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#cacaca'
+            />
+            <path
+                d='M839.70638,328.61772a4.459,4.459,0,0,0-6.04285-1.43262c1.47-2.04278.36243-5.53375-1.94165-6.94482-2.61035-1.59876-5.908-1.33392-8.95251-1.016L797.81246,321.831c-2.8775.30053-5.83478.62445-8.3963,1.96954s-4.6687,3.97016-4.52069,6.85962a14.35827,14.35827,0,0,0-6.12183-.83026c-2.05145.3233-4.07452,1.82269-4.2934,3.88794-.35125,3.31409,3.5857,5.228,6.79212,6.13653q8.93041,2.53033,17.86084,5.0608l-1.8357.39233a76.07616,76.07616,0,0,0,34.45551-4.98456c2.50818-.98889,5.04779-2.17328,6.804-4.21893S841.07973,330.93785,839.70638,328.61772Z'
+                transform='translate(-249.62661 -151.11166)'
+                fill='#e6e6e6'
+            />
+        </svg>
+    )
+}
+
+export const PeopleExpenseIllustration = ({ width, height }) => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            data-name='Layer 1'
+            width={width}
+            height={height}
+            viewBox='0 0 312.32533 689.35327'
+        >
+            <path
+                d='M756.73957,798.06625h-121.03l-1.18-5.1-6.47-28.06-6.46,28.06-1.17,5.1h-124.5c-.27-1.72-.53-3.42-.79-5.1-3.63-23.17-7.22-42.46-10.26-47.98-8.82-15.98-.6-5.97,8.81-15.39,9.41-9.41,0,0-5.51-12.08-5.52-12.08,24.37-103.88,24.37-103.88l203.9,3.01s37.66,65.89,38.05,74.98c.22,5.17,1.27,57.32,2.13,101.34C756.66956,794.67623,756.69953,796.38619,756.73957,798.06625Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M637.8194,677.4775s86.78375,39.925,100.456,9.29968,6.90448-71.56005-1.56573-76.91809-17.77167-30.04309-5.38013-47.1029-7.09991-45.88031-12.55743-48.82343c-4.15564-2.241-15.46167-128.52191-20.61933-188.64374a48.38235,48.38235,0,0,0-32.23834-41.53308h0L645.39905,250.6226l-52.62915-.47348L566.478,286.60073l-37.9447,14.19278L500.05069,418.42129s-.275,51.957,5.71109,107.43243c1.86252,17.26062,4.33118,34.86188,7.60285,51.34473,7.706,38.823-25.86032,70.79376-6.67206,78.14258C560.21076,675.83761,637.8194,677.4775,637.8194,677.4775Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M528.88448,601.84869l166.95367-6a8.63685,8.63685,0,0,1,8.63686,8.63684v83.71454a8.63683,8.63683,0,0,1-8.63686,8.63684H538.88448a8.63683,8.63683,0,0,1-8.63685-8.63684c-9.255-23.5864-16.99663-46.545-10-77.71454A8.63685,8.63685,0,0,1,528.88448,601.84869Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#3f3d56'
+            />
+            <rect x='103.11943' y='503.42215' width='137.38036' height='2' fill='#2f2e41' />
+            <rect x='103.12012' y='532.30344' width='137.37998' height='4.28271' fill='#2f2e41' />
+            <rect x='103.11966' y='442.02643' width='126.00779' height='61.48572' fill='#6c63ff' />
+            <rect
+                x='103.11966'
+                y='442.02643'
+                width='126.00779'
+                height='61.48572'
+                opacity='0.2'
+                style={{ isolation: 'isolate' }}
+            />
+            <rect x='142.21244' y='442.02643' width='86.91502' height='61.48572' fill='#6c63ff' />
+            <path
+                d='M659.959,561.0071H642.7959a.59381.59381,0,0,1,0-1.18762H659.959a.59381.59381,0,0,1,0,1.18762Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#fff'
+            />
+            <path
+                d='M575.32117,603.89529H558.15808a.59382.59382,0,0,1,0-1.18763h17.16309a.59381.59381,0,0,1,0,1.18763Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#fff'
+            />
+            <path
+                d='M608.34131,561.0071H555.53473a.61747.61747,0,0,1-.619-.49353.55931.55931,0,0,1,.58558-.6651h53.12207a.6176.6176,0,0,1-.28205,1.15863Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#fff'
+            />
+            <path
+                d='M692.98249,695.92709,586.32056,719.582a12.85728,12.85728,0,0,1-12.85724-12.85724c-10.47538-26.79837-10.65282-51.91853,0-75.27374a12.85728,12.85728,0,0,1,12.85724-12.85724l106.66193-23.6549a12.85724,12.85724,0,0,1,12.85723,12.85721c9.85742,28.11328,18.00751,55.70313,0,75.27374A12.85729,12.85729,0,0,1,692.98249,695.92709Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M596.66327,634.57033,598.02828,643l69.6-1.81462v0c.23338-4.41383-31.37986-7.71824-64.90238-6.784Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M596.66327,655.57033,598.02828,664l69.6-1.81462v0c.23338-4.41383-31.37986-7.71824-64.90238-6.784Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M596.66327,680.57033,598.02828,689l69.6-1.81462v0c.23338-4.41383-31.37986-7.71824-64.90238-6.784Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#3f3d56'
+            />
+            <path
+                d='M719.04871,508.25671l-1.96136,18.81839-.45978,4.35223-.70484,6.98785L700.78229,683.65987a19.77416,19.77416,0,0,1-8.67364,22.5575,18.96429,18.96429,0,0,1-25.99012-6.5589c-.092-.1532-.15338-.3064-.24524-.45959a19.88476,19.88476,0,0,1,6.8653-26.84839,20.14915,20.14915,0,0,1,2.85034-1.34857l-.12256-4.07636-.36779-12.78052-.27594-9.65442L674.2403,625.274l-.2146-7.44782-.7049-24.15106-.67419-23.29321-.82758-28.22742-.64373-21.6994,7.81543-1.99225Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#a0616a'
+            />
+            <path
+                d='M734.18921,483.70892c-.70483,3.59253-1.34857,5.914-1.80816,6.66-3.9538,6.19043-6.375-2.45941-4.19886,14.12173,1.747,13.45835-8.735,11.77243-4.96508,14.64664,6.22161,4.75318-2.36,6.85346-6.12976,9.42352-1.68586,1.13312,2.98174.864,4.91265,2.43919.91858.71967-3.44384,2.79952-2.36889,3.25362,4.107,1.93451,1.747,3.39924-3.70838,4.53223-9.04144,1.8515-26.542,2.76355-37.11584,3.1228-3.00354.13813-5.45538.19349-6.98779.24878-1.134.02753-1.77771.02753-1.77771.02753s-2.66645-1.71338-3.76972-6.93628c-.79693-3.64789-.79693-9.03674,1.40973-16.7746,1.34857-4.75332,3.64722-8.1524,4.99579-10.88827a7.18976,7.18976,0,0,0,1.04211-4.25592v-.02768c-.2146-1.40924-1.226-2.70819-3.52472-4.08994-4.13751-2.48706-1.83886-7.71017,1.37922-12.6293,3.06488-4.615,6.92663-9.00912,6.92663-10.61182,0-3.39923-25.49988-75.63751-25.49988-75.63751L662.379,291.25741a36.07551,36.07551,0,0,1,24.45764-6.10727c5.91541.77368,12.50476,3.26092,19.2782,8.84311,5.91522,4.89148,11.98358,12.18722,17.807,22.74386C748.6861,361.589,739.46088,457.23434,734.18921,483.70892Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#e6e6e6'
+            />
+            <circle cx='166.11047' cy='74.33239' r='54.73784' fill='#a0616a' />
+            <path
+                d='M667.48944,210.89437s-15.18865,12.40323-23.51709,16.02173c2.30585-8.00812,4.70838-21.75676,7.00782-29.74213,1.19574-4.20175,2.417-8.49492,2.22247-12.83758-.211-4.37193-2.08282-8.88634-5.78925-11.15169a10.0547,10.0547,0,0,0-14.60894,12.26892c-3.62805-.24716-5.677-4.392-6.12116-8.01553-.44421-3.62372-.10785-7.57187-2.11053-10.644-1.93042-2.978-5.74182-4.336-9.31177-4.17271-3.54083.14691-6.90338,1.60035-10.07574,3.25485-9.07154,4.693-17.30988,11.171-27.07019,14.26726a32.90308,32.90308,0,0,1-6.00843,1.30764,25.779,25.779,0,0,0-5.54211-16.28318,44.45273,44.45273,0,0,1-4.571,15.78131,18.86245,18.86245,0,0,1-11.69421-7.47393c-3.055-4.42526-3.80634-10.13075-3.08752-15.45118a36.53516,36.53516,0,0,1,1.77649-7.16989,22.52994,22.52994,0,0,0,4.1748-1.449,24.51589,24.51589,0,0,0,7.86627-5.49725,40.18032,40.18032,0,0,1-9.015-.63959c.07385-.17648.14135-.3302.21527-.50682,1.70111-3.7158,3.58392-7.55336,6.86865-9.96522,3.28467-2.41173,8.37238-2.89155,11.2254.04786a52.08774,52.08774,0,0,1,71.90845-15.98247q.38984.248.77515.50294a24.75864,24.75864,0,0,0-4.80207,8.29846,42.21844,42.21844,0,0,1,10.21961-3.683c2.5,2.79165,4.08038,6.14056,3.40253,9.72244,4.14514,2.01787,12.22863,5.38,8.40777,7.59306a53.38331,53.38331,0,0,1,7.25415,71.59871Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#2f2e41'
+            />
+            <path
+                d='M599.628,586.88137a18.96931,18.96931,0,0,1-20.1289,17.70917,4.20385,4.20385,0,0,0-.51392-.06433,19.90236,19.90236,0,0,1-18.2445-20.8783,21.3391,21.3391,0,0,1,.51391-3.10486l-3.42615-2.20569-10.74969-6.95947-8.0943-5.22491-16.1673-10.42847-6.25287-4.02576-.83514-.53534-19.465-12.59124-15.632-10.08576-1.32764-.85663-2.61255-1.69165-23.72638-15.31082-6.38123-4.11139c-.02145-.02142,0-.02142,0-.02142-.21408-.40686,13.21219-7.06647,12.99811-7.47336-3.38339-6.70248-7.66613-12.33429-2.14135-36.40323l-5.63178-10.12869,14.39,12.29157,3.31909,2.84793,5.35342,4.53967,66.211,56.48929L585.966,566.94534A19.79572,19.79572,0,0,1,599.628,586.88137Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#a0616a'
+            />
+            <path
+                d='M587.50776,379.36157,513.26661,493.51785,543.09583,526.859l-2.01288,1.79871-25.35382,22.5058-7.17352,6.35992-13.27655-2.69818s-2.63385-4.00433-16.48856-6.06012c-13.85458-2.05579-5.78159-6.76666.34268-15.97461,1.45612-2.20557-4.19711,1.26349-4.56112.49261-.17133-.47119-3.59748-2.0343-4.28275-2.14135-2.82663-.53534-.0643-2.14136-4.13281-2.97657-6.53116-1.32763-6.91657,3.29767-19.57215-21.6492-.02145-.02142,0-.02142,0-.02142-.21408-.40686-.40686-.7923-.62094-1.19916-3.38339-6.70248-.9422-25.33239,4.58255-49.40133,8.43686-36.70307,24.02607-86.0401,37.00275-124.82026a55.95274,55.95274,0,1,1,99.959,48.28782Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#e6e6e6'
+            />
+            <path
+                d='M617.01923,672.38423a1.52141,1.52141,0,0,0,1.423-2.47885l-.05172-.42421.05352-.09354a4.08563,4.08563,0,0,1,7.47411.968c.85264,3.09933,2.0387,6.24569,2.0305,9.4083a12.08649,12.08649,0,0,1-.72914,4.098,48.45968,48.45968,0,0,0,6.88456-19.41934,46.7746,46.7746,0,0,0,.36307-5.21325q.02739-2.14377-.13386-4.28075a49.02015,49.02015,0,0,0-7.0533-21.82651,13.04571,13.04571,0,0,1-4.681-6.2621,9.94931,9.94931,0,0,1-.55868-2.811c.25774.06731,1.48572-3.83269,1.31887-4.10245.43492-.50826,1.12324-.70158,1.59012-1.19226,2.322-2.44017,4.99608-1.6269,5.96335,2.20746,2.85458,1.92158,2.5633,4.47657.38455,6.69506-1.38614,1.41148-1.81523,3.5032-3.0694,5.04315.09015.151.18509.29828.27522.44923a49.33912,49.33912,0,0,1,4.09152,8.73679,20.51008,20.51008,0,0,1,2.40339-9.29924c1.51559-2.659,3.9923-4.74763,6.21663-6.93143a4.20968,4.20968,0,0,1,7.1815,2.98638l-.002.10867q-.4579.18742-.90163.40685a2.06369,2.06369,0,0,0,.357,3.83l.04094.01167a20.53192,20.53192,0,0,1-.92,2.97861c1.265,9.83223-4.48714,12.63213-12.09736,11.83349-.17836.06485-.35249.1302-.53029.1908a50.3585,50.3585,0,0,1,1.09807,12.9951,47.75685,47.75685,0,0,1-.99651,7.64744l.02428-.0886a12.62152,12.62152,0,0,1,5.184-6.69321c3.62954-2.28879,8.40241-2.69951,12.22461-4.42545a2.36864,2.36864,0,0,1,3.312,2.73888l-.02644.093a13.85692,13.85692,0,0,0-1.63579.5493q-.458.18755-.90164.40685a2.06369,2.06369,0,0,0,.357,3.83l.041.01165.08429.02374a20.55119,20.55119,0,0,1-4.48912,5.40142c-2.58486,8.11928-9.29686,8.0765-16.06621,4.76621l-.00437-.00052a50.343,50.343,0,0,1-4.58715,9.3737l-11.98671-1.50717c-.02628-.13861-.04764-.281-.06955-.419a13.7266,13.7266,0,0,0,3.34166.2195c-.75216-1.20311-1.50317-2.4148-2.25532-3.6179a.742.742,0,0,1-.04777-.07147c-.3809-.61518-.76662-1.22672-1.14753-1.84195l-.00015-.00072a20.10637,20.10637,0,0,1,1.22349-5.00838Z'
+                transform='translate(-444.41424 -108.71297)'
+                fill='#f2f2f2'
+            />
+        </svg>
+    )
+}
