--- src/components/svg/ArrowSVG.jsx
@@ -1,16 +1,0 @@
-const ArrowSVG = () => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            fill='none'
-            viewBox='0 0 24 24'
-            strokeWidth={1.5}
-            stroke='currentColor'
-            className='w-6 h-6 bg-gray-100 rounded p-1 dark:bg-slate-600'
-        >
-            <path strokeLinecap='round' strokeLinejoin='round' d='M17.25 8.25L21 12m0 0l-3.75 3.75M21 12H3' />
-        </svg>
-    )
-}
-
-export default ArrowSVG
