--- src/components/svg/ChevronDownSVG.jsx
@@ -1,16 +1,0 @@
-const ChevronDownSVG = () => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            fill='none'
-            viewBox='0 0 24 24'
-            strokeWidth={1.5}
-            stroke='currentColor'
-            className='w-4 h-4'
-        >
-            <path strokeLinecap='round' strokeLinejoin='round' d='M19.5 8.25l-7.5 7.5-7.5-7.5' />
-        </svg>
-    )
-}
-
-export default ChevronDownSVG
