--- src/components/svg/ChevronRight.jsx
@@ -1,16 +1,0 @@
-const ChevronRight = () => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            fill='none'
-            viewBox='0 0 24 24'
-            strokeWidth={1.5}
-            stroke='currentColor'
-            className='w-5 h-5'
-        >
-            <path strokeLinecap='round' strokeLinejoin='round' d='M8.25 4.5l7.5 7.5-7.5 7.5' />
-        </svg>
-    )
-}
-
-export default ChevronRight
