--- src/components/svg/ChevronUpSVG.jsx
@@ -1,16 +1,0 @@
-const ChevronUpSVG = () => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            fill='none'
-            viewBox='0 0 24 24'
-            strokeWidth={1.5}
-            stroke='currentColor'
-            className='w-4 h-4'
-        >
-            <path strokeLinecap='round' strokeLinejoin='round' d='M4.5 15.75l7.5-7.5 7.5 7.5' />
-        </svg>
-    )
-}
-
-export default ChevronUpSVG
