--- src/components/svg/CustomizationSVG.jsx
@@ -1,171 +1,0 @@
-const CustomizationSVG = ({ width, height }) => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            data-name='Layer 1'
-            width={width}
-            height={height}
-            viewBox='0 0 637.60339 599.7033'
-        >
-            <path
-                d='M867.33431,411.24994a9.50264,9.50264,0,0,0-9.50856-9.23358,9.11579,9.11579,0,0,0-1.49961.14972l-16.62067-28.6843-17.52229,3.31264,27.01655,36.83234a9.47629,9.47629,0,0,0,9.177,7.04467,9.06878,9.06878,0,0,0,8.96388-9.17249Q867.33916,411.37441,867.33431,411.24994Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#a0616a'
-            />
-            <polygon
-                points='506.295 581.028 492.409 581.026 485.802 527.463 506.298 527.464 506.295 581.028'
-                fill='#a0616a'
-            />
-            <path
-                d='M811.38256,744.135l-42.70332-.00161v-.54a16.62223,16.62223,0,0,1,16.62132-16.62106h.001l7.8003-5.91772,14.55366,5.91862,3.72771.00013Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#2f2e41'
-            />
-            <polygon
-                points='520.928 501.241 511.57 511.5 467.542 480.287 481.354 465.144 520.928 501.241'
-                fill='#a0616a'
-            />
-            <path
-                d='M836.55922,659.26189l-28.77693,31.55094-.399-.36392a16.62224,16.62224,0,0,1-1.08041-23.48107l.00069-.00076.88386-9.751,14.18008-6.76492,2.512-2.75419Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M721.61412,373.89681l-26.01594,47.37008,2,34-11,110,103,86,13-19-58-74,.69542-5.24728a7.63771,7.63771,0,0,1,1.32153-9.97152c4.983-4.7812.74817-5.64527.74817-5.64527l8.23488-62.13593,59-11.79528-24.748-93.38123Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M756.59818,441.26689l31,261,27,2.71654v-91.218s-5-4.96309,0-7.23081,0-10.77166,0-10.77166V463.47161l-25-39.20472Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M701.59818,223.26689l-24,9-14,32s-32,16-20,38S703.8856,372.5543,703.8856,372.5543c-5.90778,11.52372-3.88685,20.38877,5.71258,26.71259,12.53272,20.10248,47.7218,10.25409,87-5,7.26868-17.72495-3.98376-45.17648-15.045-72.52756l3.045-36.47244s-34.49606-50.19678-40-50-27-2-27-2Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#3f3d56'
-            />
-            <path
-                d='M735.59818,240.36925l0,0a19.239,19.239,0,0,1,29.499,3.69812l52.31369,82.87314c5.02063,2.37748,7.544,7.74,9.04346,14.32636,3.72864-1.26514,6.5528,1.19434,7.41266,11.7428l19.73114,31.25723-18,20-24.89429-32.99936c-5.78766-1.53006-3.09631-9.82547-7.101-9.413s-5.94073-7.87488-5.94073-7.87488c-5.05134-.057-6.5351-2.427-5.064-6.71271l-51-72.88189Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#3f3d56'
-            />
-            <path
-                d='M513.14779,238.50371a9.50264,9.50264,0,0,0,11.72293,6.18424,9.11591,9.11591,0,0,0,1.39685-.56572l24.02179,22.8471,15.8817-8.11024-36.29065-27.73995a9.47629,9.47629,0,0,0-10.7887-4.17713,9.06877,9.06877,0,0,0-6.02,11.32453Q513.10811,238.38564,513.14779,238.50371Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#a0616a'
-            />
-            <path
-                d='M661.59818,273.26689l-33,10-23.25984-3.69678-5.21727-3.72738c-2.24658-6.07431-6.77381-8.9307-12.80652-9.66294,0,0-4.87986,1.67223-4.79812-3.62034s-31.91825-24.29256-31.91825-24.29256l-14,25,18,22c6.38947.19733,10.32663-.64578,7,7,1.35757,7.363.41861,5.6539,5.90286,5.78845l28.09714,28.21155,59,9,37.47244-13.199Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#3f3d56'
-            />
-            <circle cx='365.2104' cy='42.75712' r='30.71644' fill='#a0616a' />
-            <path
-                d='M689.51592,190.78444q-.03947-.51366-.10535-1.02726A2.30264,2.30264,0,0,0,689.51592,190.78444Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M627.06221,170.29381c-.53923-3.68527,1.51333-7.86251,5.10168-8.88461,2.23582-.63832,4.62236-.02657,6.94171.01382,6.50608.11,18.07-10.72993,24.35075-12.39064,6.29227-1.656,14.43492,1.80493,14.94424,8.29041,16.08953-4.79045,28.37513,6.04782,29.12537,23.14065.58674,13.3678-10.78155,31.58743-22.77064,37.53377-1.46784-7.7789,3.96911-19.43892,2.50465-27.2097a15.75444,15.75444,0,0,0-2.02937-6.06981c-1.20062-1.7823-3.40319-3.07018-5.48577-2.54344a3.54183,3.54183,0,0,0-2.46459,4.10189,14.4489,14.4489,0,0,0-24.16753-1.12l1.69754-6.988a33.62333,33.62333,0,0,1-16.05187,9.82088,4.46182,4.46182,0,0,1-.78031-1.482,3.36418,3.36418,0,0,1,.66252-3.245,7.79112,7.79112,0,0,0-2.98254,5.3456,15.316,15.316,0,0,1-10.738-2.11122C621.83619,184.39655,627.60149,173.97905,627.06221,170.29381Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M907.70771,747.29932H874.52119l-.14258-.25879c-.42432-.76953-.834-1.585-1.2168-2.42285-3.41846-7.31836-4.86328-15.68848-6.13818-23.07325l-.96-5.5664a3.43688,3.43688,0,0,1,5.41015-3.36231q7.56519,5.5049,15.13623,10.999c1.91114,1.39062,4.09375,3,6.18409,4.73925.20166-.97949.41259-1.96191.62353-2.93066a3.43916,3.43916,0,0,1,6.28076-1.08594l3.88281,6.23828c2.832,4.55567,5.33155,9.04493,4.82227,13.88672a.75716.75716,0,0,1-.01318.17578,10.94679,10.94679,0,0,1-.56348,2.33106Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#f0f0f0'
-            />
-            <path
-                d='M941.0171,748.01709l-315.3575.30731a1.19069,1.19069,0,0,1,0-2.38135l315.3575-.30731a1.19069,1.19069,0,0,1,0,2.38135Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#cacaca'
-            />
-            <path
-                d='M580.59818,621.26689h-253a23.02588,23.02588,0,0,1-23-23v-340a23.02619,23.02619,0,0,1,23-23h253a23.02619,23.02619,0,0,1,23,23v340A23.02588,23.02588,0,0,1,580.59818,621.26689Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#fff'
-            />
-            <path
-                d='M580.59818,621.26689h-253a23.02588,23.02588,0,0,1-23-23v-340a23.02619,23.02619,0,0,1,23-23h253a23.02619,23.02619,0,0,1,23,23v340A23.02588,23.02588,0,0,1,580.59818,621.26689Zm-253-384a21.02359,21.02359,0,0,0-21,21v340a21.0239,21.0239,0,0,0,21,21h253a21.0239,21.0239,0,0,0,21-21v-340a21.02359,21.02359,0,0,0-21-21Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <circle cx='60.5' cy='189.89943' r='9' fill='#ccc' />
-            <path
-                d='M550.09818,339.52052H395.854a1,1,0,0,1,0-2H550.09818a1,1,0,0,1,0,2Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#ccc'
-            />
-            <circle cx='61' cy='356.89943' r='9' fill='#ccc' />
-            <circle cx='60.56127' cy='261.49861' r='9' fill='#ccc' />
-            <path
-                d='M548.09818,364.1197h-40a4,4,0,0,1,0-8h40a4,4,0,0,1,0,8Z'
-                transform='translate(-304.59818 -148.62109)'
-                className='fill-primary'
-            />
-            <path
-                d='M548.09818,463.1197h-40a4,4,0,0,1,0-8h40a4,4,0,0,1,0,8Z'
-                transform='translate(-304.59818 -148.62109)'
-                className='fill-primary'
-            />
-            <path
-                d='M519.295,289.52052H404.4014a10.55322,10.55322,0,0,1,0-21.10644H519.295a10.55322,10.55322,0,1,1,0,21.10644Z'
-                transform='translate(-304.59818 -148.62109)'
-                className='fill-primary'
-            />
-            <path
-                d='M548.09818,588.1197h-40a4,4,0,1,1,0-8h40a4,4,0,0,1,0,8Z'
-                transform='translate(-304.59818 -148.62109)'
-                className='fill-primary'
-            />
-            <path
-                d='M540.30814,563.52052H405.73832a11.11424,11.11424,0,0,1-2.51953-.31933,11.24845,11.24845,0,0,1-8.62061-10.97071V506.49025a11.29752,11.29752,0,0,1,11.29-11.27979H540.30814a11.144,11.144,0,0,1,5.20264,1.27344,10.81252,10.81252,0,0,1,1.8042,1.17725,11.16537,11.16537,0,0,1,4.2832,8.8291v45.74023A11.30252,11.30252,0,0,1,540.30814,563.52052ZM405.88822,497.21046a9.29555,9.29555,0,0,0-9.29,9.27979v45.74023a9.25724,9.25724,0,0,0,7.09473,9.02832,9.13326,9.13326,0,0,0,2.05761.26172H540.30814a9.30086,9.30086,0,0,0,9.29-9.29V506.49025a9.18459,9.18459,0,0,0-3.52783-7.26367,8.86057,8.86057,0,0,0-1.481-.96778,9.16856,9.16856,0,0,0-4.28125-1.04834Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M531.28818,562.52052q9.165-22.275,19.31-44.11v24.68a119.89862,119.89862,0,0,1-15.92,19.43Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M546.68821,498.44051q-18.135,32.535-37.95,64.08h-2.3q20.145-32.04,38.61-65.15A9.82861,9.82861,0,0,1,546.68821,498.44051Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M457.96818,522.1205a407.71473,407.71473,0,0,1-26.11,40.4h-13.71q13.47006-16.41,26.96-32.82c2.58-3.15,5.38-6.45,9.25-7.73Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M439.7482,519.88051c-1.43,1.49-2.8,3.04-4.09,4.65-2.84,3.42-5.64,6.87-8.45,10.31-1.75,2.13995-3.5,4.28-5.23,6.42a12.77008,12.77008,0,0,1-1.56-1.18,13.99258,13.99258,0,0,1,17.88-21.52A12.89785,12.89785,0,0,1,439.7482,519.88051Z'
-                transform='translate(-304.59818 -148.62109)'
-                className='bg-primary'
-            />
-            <path
-                d='M517.07816,529.60048l-26.28-9.61-13.05-4.77c-4.25-1.55-8.49-3.32-12.88-4.44a20.3055,20.3055,0,0,0-11.66.15,26.80777,26.80777,0,0,0-9.09,5.64q-1.74,1.605-3.37,3.31c-1.43,1.49-2.8,3.04-4.09,4.65-2.84,3.42-5.64,6.87-8.45,10.31-1.75,2.13995-3.5,4.28-5.23,6.42q-8.67,10.605-17.24,21.26a10.02168,10.02168,0,0,1-2.28-.29q8.94-11.09994,17.96-22.15c.69-.86,1.4-1.73,2.1-2.59q4.395-5.38494,8.8-10.76c2.27-2.76995,4.52-5.56,6.98-8.17.22-.24.44-.47.67-.71,4.99-5.16,11.22-9.72,18.68-9.84,4.39-.07,8.63,1.32,12.71,2.79,4.39,1.59,8.77,3.21,13.16,4.82l26.46,9.68,6.62,2.42C518.76816,528.14052,518.26816,530.04048,517.07816,529.60048Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M431.63572,438.31194H400.44741a6.29506,6.29506,0,0,1-6.288-6.288V402.84782a6.295,6.295,0,0,1,6.288-6.288h31.18831a6.295,6.295,0,0,1,6.288,6.288V432.024A6.29506,6.29506,0,0,1,431.63572,438.31194Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M487.26687,438.31194H456.07856a6.29506,6.29506,0,0,1-6.288-6.288V402.84782a6.295,6.295,0,0,1,6.288-6.288h31.18831a6.295,6.295,0,0,1,6.288,6.288V432.024A6.29506,6.29506,0,0,1,487.26687,438.31194Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M542.898,438.31194h-31.1883a6.29506,6.29506,0,0,1-6.288-6.288V402.84782a6.29505,6.29505,0,0,1,6.288-6.288H542.898a6.29506,6.29506,0,0,1,6.288,6.288V432.024A6.29506,6.29506,0,0,1,542.898,438.31194Z'
-                transform='translate(-304.59818 -148.62109)'
-                fill='#e6e6e6'
-            />
-        </svg>
-    )
-}
-
-export default CustomizationSVG
