--- src/components/svg/EmptyGroupSVG.jsx
@@ -1,220 +1,0 @@
-const EmptyGroupSVG = ({ width, height }) => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            width={width}
-            height={height}
-            viewBox='0 0 583.04004 614.68419'
-        >
-            <g>
-                <polygon
-                    points='303.29327 487.86387 284.26402 486.81764 286.88489 545.20564 299.77945 545.91545 303.29327 487.86387'
-                    fill='#a0616a'
-                />
-                <path
-                    d='M285.19656,526.74635l6.61317,2.9996v4.00488s1.58398-4.63849,10.13756-5.66809l1.0296,3.88077s3.32639,3.80157,2.772,10.69196l-2.1384,18.45352-6.45264-.13477-1.06961-12.83548s-1.78371-2.9093-9.94127,10.95065c0,0-3.00665,4.58157-16.98299-.07141-1.92096-.63953-3.25821-2.4288-3.23743-4.45331,.01147-1.11725,.44595-2.1507,1.84608-2.52405,3.564-.95041,7.04877,.23761,9.34557-2.61359s8.07837-22.68073,8.07837-22.68073l-.00003,.00003Z'
-                    fill='#2f2e41'
-                />
-            </g>
-            <g>
-                <polygon
-                    points='255.65484 541.43907 236.62559 540.39283 239.24646 598.78084 252.14102 599.49064 255.65484 541.43907'
-                    fill='#a0616a'
-                />
-                <path
-                    d='M237.55813,580.32155l6.61317,2.9996v4.00488s1.58398-4.63849,10.13756-5.66809l1.0296,3.88077s3.32639,3.80157,2.772,10.69196l-2.1384,18.45352-6.45264-.13477-1.06961-12.83548s-1.78371-2.9093-9.94127,10.95065c0,0-3.00665,4.58157-16.98299-.07141-1.92096-.63953-3.25821-2.4288-3.23743-4.45331,.01147-1.11725,.44595-2.1507,1.84608-2.52405,3.564-.95041,7.04877,.23761,9.34557-2.61359s8.07837-22.68073,8.07837-22.68073l-.00003,.00003Z'
-                    fill='#2f2e41'
-                />
-            </g>
-            <g>
-                <path
-                    d='M494.45289,366.96353c-.8443-1.42365-4.49683,.10068-5.16833-1.19135-.66913-1.28735,2.72449-3.24887,4.66333-7.7298,.34967-.80807,2.55298-5.90033,.92188-7.47162-3.09174-2.97836-17.23962,9.31259-22.19073,5.29041-1.08661-.88269-1.8457-2.64484-5.13428-5.50583-1.30829-1.13815-2.09937-1.65204-2.86499-1.46085-1.08636,.27127-1.17242,1.72522-2.33734,4.58398-1.74701,4.28729-2.78027,4.02841-3.79816,7.44897-.75562,2.53925-.58301,4.01553-1.47217,4.30313-1.25824,.40701-2.41638-2.28632-4.01099-2.04517-1.6272,.24612-2.61096,3.38205-2.92169,5.72998-.58313,4.40607,1.03497,7.31348,1.93182,10.37064,.97461,3.32236,1.55304,8.37494-.69769,15.52759l-22.27719,62.76614c4.91812-12.87943,19.15842-48.18536,25.14218-61.3053,1.72742-3.78745,3.6347-7.62909,7.55096-9.14526,3.77045-1.45978,8.94867-.58823,15.55029-2.17862,.771-.1857,2.91052-.73007,3.17981-1.98849,.22266-1.04037-1.03583-1.62109-.84247-2.59549,.25934-1.30701,2.73413-1.32306,5.77533-2.83096,2.14447-1.06332,3.48047-2.32785,4.62933-3.41531,.34637-.32785,5.49512-5.26166,4.37115-7.1568h-.00006Z'
-                    fill='#f2f2f2'
-                />
-                <path
-                    d='M406.80777,400.00625c-.94235,.0036-1.24805,2.23611-2.07593,2.1929-.82492-.04303-.80847-2.2746-2.4483-4.5191-.29572-.40475-2.1593-2.95547-3.40115-2.60825-2.35385,.65808-.4035,11.14828-3.80618,12.41782-.74673,.27863-1.83066,.14276-4.18369,.93134-.9361,.31372-1.41656,.55371-1.54364,.98465-.1803,.61145,.50845,1.07324,1.57544,2.46979,1.60016,2.09445,1.17496,2.52692,2.56009,4.01343,1.02826,1.10352,1.80264,1.44476,1.68723,1.96414-.16327,.73499-1.81934,.52631-2.16107,1.37851-.34869,.86966,.90659,2.25729,1.9693,3.08725,1.99429,1.5575,3.88812,1.6022,5.64737,2.04413,1.91183,.48026,4.55859,1.65421,7.41974,4.82266l24.37836,29.04382c-4.90228-6.12997-18.12189-23.30609-22.83472-30.0285-1.3605-1.94061-2.69568-3.98517-2.30984-6.34479,.37143-2.27176,2.29327-4.56177,3.41748-8.26077,.13132-.43201,.48145-1.63916-.05847-2.13443-.44638-.40945-1.09451,.04068-1.51697-.33539-.56665-.50443,.13953-1.72369,.27689-3.6514,.09686-1.35931-.13831-2.37988-.34055-3.25754-.06097-.26459-.99695-4.21518-2.25137-4.21036v.00009Z'
-                    fill='#f2f2f2'
-                />
-                <path
-                    d='M425.6074,404.22635l-.56775-5.06818,.2594-.38422c1.20035-1.77667,1.81216-3.51175,1.81949-5.15872,.00122-.26184-.01102-.52371-.02325-.79044-.04895-1.05719-.11011-2.37134,.5751-3.90329,.38422-.85406,1.46344-2.83142,3.08594-2.5867,.43683,.06119,.76721,.26184,1.01804,.49677,.03671-.05872,.07465-.11746,.11502-.18109,.50534-.77087,.90546-1.10123,1.29089-1.41937,.29611-.24472,.60202-.49677,1.08167-1.0939,.21045-.26184,.37442-.49188,.5127-.68521,.41971-.58487,.96786-1.28235,1.93451-1.27011,1.03271,.04895,1.57355,.90302,1.93207,1.46832,.63995,1.00824,.92993,1.73508,1.12204,2.21716,.06976,.17621,.14929,.37442,.18967,.43561,.33282,.48944,3.03824,.0318,4.05746-.13461,2.28937-.38177,4.27039-.71213,5.08655,.73172,.5849,1.03271,.15051,2.39825-1.33008,4.1676-.4613,.55063-.95197,.96909-1.38757,1.29703,.36096,.21536,.68396,.54572,.81494,1.06454h0c.30835,1.22849-.74762,2.45944-3.13733,3.66348-.59344,.30099-1.39246,.70236-2.51205,.89569-.52737,.09055-1.00214,.11258-1.40836,.12482-.00858,.23492-.06363,.48944-.20679,.75128-.41971,.77087-1.29333,1.13794-2.60995,1.04984-1.44876-.07587-2.64175-.35239-3.69406-.59467-.91891-.21045-1.7106-.3891-2.33218-.3475-1.15262,.09299-2.04099,1.01559-3.08347,2.21228l-2.6026,3.04187h-.00003Z'
-                    fill='#f2f2f2'
-                />
-                <path
-                    d='M428.24109,374.05041l-4.75842,1.83472-.43744-.15347c-2.02304-.71045-3.85654-.86215-5.45151-.45148-.2536,.06525-.50378,.14349-.7587,.22299-1.01019,.31549-2.26587,.70801-3.92151,.43378-.92361-.155-3.11002-.69739-3.28485-2.32889-.05161-.43805,.05869-.80853,.22232-1.11075-.06613-.0206-.13257-.04239-.20435-.06531-.87384-.2933-1.29489-.59653-1.70041-.88867-.31183-.22437-.63324-.45633-1.33249-.76883-.30667-.13715-.57077-.2374-.79285-.32211-.67221-.2576-1.4859-.61096-1.71924-1.5491-.2146-1.01135,.47437-1.75113,.93024-2.24133,.81296-.87476,1.44244-1.33963,1.86005-1.64774,.15274-.11215,.32431-.23938,.37326-.29395,.38901-.44608-.73987-2.94693-1.15936-3.89062-.94998-2.11765-1.77203-3.95007-.5824-5.10577,.85059-.8277,2.28165-.75391,4.36868,.22946,.64963,.30655,1.17886,.67502,1.60657,1.01321,.11676-.40378,.35437-.80005,.82303-1.05829h0c1.11011-.60986,2.56863,.09933,4.33945,2.10547,.44168,.49768,1.03259,1.16876,1.50357,2.2027,.22134,.48715,.36307,.9408,.47797,1.33066,.22943-.0513,.48956-.06259,.77917,.00946,.85211,.21045,1.4288,.9624,1.67752,2.2583,.2941,1.42062,.32922,2.64474,.36179,3.72409,.02951,.94226,.0575,1.75336,.25543,2.34406,.38232,1.09134,1.50006,1.71661,2.92206,2.42148l3.60254,1.74591-.00012,.00003Z'
-                    fill='#f2f2f2'
-                />
-                <path
-                    d='M425.6074,359.11941l-.56775-5.06818,.2594-.38422c1.20035-1.77667,1.81216-3.51175,1.81949-5.15872,.00122-.26184-.01102-.52371-.02325-.79044-.04895-1.05719-.11011-2.37134,.5751-3.90329,.38422-.85406,1.46344-2.83142,3.08594-2.5867,.43683,.06119,.76721,.26184,1.01804,.49677,.03671-.05872,.07465-.11746,.11502-.18109,.50534-.77087,.90546-1.10123,1.29089-1.41937,.29611-.24472,.60202-.49677,1.08167-1.0939,.21045-.26184,.37442-.49188,.5127-.68521,.41971-.58487,.96786-1.28235,1.93451-1.27011,1.03271,.04895,1.57355,.90302,1.93207,1.46832,.63995,1.00824,.92993,1.73508,1.12204,2.21716,.06976,.17621,.14929,.37442,.18967,.43561,.33282,.48944,3.03824,.0318,4.05746-.13461,2.28937-.38177,4.27039-.71213,5.08655,.73172,.5849,1.03271,.15051,2.39825-1.33008,4.1676-.4613,.55063-.95197,.96909-1.38757,1.29703,.36096,.21536,.68396,.54572,.81494,1.06454h0c.30835,1.22849-.74762,2.45944-3.13733,3.66348-.59344,.30099-1.39246,.70236-2.51205,.89569-.52737,.09055-1.00214,.11258-1.40836,.12482-.00858,.23492-.06363,.48944-.20679,.75128-.41971,.77087-1.29333,1.13794-2.60995,1.04984-1.44876-.07587-2.64175-.35239-3.69406-.59467-.91891-.21045-1.7106-.3891-2.33218-.3475-1.15262,.09299-2.04099,1.01559-3.08347,2.21228l-2.6026,3.04187h-.00003Z'
-                    fill='#f2f2f2'
-                />
-                <path
-                    d='M430.79303,457.49716l-1.0621-.66565-.25696-1.22604,.25696,1.22604-1.23828,.14194c-.01959-.11502-.08688-.37933-.18842-.79291-.55429-2.26611-2.24408-9.16235-3.65369-20.02304-.9838-7.58145-1.54053-15.37579-1.65555-23.17017-.11502-7.80658,.25452-13.71661,.55063-18.46661,.22391-3.5827,.49557-6.99411,.7623-10.31744,.70969-8.87848,1.379-17.26505,.88098-26.54974-.11011-2.07278-.3414-6.38721-2.862-10.97818-1.46222-2.66257-3.47626-5.02655-5.98709-7.02347l1.56131-1.96265c2.77145,2.20737,4.99963,4.8259,6.62338,7.78214,2.79471,5.09018,3.04678,9.79129,3.16791,12.05005,.50656,9.44867-.17007,17.91602-.88712,26.88013-.26431,3.31107-.53595,6.70779-.75864,10.27582-.29367,4.70599-.65952,10.56458-.54572,18.27084,.11258,7.70135,.66318,15.40027,1.63351,22.88629,1.39246,10.72366,3.0578,17.51712,3.60474,19.74896,.29123,1.1918,.35239,1.44141,.05383,1.91373v-.00003Z'
-                    fill='#f2f2f2'
-                />
-                <path
-                    d='M411.74396,343.63354c-.10278,0-.20679-.00244-.31201-.0098-2.13397-.11502-4.11008-1.39246-5.87451-3.79562-.82715-1.13062-1.25052-2.42029-2.09482-4.99231-.13092-.39645-.76721-2.40805-1.15387-5.17584-.2533-1.8085-.22147-2.56467,.13705-3.23276,.39767-.74396,1.04129-1.26276,1.76566-1.6127-.0318-.23984-.00613-.487,.0979-.73907,.42703-1.04495,1.56253-.9079,2.17679-.84183,.31079,.03915,.69867,.09055,1.11716,.06851,.65829-.0318,1.01193-.2276,1.54785-.52127,.5127-.28143,1.15018-.63138,2.09604-.82471,1.86478-.3891,3.42731,.1395,3.94244,.31323,2.71027,.90302,4.04645,3.01007,5.59308,5.44992,.30835,.48944,1.36676,2.271,2.06543,4.71332,.50412,1.76199,.43317,2.5451,.28143,3.12753-.30835,1.19669-1.03638,1.8819-2.88525,3.39429-1.93085,1.58334-2.89993,2.37625-3.73199,2.85834-1.93695,1.11838-3.15323,1.82071-4.76837,1.82071v.00006Z'
-                    fill='#f2f2f2'
-                />
-            </g>
-            <g>
-                <polygon
-                    points='146.01363 451.57101 157.87274 451.56985 163.51419 405.82702 146.01125 405.82824 146.01363 451.57101'
-                    fill='#ffb6b6'
-                />
-                <path
-                    d='M186.64445,451.302h0c.36925,.62189,.56409,2.62906,.56409,3.35226h0c0,2.22293-1.80202,4.02502-4.02502,4.02502h-36.7263c-1.51648,0-2.7459-1.22934-2.7459-2.74591v-1.52902s-1.8168-4.59555,1.92369-10.2598c0,0,4.64894,4.43521,11.59566-2.51151l2.04852-3.71106,14.82851,10.84482,8.21921,1.01172c1.7982,.22131,3.39246-.03436,4.31743,1.52347h.00012Z'
-                    fill='#2f2e41'
-                />
-            </g>
-            <g>
-                <polygon
-                    points='101.01362 451.57101 112.87274 451.56985 118.51419 405.82702 101.01124 405.82824 101.01362 451.57101'
-                    fill='#ffb6b6'
-                />
-                <path
-                    d='M141.64445,451.302h0c.36925,.62189,.56409,2.62906,.56409,3.35226h0c0,2.22293-1.80202,4.02502-4.02502,4.02502h-36.72631c-1.51648,0-2.7459-1.22934-2.7459-2.74591v-1.52902s-1.81681-4.59555,1.92369-10.2598c0,0,4.64894,4.43521,11.59566-2.51151l2.04852-3.71106,14.82851,10.84482,8.21921,1.01172c1.7982,.22131,3.39246-.03436,4.31743,1.52347h.00014Z'
-                    fill='#2f2e41'
-                />
-            </g>
-            <polygon
-                points='92.27986 205.85551 77.27986 239.85551 96.27986 425.85553 119.77986 427.35553 129.27986 262.85553 138.27986 427.85553 166.27986 424.85553 172.92159 215.39489 92.27986 205.85551'
-                fill='#2f2e41'
-            />
-            <path
-                d='M151.3783,70.35552l-30.09843-5.5-6.07874,6.07874-30.92126,12.92126,5,97.00001s-10,6.66142-5,15.8307,4,20.1693,4,20.1693h84.64174l-3.64172-55,10-70-23.36613-11.0625-4.53543-10.4375h-.00002Z'
-                fill='#ccc'
-            />
-            <path
-                id='b-281'
-                d='M129.56279,183.36398c8.55096,2.04129,16.53548-.71614,17.83366-6.15823,1.2982-5.44209-4.58109-11.5069-13.13538-13.54778-3.4059-.86212-6.96874-.89606-10.39044-.09895l-36.34-8.1915-3.48054,17.13522,36.19433,6.25613c2.69377,2.2572,5.88904,3.83629,9.31834,4.60509l.00002,.00002Z'
-                fill='#ffb6b6'
-            />
-            <path
-                d='M90.27986,88.85552l-6-5s-12,6-14,11-14,53.00001-14,53.00001c0,0-2.35433,26.30624,3.82283,29.65312s45.17717,1.34688,45.17717,1.34688l3-20-16-6-2-64.00001Z'
-                fill='#ccc'
-            />
-            <circle cx='138.21861' cy='35.58432' r='26.43324' fill='#ffb6b6' />
-            <path
-                d='M119.26031,59.04964c-3.16664,.28297-8.33604-7.81783-8.10563-15.66315,.14823-5.04705-4.23354-6.03942-4.55911-12.54824-.18969-3.79259,5.63598-4.97511,6.03616-8.01702,.72506-5.51164,4.28094-9.22043,4.72488-9.6713,3.83216-3.89193,6.53676-2.62671,11.0433-7.13736,2.71317-2.71563,2.01797-3.45962,3.84521-4.66415,5.27354-3.47634,14.74936,.29135,20.41591,5.36658,1.96088,1.75623,9.27003-.16323,12.46283,4.79619,3.76167,5.8431,2.88828,17.7594,1.27522,19.33202-.98001,.95544-3.21809-1.37882-9.284-2.87694-5.61601-1.38702-.31822,5.4722-6.81078,4.32115-7.67747-1.36114-9.18057-3.69693-12.05589-2.29598-3.03947,1.48093-5.41446,6.06658-4.72488,9.6713,.34218,1.78881,1.33388,2.78625,.77461,3.69587-1.021,1.66068-2.19258,.03197-3.50221,1.69859-1.52328,1.9385,2.26202,6.9115,1.54924,7.39175-.14362,.09678-.57838,.08416-4.66415-3.84522-1.63144-1.56899-2.10973-2.53248-2.4289-2.38459-.5167,.2394-3.90345,3.39443-2.61026,5.88597,2.01246,3.87731-.56309,3.9981,0,6.60831,.58094,2.69284-1.36913,.1564-3.38159,.33623h.00003Z'
-                fill='#2f2e41'
-            />
-            <polygon
-                points='113.77986 139.35551 109.77986 142.35551 130.77986 171.35551 134.77986 170.35551 113.77986 139.35551'
-                fill='#3f3d56'
-            />
-            <g>
-                <path
-                    d='M167.77986,165.90214v50.61444c0,4.47604-3.64136,8.11742-8.11742,8.11742h-50.61444c-4.47605,0-8.11741-3.64136-8.11741-8.11742v-50.61444c0-4.47604,3.64136-8.11742,8.11741-8.11742h50.61444c4.47604,0,8.11742,3.64136,8.11742,8.11742Z'
-                    fill='#fff'
-                />
-                <g>
-                    <path
-                        d='M141.09291,178.45652h-16.34016c-.41195,0-.74714-.33521-.74714-.74715s.33519-.74699,.74714-.74699h16.34016c.41196,0,.74699,.33504,.74699,.74699s-.33504,.74715-.74699,.74715Z'
-                        fill='#3f3d56'
-                    />
-                    <path
-                        d='M141.09291,194.38748h-16.34016c-.41195,0-.74714-.33521-.74714-.74715s.33519-.74699,.74714-.74699h16.34016c.41196,0,.74699,.33504,.74699,.74699s-.33504,.74715-.74699,.74715Z'
-                        fill='#3f3d56'
-                    />
-                    <path
-                        d='M132.97549,216.35223h-16.34016c-.41195,0-.74714-.33521-.74714-.74715s.33519-.74699,.74714-.74699h16.34016c.41196,0,.74699,.33504,.74699,.74699s-.33504,.74715-.74699,.74715Z'
-                        fill='#3f3d56'
-                    />
-                    <path
-                        d='M149.86121,186.42723h-33.87675c-.41195,0-.74714-.33521-.74714-.74715s.33519-.74699,.74714-.74699h33.87675c.41196,0,.74699,.33504,.74699,.74699s-.33504,.74715-.74699,.74715Z'
-                        fill='#3f3d56'
-                    />
-                </g>
-            </g>
-            <path
-                id='c-282'
-                d='M147.35931,187.47883c-7.79353,4.06776-12.01196,11.38629-9.42239,16.34572s11.00543,5.68114,18.80121,1.61087c3.13638-1.58319,5.84802-3.89445,7.90804-6.74046l32.80383-17.65234-8.61185-15.21732-31.42445,19.01768c-3.51389,.06303-6.96155,.96686-10.05438,2.63585Z'
-                fill='#ffb6b6'
-            />
-            <path
-                d='M173.27986,92.85552s9-6,15,10c4.73351,12.62267,13.51254,49.82973,17.00784,64.97489,.92789,4.02048-.41492,8.22028-3.49884,10.96156l-4.009,3.56355-19.5,14.5-13.29527-19.34688,16.14764-9.73579-18.41534-39.94857,10.56299-34.96875-.00002-.00002Z'
-                fill='#ccc'
-            />
-            <path
-                d='M583.04004,458.35769h0c0,.46393-.3761,.84003-.84003,.84003H0v-1.68005H582.20001c.46393,0,.84003,.3761,.84003,.84003Z'
-                fill='#e6e6e6'
-            />
-            <path
-                id='d-283'
-                d='M267.57451,453.80015c.21127-2.10977-.08593-4.23945-.86671-6.21081l3.29375-21.91443-10.47333-1.26228-2.21318,22.3655c-1.10652,1.80791-1.76392,3.85453-1.91732,5.96864-.45649,5.28614,1.89944,9.80496,5.26236,10.09546s6.45878-3.7578,6.9144-9.04212l.00004,.00004Z'
-                fill='#a0616a'
-            />
-            <path
-                d='M297.88696,324.33156l-8.64394,45.41342-10.7203,23.82288-6.38481,45.93698-16.40538,4.96443,5.14459-58.24861s16.5694-54.12172,21.01169-59.63294c4.44229-5.51122,15.99818-2.25617,15.99818-2.25617h-.00004Z'
-                className='fill-primary'
-            />
-            <g>
-                <rect
-                    x='301.15348'
-                    y='284.71108'
-                    width='35.51393'
-                    height='37.45106'
-                    transform='translate(-64.66906 519.40978) rotate(-72.95725)'
-                    fill='#2f2e41'
-                />
-                <circle
-                    cx='324.54249'
-                    cy='286.41608'
-                    r='18.26031'
-                    fill='#a0616a'
-                />
-                <path
-                    d='M341.46311,260.71066c-5.62235-2.72288-16.25609-1.44609-20.05826-1.02668-3.39636,.37463-5.75092,2.34433-8.06601,4.28104-5.25501,4.39607-7.56036,9.75249-8.29566,11.63967-3.47561,8.92126-.05273,12.67283-3.68572,16.42966-3.89223,4.02486-9.34954,1.29932-11.7205,5.18687-2.15402,3.53162,1.84003,6.6167,.29765,12.24782-1.77974,6.49804-8.61438,7.92623-7.91716,10.40497,.77383,2.75071,10.01076,3.90835,17.3759,.59896,.99343-.44644,9.7251-4.50009,10.14773-11.07192,.27098-4.21504-3.18571-4.65001-4.14143-10.72465-.66447-4.22348,.40905-7.80999,2.12212-13.53215,1.19361-3.98737,1.79045-5.98105,3.07757-7.83633,.62101-.89515,6.46723-9.0362,15.13616-7.51664,7.64169,1.33946,12.77845,9.362,11.7885,14.41958-.58468,2.98679-3.36435,5.23691-2.27097,7.40823,.80617,1.60107,2.87733,1.48907,3.13636,2.98754,.24318,1.40657-1.34957,2.84796-2.55945,3.94297-1.5289,1.38371-2.89986,2.04237-3.41566,2.32976-3.67727,2.04933-5.04017,8.16432-5.01966,11.96856,.03074,5.67912,3.17937,12.66459,9.55794,15.08653,7.16397,2.72018,15.90092-1.26615,16.2312-4.47949,.19184-1.86605-2.55647-2.45017-2.99673-5.6462-.49671-3.60511,2.65553-5.36226,1.65361-7.59748-.90774-2.02505,.41549,.8649-2.09093-1.9917-.3062-.34902-1.92277-2.25343-1.95127-4.65034-.03708-3.1224,2.64299-5.18211,4.12304-6.84048,5.53937-6.2068,6.41259-17.91827,1.28045-23.63984-2.91304-3.2476-1.11012-7.23073-11.73889-12.37827h.00004Z'
-                    fill='#2f2e41'
-                />
-                <path
-                    d='M344.14013,278.41544c1.65079-3.74302-2.71566-8.50512-3.28521-9.11146-4.25985-4.5351-10.37773-4.69967-13.3925-4.78079-2.63874-.07095-8.29649-.22326-12.57672,3.57368-2.31485,2.05344-4.47499,5.48212-3.50568,7.02973,.59292,.94674,2.31049,.99311,5.74542,1.08588,3.08901,.08343,4.2067-.32351,5.93467,.46853,1.27631,.58506,1.59134,1.23102,2.70825,2.18093,3.02192,2.56993,6.73696,2.96166,7.83632,3.07756,2.59571,.27369,8.91057,.15998,10.5354-3.52406h.00004Z'
-                    fill='#2f2e41'
-                />
-            </g>
-            <polygon
-                points='300.55889 320.31249 285.07402 325.07707 286.26517 339.37079 281.50059 427.51546 361.30724 429.89775 362.32575 421.69449 354.16038 392.15161 357.73381 333.41508 335.10207 321.50364 329.14635 309.5922 307.70576 308.40105 300.55889 320.31249'
-                className='fill-primary'
-            />
-            <path
-                d='M335.69764,412.62616l-54.79263,16.67602s-33.09295,9.46931-50.96011,18.99846c-17.86716,9.52915-15.74396,18.17645-15.74396,18.17645l22.63174,101.24724,21.44059,1.19114-2.90752-93.13432,23.15697-3.15697,3.57343,34.10235,19.0583,4.76458,19.0583-65.5129h0c1.36875,8.21254,8.972,13.8824,17.23186,12.83622,6.89884-.87378,15.31868-2.21206,23.26704-4.1981,17.86716-4.46445,0-36.03445,0-36.03445l-25.01402-5.95572Z'
-                fill='#2f2e41'
-            />
-            <path
-                d='M279.1183,463.24978s3.57343-28.58746,9.52915-60.74835c5.95572-32.16089,27.39631-23.82288,27.39631-23.82288l38.11661,40.4989-30.96975,20.24945-44.07233,23.82288Z'
-                fill='#2f2e41'
-            />
-            <path
-                id='e-284'
-                d='M311.44571,432.36354c1.56425-1.43139,2.76284-3.21669,3.4955-5.20642l17.06722-14.13507-6.96337-7.92437-16.56275,15.19177c-2.03012,.60949-3.88474,1.69633-5.40875,3.16956-3.865,3.63503-5.12245,8.57354-2.80999,11.03246s7.31902,1.50632,11.18217-2.12792h-.00004Z'
-                fill='#a0616a'
-            />
-            <path
-                d='M325.61963,353.65488l13.63203,26.55539-23.40354,30.17987,11.7804,12.33214,29.28303-24.69027,13.73611-12.93675c4.98188-4.69196,2.38952-10.878-.07906-17.26074-9.85693-25.48588-7.47464-35.01503-17.00379-37.39732-7.75185-1.93797-27.94514,23.21766-27.94514,23.21766l-.00004,.00002Z'
-                className='fill-primary'
-            />
-            <g>
-                <polygon
-                    points='263.17418 403.63754 265.4026 397.39802 306.85082 404.52889 342.05949 386.70171 343.84221 391.60419 305.51375 414.33384 263.17418 403.63754'
-                    fill='#e6e6e6'
-                />
-                <polygon
-                    points='345.62492 388.48443 305.95946 408.54 260.62257 401.98859 264.95692 421.01904 303.28536 427.70424 339.52052 410.34114 345.62492 388.48443'
-                    fill='#3f3d56'
-                />
-            </g>
-            <g>
-                <polygon
-                    points='104.61118 172.30328 106.8396 166.06376 148.28781 173.19463 183.49649 155.36745 185.27921 160.26993 146.95074 182.99959 104.61118 172.30328'
-                    fill='#e6e6e6'
-                />
-                <polygon
-                    points='187.06192 157.15017 147.39645 177.20575 102.05957 170.65433 106.39392 189.68478 144.72235 196.36998 180.95752 179.00688 187.06192 157.15017'
-                    fill='#3f3d56'
-                />
-            </g>
-        </svg>
-    )
-}
-
-export default EmptyGroupSVG
