--- src/components/svg/MenuSVG.jsx
@@ -1,27 +1,0 @@
-const MenuSVG = ({ width, height }) => {
-    return (
-        <svg
-            part='svg'
-            xmlns='http://www.w3.org/2000/svg'
-            aria-hidden='true'
-            aria-labelledby='actionDots'
-            focusable='false'
-            viewBox='0 0 24 24'
-            className='h-icon icon-primary cursor-pointer d-none d-md-flex fill-black dark:fill-gray-200'
-            width={width}
-            height={height}
-            // dataV68e5bb38=''
-            // block=''
-        >
-            <g>
-                <g>
-                    <circle cx='12.1091' cy='4' r='2'></circle>
-                    <circle cx='12.1091' cy='20' r='2'></circle>
-                    <circle cx='12.1091' cy='12' r='2'></circle>
-                </g>
-            </g>
-        </svg>
-    )
-}
-
-export default MenuSVG
