--- src/components/svg/PeopleSVG.jsx
@@ -1,156 +1,0 @@
-const PeopleSVG = ({ width, height }) => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            data-name='Layer 1'
-            width={width}
-            height={height}
-            viewBox='0 0 838 638.00376'
-        >
-            <path
-                d='M689.13548,323.255a5.09028,5.09028,0,0,0,3.68913-8.8227c-.12836-.51014-.22044-.8765-.34879-1.38664q.06888-.16659.13859-.33286c4.6863-11.17449,20.5687-11.09739,25.21513.09369,4.12406,9.93267,9.37471,19.88226,10.66755,30.38441a40.44007,40.44007,0,0,1-.70989,13.908,162.132,162.132,0,0,0,14.74848-67.33847,156.48736,156.48736,0,0,0-.97069-17.45764q-.80412-7.128-2.23118-14.15449a164.01034,164.01034,0,0,0-32.52468-69.51166,43.64774,43.64774,0,0,1-18.153-18.83392,33.29411,33.29411,0,0,1-3.02794-9.09827c.88371.11592,3.3322-13.34322,2.66567-14.169,1.23157-1.86885,3.43587-2.79784,4.78094-4.62148,6.68954-9.06961,15.906-7.486,20.71739,4.83888,10.27825,5.18745,10.37774,13.79049,4.07105,22.06463-4.01234,5.26406-4.56364,12.38693-8.08418,18.02253.36227.46365.739.91282,1.10112,1.37648a165.06913,165.06913,0,0,1,17.22906,27.29491c-1.36865-10.69089.652-23.57059,4.09674-31.873,3.9213-9.45971,11.27119-17.4267,17.74364-25.6046,7.77447-9.82292,23.71674-5.536,25.08643,6.91609q.01989.18086.03883.36158-1.44205.8135-2.8232,1.727a6.90459,6.90459,0,0,0,2.78379,12.56506l.1408.0217a68.70125,68.70125,0,0,1-1.81092,10.27183c8.30335,32.11122-9.623,43.80679-35.21939,44.33218-.565.28972-1.11566.57943-1.6807.8548a168.48694,168.48694,0,0,1,9.0693,42.68053,159.80436,159.80436,0,0,1-.11592,25.80246l.04349-.30421a42.227,42.227,0,0,1,14.41527-24.38271c11.09331-9.11275,26.76607-12.46849,38.73379-19.79336,5.76092-3.526,13.14312,1.03057,12.13769,7.70959q-.02407.1599-.04894.31988a46.37722,46.37722,0,0,0-5.201,2.50624q-1.44205.8137-2.8232,1.727a6.90468,6.90468,0,0,0,2.7838,12.56519l.14091.02164c.10137.01448.18834.029.2896.04359a68.7599,68.7599,0,0,1-12.6476,19.80451c-5.19187,28.03184-27.491,30.69143-51.34422,22.52828h-.01453a168.43679,168.43679,0,0,1-11.31491,33.03181H690.14836c-.145-.4491-.27531-.91275-.40575-1.36192a45.9261,45.9261,0,0,0,11.18457-.66624c-2.999-3.67994-5.99787-7.38886-8.99686-11.06867a2.51118,2.51118,0,0,1-.18835-.21734c-1.52118-1.88328-3.05689-3.75226-4.57807-5.6356l-.00081-.00229a67.27051,67.27051,0,0,1,1.97111-17.13665Z'
-                transform='translate(-181 -130.99812)'
-                fill='#f2f2f2'
-            />
-            <path
-                d='M818.913,219.8102c-7.51576-27.40163,53.10245-50.89907,69.02069-28.83249,2.21348-3.1372,10.39357-5.04278,14.14681-4.08041s6.90971,3.36827,9.98934,5.69723c4.23443,3.25279,8.62274,6.62107,11.22108,11.29814,2.61772,4.65781,3.02183,11.0095-.50033,15.05138-2.79088,3.21426-11.02877,5.293-15.22466,6.13988-2.92551.59667-5.38921,1.07785-7.50637,1.42428a14.35164,14.35164,0,0,1-2.50224-9.48889,11.64833,11.64833,0,0,0-5.08119,8.5458c-.05779.71214,23.65486,87.96005,23.65486,87.96005a51.72043,51.72043,0,0,1-8.95,4.02262,112.54468,112.54468,0,0,1-13.95427-30.41067,97.45318,97.45318,0,0,0,.26949,21.65318,119.1445,119.1445,0,0,0,2.0402,11.991c-28.178,6.08221-70.56044,4.08053-88.99936-21.7878q-.00491-.16053-.00929-.3213a4.40184,4.40184,0,0,0-8.50333-1.53505c-.44736,1.1413-.85908,2.12916-1.2265,2.91492C791.42794,280.84323,798.37617,258.61265,818.913,219.8102Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M770.96283,679.20736v.89737A3.809,3.809,0,0,1,770.96283,679.20736Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M798.5576,524.13266a15.32366,15.32366,0,0,1,2.183-.424Q799.47155,523.91547,798.5576,524.13266Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M797.919,524.2954c.19365-.0543.40921-.1086.63857-.16274C798.34394,524.18828,798.13849,524.2292,797.919,524.2954Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M1001.7905,768.99812H898.50437a8.21329,8.21329,0,0,1-2.2114-10.1149c2.82162-6.47111-2.70755-12.38425-2.70755-12.38425a17.5668,17.5668,0,0,1-3.53986-16.19126c2.73658-9.55783-7.98759-36.535-7.98759-36.535s-5.78127,34.83426-12.86549,75.22543H770.96283a18.25939,18.25939,0,0,1,0-18.48907c5.72448-9.70643,0-11.56794,0-11.56794s-5.72448-10.86155,0-18.748,0-40.08836,0-40.08836c.21572,1.42376,1.84656-1.16486,4.35751-6.59238,3.82253-8.25772,9.66421-23.10776,15.56624-40.348,4.21947-12.30455,8.47344-25.84306,12.02848-39.07958h.00864q1.39788-5.216,2.649-10.33723c1.63084-6.69591,3.0287-13.202,4.09-19.31114,2.30386-13.15884,3.06319-24.46254,1.29431-31.8142-1.4669-6.1178-4.67677-9.49162-10.21641-8.91351,2.96829-.49184,7.31716-.95778,12.67563-1.39785.96642-.07768,1.96733-.15531,3.00279-.24163,39.96837-3.02006,125.72089-4.7544,125.72089-4.7544L968.40587,564.35l33.26382,59.58155S1000.76368,703.8597,1001.7905,768.99812Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M891.54674,314.01971,831.0329,301.18327l-12.01528,24.1852L778.55553,348.7939s26.08156,106.64135,25.93277,121.63984c-.37777,38.079-61.30588,123.44042-15.28485,116.87355s55.36918-34.07334,70.14317-29.724c13.3243,3.92253,29.59359,21.37211,40.48552,8.89216s14.72832,1.93129,21.61565,2.42934c3.27307.23669,1.30185-15.75881,9.53707.49707s51.90445-5.62623,37.10363-20.42705c-24.65-24.65-29.65972-206.36132-29.65972-206.36132l-27.4406-8.44326Z'
-                transform='translate(-181 -130.99812)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M777.41122,701.29557a14.526,14.526,0,0,0-1.49594-20.20536,13.9357,13.9357,0,0,0-1.866-1.351l10.11793-49.6564-23.39032-13.99921-6.50593,69.52164a14.48576,14.48576,0,0,0,3.33874,17.36689,13.86282,13.86282,0,0,0,19.556-1.38555Q777.29042,701.44274,777.41122,701.29557Z'
-                transform='translate(-181 -130.99812)'
-                fill='#a0616a'
-            />
-            <path
-                d='M850.0367,439.55249c-.14673-1.59635-2.1831-2.32114-5.00472-3.3221a30.95043,30.95043,0,0,1-7.99881-3.788c-6.24726-4.51282-3.788-9.23272-.38831-15.75606.66443-1.26845,1.35472-2.58862,2.00188-3.96061,1.92419-4.06416-1.139-20.82982-9.1206-49.83967-2.45922-8.93073-2.45058-15.23835-4.15042-19.42332-1.69126-4.18492-5.08235-6.24721-14.32376-6.67864a42.251,42.251,0,0,0-30.41634,10.81184c-.2243.19845-.45729.41417-.68165.62125q-.6342.59541-1.24254,1.21668a42.21984,42.21984,0,0,0-12.27006,29.02707l-.04318,1.85521-1.90692,88.63448-5.35848,77.77956-7.3862,107.14316,5.47065.26754L779.048,655.228l2.649.12945,6.28171-26.35218,7.55879-31.71065,2.77848-11.64881,8.62012-36.13721,4.021-16.88647,2.45922-10.31136L814.4172,518.1l4.88385-20.46737c.11217-.43144,11.34682-44.75732,24.29858-52.16944C848.43171,442.702,850.17474,441.097,850.0367,439.55249Z'
-                transform='translate(-181 -130.99812)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M995.35486,700.7676a14.526,14.526,0,0,1-.68275-20.24915,13.93416,13.93416,0,0,1,1.71013-1.54361L980.98979,630.6923l21.75153-16.43034,13.93485,68.4208a14.48576,14.48576,0,0,1-1.45425,17.625,13.86281,13.86281,0,0,1-19.59166.72275Q995.49076,700.901,995.35486,700.7676Z'
-                transform='translate(-181 -130.99812)'
-                fill='#a0616a'
-            />
-            <path
-                d='M1005.99554,595.99963S989.031,536.3119,994.20466,529.85028s-10.10621-7.7235-3.52169-19.75745-2.50376-14.04665-2.50376-14.04665l-4.93184-27.66868-8.81682-67.96132-.26734-2.02117-2.56448-19.77218a42.34076,42.34076,0,0,0-16.696-28.61684c-.23307-.17843-.47114-.34983-.70925-.52125q-1.62813-1.17213-3.32715-2.172a42.26284,42.26284,0,0,0-28.10644-5.33868c-18.28571,2.84786-12.59457,9.6669-15.56046,27.935-4.819,29.7033-6.0699,46.69679-3.71994,50.53631.79163,1.2929,1.62357,2.5298,2.41661,3.71735,4.08278,6.129,7.03442,10.55052,1.30465,15.7181a31.25081,31.25081,0,0,1-7.54427,4.6162c-2.69312,1.304-4.64259,2.23962-4.61646,3.84262.03034,1.55289,1.934,2.95523,7.0314,5.179,6.17691,2.70319,12.81133,12.92324,18.24528,23.36341a257.87976,257.87976,0,0,1,11.51823,25.89924L978.53972,634.1368l7.61783,21.422,2.62767-.41552,21.58-3.42133,5.41173-.85662Z'
-                transform='translate(-181 -130.99812)'
-                fill='#e6e6e6'
-            />
-            <circle cx='685.01598' cy='119.90786' r='49.37021' fill='#a0616a' />
-            <path
-                d='M791.9668,223.65966c-7.51577-27.40164,53.10245-50.89908,69.02069-28.8325,2.21348-3.1372,10.39357-5.04277,14.14681-4.08041s6.90971,3.36827,9.98934,5.69723c4.23443,3.2528,8.62274,6.62107,11.22108,11.29814,2.61772,4.65781,3.02183,11.0095-.50033,15.05138-2.79088,3.21427-11.02877,5.293-15.22466,6.13989-2.92551.59666-5.38921,1.07784-7.50637,1.42428a14.35166,14.35166,0,0,1-2.50224-9.48889,11.64829,11.64829,0,0,0-5.08119,8.54579c-.0578.71214,23.65486,87.96006,23.65486,87.96006a51.72045,51.72045,0,0,1-8.95,4.02261,112.54462,112.54462,0,0,1-13.95427-30.41066,97.45318,97.45318,0,0,0,.26949,21.65318,119.14215,119.14215,0,0,0,2.0402,11.991c-28.178,6.0822-70.56044,4.08053-88.99936-21.78781q-.00492-.16051-.00929-.32129a4.40184,4.40184,0,0,0-8.50333-1.535c-.44736,1.14129-.85908,2.12916-1.2265,2.91492C764.48175,284.69268,771.43,262.4621,791.9668,223.65966Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M810.43534,197.849'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M547.1209,394.255a5.09028,5.09028,0,0,1-3.68913-8.8227c.12835-.51014.22043-.8765.34879-1.38664q-.06889-.16659-.13859-.33286c-4.68631-11.17449-20.5687-11.09739-25.21514.09369-4.12406,9.93267-9.37471,19.88226-10.66754,30.38441a40.43982,40.43982,0,0,0,.70989,13.908A162.13182,162.13182,0,0,1,493.7207,360.7605a156.48974,156.48974,0,0,1,.97068-17.45764q.80413-7.128,2.23119-14.15449a164.01022,164.01022,0,0,1,32.52468-69.51166,43.64785,43.64785,0,0,0,18.153-18.83392,33.29478,33.29478,0,0,0,3.02794-9.09827c-.88372.11592-3.3322-13.34322-2.66567-14.169-1.23157-1.86885-3.43587-2.79784-4.78094-4.62148-6.68954-9.06961-15.906-7.486-20.7174,4.83888-10.27825,5.18745-10.37774,13.79049-4.07105,22.06463,4.01234,5.26406,4.56364,12.38693,8.08419,18.02253-.36227.46365-.739.91282-1.10113,1.37648a165.06908,165.06908,0,0,0-17.229,27.29491c1.36864-10.69089-.652-23.57059-4.09675-31.873-3.9213-9.45971-11.27118-17.4267-17.74364-25.6046-7.77447-9.82292-23.71674-5.536-25.08643,6.91609q-.01988.18086-.03883.36158,1.44205.8135,2.8232,1.727a6.90459,6.90459,0,0,1-2.78379,12.56506l-.14079.0217a68.7,68.7,0,0,0,1.81092,10.27183c-8.30336,32.11122,9.623,43.80679,35.21939,44.33218.565.28972,1.11566.57943,1.6807.8548a168.48694,168.48694,0,0,0-9.0693,42.68053,159.80436,159.80436,0,0,0,.11591,25.80246l-.04348-.30421A42.227,42.227,0,0,0,476.379,349.88019c-11.0933-9.11275-26.76607-12.46849-38.73378-19.79336-5.76092-3.526-13.14312,1.03057-12.1377,7.70959q.02407.1599.04895.31988a46.37792,46.37792,0,0,1,5.201,2.50624q1.44205.8137,2.82321,1.727a6.90468,6.90468,0,0,1-2.7838,12.56519l-.14091.02164c-.10138.01448-.18835.029-.28961.04359A68.75993,68.75993,0,0,0,443.014,374.78445c5.19187,28.03184,27.491,30.69143,51.34421,22.52828h.01454a168.43742,168.43742,0,0,0,11.3149,33.03181H546.108c.145-.4491.27532-.91275.40576-1.36192a45.9261,45.9261,0,0,1-11.18457-.66624c2.999-3.67994,5.99787-7.38886,8.99686-11.06867a2.51291,2.51291,0,0,0,.18835-.21734c1.52117-1.88328,3.05689-3.75226,4.57807-5.6356l.00081-.00229a67.27051,67.27051,0,0,0-1.97111-17.13665Z'
-                transform='translate(-181 -130.99812)'
-                fill='#f2f2f2'
-            />
-            <path
-                d='M358.51608,190.97871c-.9174,2.13788-29.20421,1.90879-30.49973,3.81247-9.98865,14.6971,26.0185,77.22955,16.02986,91.94573l-1.39152,1.65842a65.75661,65.75661,0,0,0,9.35956-5.28031c.2669-6.02364.5338-12.06649.81966-18.09013a49.72123,49.72123,0,0,1,7.16746,12.60017c2.32566-1.63933,4.23341-4.67832,14.20146-10.65578A62.515,62.515,0,0,1,388.0165,255.5982Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M379.05672,311.42954a23.4277,23.4277,0,0,0-35.49309,13.858L300.57777,437.29108,226.69,395.34309a22.87524,22.87524,0,1,0-14.27186,22.7523c15.56013,16.60716,83.01892,84.69637,110.94639,58.09227,23.47441-22.362,57.3508-109.02911,65.50546-138.957A23.43014,23.43014,0,0,0,379.05672,311.42954Z'
-                transform='translate(-181 -130.99812)'
-                fill='#ffb6b6'
-            />
-            <polygon
-                points='117.341 349.883 130.906 411.195 230.098 420.167 239.214 366.781 117.341 349.883'
-                fill='#ffb6b6'
-            />
-            <path
-                d='M449.85848,638.108c-.04529.87892-.09966,1.74882-.16309,2.62779-.95143,12.70387-2.03879,27.356-3.21674,43.41254-1.85755,25.4803-3.93262,54.49451-6.02577,84.84978H232.397l40.35026-121.28154s16.29252-8.97031,7.99781-24.03914,8.02627-3.06065,8.1781-24.5811,14.777-44.41548,14.777-44.41548l2.40126-32.58436,1.36826.02721,108.27315,1.80318,1.97535.03623,3.32547,27.9722A123.668,123.668,0,0,1,449.85848,638.108Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M360.25006,281.71589l49.37257-2.54338,22.93226,36.23574s43.14762,7.21373,40.15679,58.05785-52.49777,95.46007-52.49777,95.46007l2.11627-7.25542-3.67543,10.502.47132,5.89689,5.014,6.245.51621,6.4585-4.44235,10.21851v0a5.83891,5.83891,0,0,1-5.53624,8.90024l-71.33958-7.14938c-1.62421-4.468-2.92741-9.81387-9.43789-.91444l-6.72031-.65106-27.93639-2.7066-1.6443-5.98994a23.20469,23.20469,0,0,1,3.77511-20.01465v0a30.60736,30.60736,0,0,1,1.77136-22.98081l.13362-.2677,2.96669-2.47279s-32.45673-48.76167,4.80261-85.48825l14.95415-25.42206,16.73241-20.41434Z'
-                transform='translate(-181 -130.99812)'
-                className='fill-primary'
-            />
-            <circle cx='191.538' cy='82.50891' r='46.81885' fill='#ffb6b6' />
-            <path
-                d='M408.09722,280.95294a22.5482,22.5482,0,0,1-7.54872-12.39051,48.44628,48.44628,0,0,0,.66725,17.68978c-2.53532,1.90623-4.99432,3.71718-7.32,5.35651a49.721,49.721,0,0,0-7.16745-12.60017c-.28587,6.02364-.55277,12.06649-.81967,18.09013a65.7548,65.7548,0,0,1-9.35955,5.2803l1.39151-1.65841c9.98864-14.71618-15.85987-41.80368-5.87123-56.50078,2.80222-4.11753-9.85519-36.65683-19.50073-38.12466-8.57805-1.29623-8.21586,1.18185-16.527,2.9928a41.5889,41.5889,0,0,0-3.98408-10.31279,34.07467,34.07467,0,0,1-.324,10.86555c-2.95464.11437-5.27927-1.76829-8.42552-2.24934-20.07269-3.069-18.033-24.78091-12.44776-31.6625,5.60433-6.8624,24.20918-8.92116,32.78723-11.15154.78151-.76242,2.55429-.28586,3.33591-1.01024,16.43935-15.045,38.42593-17.53657,58.21178-10.86829,21.10012,7.11122,34.25569,20.87458,41.00764,42.0924l1.93652,16.97969c1.73734,5.45956-.44817,30.86221-1.93652,36.39484C440.41735,269.67189,428.096,258.93593,408.09722,280.95294Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M476.36329,502.76855l-1.84316-143.55284a23.17268,23.17268,0,0,0-23.16985-22.80657h-.00007a23.17271,23.17271,0,0,0-23.14009,24.40154l6.486,149.05754.12313,90.73906a20.22151,20.22151,0,1,0,24.8266,7.56106Z'
-                transform='translate(-181 -130.99812)'
-                fill='#ffb6b6'
-            />
-            <path
-                d='M718,688.49187s-2.85,51.27-4.16,80.51H615.7c-2.02-26.61-2.7-46.16-2.7-46.16l-5.37,46.16h-99.2l-6.15-85.49,29.72-73.67s161.06-11.17,176,3C722.93,627.00194,718,688.49187,718,688.49187Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <circle cx='450.21597' cy='72.59881' r='53.60692' fill='#ffb6b6' />
-            <path
-                d='M693.70022,186.178c-5.50615,2.51069-9.07382,20.69444-13.4016,16.4671s-4.03893-9.7286-9.87364-11.2887c-4.71184-1.26523-9.78253.18078-14.08037,2.46133-.7674.40241-1.51139.83995-2.2365,1.27509-.47865-.31981-.955-.62073-1.46639-.89836a17.03314,17.03314,0,0,0-12.5366-1.41014,16.7126,16.7126,0,0,1,8.2586,6.0872c-1.4384.96394-2.89543,1.93037-4.38568,2.78668-12.45014,7.18478-30.16623-8.43822-43.27124-14.3074,8.2086,15.404,4.16668,21.78351-2.96071,37.71453l-.24613.90714c.40491-3.99277-11.95042-6.96835-11.23395-1.34772.5087,3.99065,3.96222,12.86082,5.89652,16.38422-13.77842-15.93173-23.19417-35.42778-21.20733-55.5396.4325-4.22486.30214-8.68257,2.48127-12.3495,2.18143-3.648,6.7293-6.208,10.75272-4.81688a35.542,35.542,0,0,1,53.50674-32.52446c5.14478,3.01886,9.43984,7.28849,14.15516,10.97139,4.71543,3.68287,10.15119,6.89311,16.13038,7.19715,3.79317.18294,7.921-.74312,11.17973,1.22146,3.77491,2.2796,4.42131,7.05234,4.98779,11.64491a17.01391,17.01391,0,0,0-9.16691,3.815,16.648,16.648,0,0,1,10.02242,2.14968,1.03627,1.03627,0,0,1,.16413.09331C687.16438,180.46141,686.95993,182.144,693.70022,186.178Z'
-                transform='translate(-181 -130.99812)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M508.6679,635.777c45.65079-53.43516,30.23132-121.4003,12.11288-190.00758l3.44756-101.35372a31.00393,31.00393,0,0,1,14.02907-24.90184L584.861,289.0678,603,262.84191l50,5,27.72649,29.322,30.69062,23.42985a50.84312,50.84312,0,0,1,18.79205,51.39009L700.83948,504.80436s12.28383,68.85522,4.78852,76.45969-26.80842.00387-11.49291,8.54042c18.37976,10.24451,35.09164,8.35414,25.79051,17.65527s2.47025,22.23225,2.47025,22.23225Z'
-                transform='translate(-181 -130.99812)'
-                className='fill-primary'
-            />
-            <path
-                d='M528.18594,713.43646l7.61052-54.22065-30.36735-8.66909-6.00663,57.61172a18.32876,18.32876,0,1,0,28.76346,5.278Z'
-                transform='translate(-181 -130.99812)'
-                fill='#ffb6b6'
-            />
-            <path
-                d='M586,298.84191s-86.4384,28.13284-88.4995,48.71733S496.12293,380.74489,487,463.11505l.48061,221.88319s57.95172,12.40752,52.0085-.8585-3.86948-19.50742-3.86948-19.50742,12.42491-4.90311,9.31468-20.47947,12.48716-92.76777,12.48716-92.76777L567.266,412.94623Z'
-                transform='translate(-181 -130.99812)'
-                className='fill-primary'
-            />
-            <path
-                d='M692.58051,710.565l-4.05011-54.60215,30.86907-6.66559,2.228,57.88114a18.32876,18.32876,0,1,1-29.047,3.3866Z'
-                transform='translate(-181 -130.99812)'
-                fill='#ffb6b6'
-            />
-            <path
-                d='M661.99007,293.07805s84.41465,33.72272,85.12583,54.39792-.79456,33.20472,2.92474,115.99505l-14.983,221.37726s-58.63881,8.593-51.84116-4.2562,5.1363-19.21278,5.1363-19.21278-12.07784-5.70478-7.95612-21.04453S674,546.94918,674,546.94918l-.77441-138.78628Z'
-                transform='translate(-181 -130.99812)'
-                className='fill-primary'
-            />
-        </svg>
-    )
-}
-
-export default PeopleSVG
