--- src/components/svg/PieChartSVG.jsx
@@ -1,107 +1,0 @@
-const PieChartSVG = ({ width, height }) => {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            data-name='Layer 1'
-            width={width}
-            height={height}
-            viewBox='0 0 312.32533 689.35327'
-        >
-            <path
-                d='M756.73957,798.06625h-121.03l-1.18-5.1-6.47-28.06-6.46,28.06-1.17,5.1h-124.5c-.27-1.72-.53-3.42-.79-5.1-3.63-23.17-7.22-42.46-10.26-47.98-8.82-15.98-.6-5.97,8.81-15.39,9.41-9.41,0,0-5.51-12.08-5.52-12.08,24.37-103.88,24.37-103.88l203.9,3.01s37.66,65.89,38.05,74.98c.22,5.17,1.27,57.32,2.13,101.34C756.66956,794.67623,756.69953,796.38619,756.73957,798.06625Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M637.8194,677.4775s86.78375,39.925,100.456,9.29968,6.90448-71.56005-1.56573-76.91809-17.77167-30.04309-5.38013-47.1029-7.09991-45.88031-12.55743-48.82343c-4.15564-2.241-15.46167-128.52191-20.61933-188.64374a48.38235,48.38235,0,0,0-32.23834-41.53308h0L645.39905,250.6226l-52.62915-.47348L566.478,286.60073l-37.9447,14.19278L500.05069,418.42129s-.275,51.957,5.71109,107.43243c1.86252,17.26062,4.33118,34.86188,7.60285,51.34473,7.706,38.823-25.86032,70.79376-6.67206,78.14258C560.21076,675.83761,637.8194,677.4775,637.8194,677.4775Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M528.88448,601.84869l166.95367-6a8.63685,8.63685,0,0,1,8.63686,8.63684v83.71454a8.63683,8.63683,0,0,1-8.63686,8.63684H538.88448a8.63683,8.63683,0,0,1-8.63685-8.63684c-9.255-23.5864-16.99663-46.545-10-77.71454A8.63685,8.63685,0,0,1,528.88448,601.84869Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#3f3d56'
-            />
-            <rect x='103.11943' y='503.42215' width='137.38036' height='2' fill='#2f2e41' />
-            <rect x='103.12012' y='532.30344' width='137.37998' height='4.28271' fill='#2f2e41' />
-            <rect x='103.11966' y='442.02643' width='126.00779' height='61.48572' fill='#6c63ff' />
-            <rect
-                x='103.11966'
-                y='442.02643'
-                width='126.00779'
-                height='61.48572'
-                opacity='0.2'
-                style={{ isolation: 'isolate' }}
-            />
-            <rect x='142.21244' y='442.02643' width='86.91502' height='61.48572' fill='#6c63ff' />
-            <path
-                d='M659.959,561.0071H642.7959a.59381.59381,0,0,1,0-1.18762H659.959a.59381.59381,0,0,1,0,1.18762Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#fff'
-            />
-            <path
-                d='M575.32117,603.89529H558.15808a.59382.59382,0,0,1,0-1.18763h17.16309a.59381.59381,0,0,1,0,1.18763Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#fff'
-            />
-            <path
-                d='M608.34131,561.0071H555.53473a.61747.61747,0,0,1-.619-.49353.55931.55931,0,0,1,.58558-.6651h53.12207a.6176.6176,0,0,1-.28205,1.15863Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#fff'
-            />
-            <path
-                d='M692.98249,695.92709,586.32056,719.582a12.85728,12.85728,0,0,1-12.85724-12.85724c-10.47538-26.79837-10.65282-51.91853,0-75.27374a12.85728,12.85728,0,0,1,12.85724-12.85724l106.66193-23.6549a12.85724,12.85724,0,0,1,12.85723,12.85721c9.85742,28.11328,18.00751,55.70313,0,75.27374A12.85729,12.85729,0,0,1,692.98249,695.92709Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M596.66327,634.57033,598.02828,643l69.6-1.81462v0c.23338-4.41383-31.37986-7.71824-64.90238-6.784Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#3f3d56'
-            />
-            <path
-                d='M596.66327,655.57033,598.02828,664l69.6-1.81462v0c.23338-4.41383-31.37986-7.71824-64.90238-6.784Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#3f3d56'
-            />
-            <path
-                d='M596.66327,680.57033,598.02828,689l69.6-1.81462v0c.23338-4.41383-31.37986-7.71824-64.90238-6.784Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#3f3d56'
-            />
-            <path
-                d='M719.04871,508.25671l-1.96136,18.81839-.45978,4.35223-.70484,6.98785L700.78229,683.65987a19.77416,19.77416,0,0,1-8.67364,22.5575,18.96429,18.96429,0,0,1-25.99012-6.5589c-.092-.1532-.15338-.3064-.24524-.45959a19.88476,19.88476,0,0,1,6.8653-26.84839,20.14915,20.14915,0,0,1,2.85034-1.34857l-.12256-4.07636-.36779-12.78052-.27594-9.65442L674.2403,625.274l-.2146-7.44782-.7049-24.15106-.67419-23.29321-.82758-28.22742-.64373-21.6994,7.81543-1.99225Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#a0616a'
-            />
-            <path
-                d='M734.18921,483.70892c-.70483,3.59253-1.34857,5.914-1.80816,6.66-3.9538,6.19043-6.375-2.45941-4.19886,14.12173,1.747,13.45835-8.735,11.77243-4.96508,14.64664,6.22161,4.75318-2.36,6.85346-6.12976,9.42352-1.68586,1.13312,2.98174.864,4.91265,2.43919.91858.71967-3.44384,2.79952-2.36889,3.25362,4.107,1.93451,1.747,3.39924-3.70838,4.53223-9.04144,1.8515-26.542,2.76355-37.11584,3.1228-3.00354.13813-5.45538.19349-6.98779.24878-1.134.02753-1.77771.02753-1.77771.02753s-2.66645-1.71338-3.76972-6.93628c-.79693-3.64789-.79693-9.03674,1.40973-16.7746,1.34857-4.75332,3.64722-8.1524,4.99579-10.88827a7.18976,7.18976,0,0,0,1.04211-4.25592v-.02768c-.2146-1.40924-1.226-2.70819-3.52472-4.08994-4.13751-2.48706-1.83886-7.71017,1.37922-12.6293,3.06488-4.615,6.92663-9.00912,6.92663-10.61182,0-3.39923-25.49988-75.63751-25.49988-75.63751L662.379,291.25741a36.07551,36.07551,0,0,1,24.45764-6.10727c5.91541.77368,12.50476,3.26092,19.2782,8.84311,5.91522,4.89148,11.98358,12.18722,17.807,22.74386C748.6861,361.589,739.46088,457.23434,734.18921,483.70892Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#e6e6e6'
-            />
-            <circle cx='166.11047' cy='74.33239' r='54.73784' fill='#a0616a' />
-            <path
-                d='M667.48944,210.89437s-15.18865,12.40323-23.51709,16.02173c2.30585-8.00812,4.70838-21.75676,7.00782-29.74213,1.19574-4.20175,2.417-8.49492,2.22247-12.83758-.211-4.37193-2.08282-8.88634-5.78925-11.15169a10.0547,10.0547,0,0,0-14.60894,12.26892c-3.62805-.24716-5.677-4.392-6.12116-8.01553-.44421-3.62372-.10785-7.57187-2.11053-10.644-1.93042-2.978-5.74182-4.336-9.31177-4.17271-3.54083.14691-6.90338,1.60035-10.07574,3.25485-9.07154,4.693-17.30988,11.171-27.07019,14.26726a32.90308,32.90308,0,0,1-6.00843,1.30764,25.779,25.779,0,0,0-5.54211-16.28318,44.45273,44.45273,0,0,1-4.571,15.78131,18.86245,18.86245,0,0,1-11.69421-7.47393c-3.055-4.42526-3.80634-10.13075-3.08752-15.45118a36.53516,36.53516,0,0,1,1.77649-7.16989,22.52994,22.52994,0,0,0,4.1748-1.449,24.51589,24.51589,0,0,0,7.86627-5.49725,40.18032,40.18032,0,0,1-9.015-.63959c.07385-.17648.14135-.3302.21527-.50682,1.70111-3.7158,3.58392-7.55336,6.86865-9.96522,3.28467-2.41173,8.37238-2.89155,11.2254.04786a52.08774,52.08774,0,0,1,71.90845-15.98247q.38984.248.77515.50294a24.75864,24.75864,0,0,0-4.80207,8.29846,42.21844,42.21844,0,0,1,10.21961-3.683c2.5,2.79165,4.08038,6.14056,3.40253,9.72244,4.14514,2.01787,12.22863,5.38,8.40777,7.59306a53.38331,53.38331,0,0,1,7.25415,71.59871Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#2f2e41'
-            />
-            <path
-                d='M599.628,586.88137a18.96931,18.96931,0,0,1-20.1289,17.70917,4.20385,4.20385,0,0,0-.51392-.06433,19.90236,19.90236,0,0,1-18.2445-20.8783,21.3391,21.3391,0,0,1,.51391-3.10486l-3.42615-2.20569-10.74969-6.95947-8.0943-5.22491-16.1673-10.42847-6.25287-4.02576-.83514-.53534-19.465-12.59124-15.632-10.08576-1.32764-.85663-2.61255-1.69165-23.72638-15.31082-6.38123-4.11139c-.02145-.02142,0-.02142,0-.02142-.21408-.40686,13.21219-7.06647,12.99811-7.47336-3.38339-6.70248-7.66613-12.33429-2.14135-36.40323l-5.63178-10.12869,14.39,12.29157,3.31909,2.84793,5.35342,4.53967,66.211,56.48929L585.966,566.94534A19.79572,19.79572,0,0,1,599.628,586.88137Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#a0616a'
-            />
-            <path
-                d='M587.50776,379.36157,513.26661,493.51785,543.09583,526.859l-2.01288,1.79871-25.35382,22.5058-7.17352,6.35992-13.27655-2.69818s-2.63385-4.00433-16.48856-6.06012c-13.85458-2.05579-5.78159-6.76666.34268-15.97461,1.45612-2.20557-4.19711,1.26349-4.56112.49261-.17133-.47119-3.59748-2.0343-4.28275-2.14135-2.82663-.53534-.0643-2.14136-4.13281-2.97657-6.53116-1.32763-6.91657,3.29767-19.57215-21.6492-.02145-.02142,0-.02142,0-.02142-.21408-.40686-.40686-.7923-.62094-1.19916-3.38339-6.70248-.9422-25.33239,4.58255-49.40133,8.43686-36.70307,24.02607-86.0401,37.00275-124.82026a55.95274,55.95274,0,1,1,99.959,48.28782Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#e6e6e6'
-            />
-            <path
-                d='M617.01923,672.38423a1.52141,1.52141,0,0,0,1.423-2.47885l-.05172-.42421.05352-.09354a4.08563,4.08563,0,0,1,7.47411.968c.85264,3.09933,2.0387,6.24569,2.0305,9.4083a12.08649,12.08649,0,0,1-.72914,4.098,48.45968,48.45968,0,0,0,6.88456-19.41934,46.7746,46.7746,0,0,0,.36307-5.21325q.02739-2.14377-.13386-4.28075a49.02015,49.02015,0,0,0-7.0533-21.82651,13.04571,13.04571,0,0,1-4.681-6.2621,9.94931,9.94931,0,0,1-.55868-2.811c.25774.06731,1.48572-3.83269,1.31887-4.10245.43492-.50826,1.12324-.70158,1.59012-1.19226,2.322-2.44017,4.99608-1.6269,5.96335,2.20746,2.85458,1.92158,2.5633,4.47657.38455,6.69506-1.38614,1.41148-1.81523,3.5032-3.0694,5.04315.09015.151.18509.29828.27522.44923a49.33912,49.33912,0,0,1,4.09152,8.73679,20.51008,20.51008,0,0,1,2.40339-9.29924c1.51559-2.659,3.9923-4.74763,6.21663-6.93143a4.20968,4.20968,0,0,1,7.1815,2.98638l-.002.10867q-.4579.18742-.90163.40685a2.06369,2.06369,0,0,0,.357,3.83l.04094.01167a20.53192,20.53192,0,0,1-.92,2.97861c1.265,9.83223-4.48714,12.63213-12.09736,11.83349-.17836.06485-.35249.1302-.53029.1908a50.3585,50.3585,0,0,1,1.09807,12.9951,47.75685,47.75685,0,0,1-.99651,7.64744l.02428-.0886a12.62152,12.62152,0,0,1,5.184-6.69321c3.62954-2.28879,8.40241-2.69951,12.22461-4.42545a2.36864,2.36864,0,0,1,3.312,2.73888l-.02644.093a13.85692,13.85692,0,0,0-1.63579.5493q-.458.18755-.90164.40685a2.06369,2.06369,0,0,0,.357,3.83l.041.01165.08429.02374a20.55119,20.55119,0,0,1-4.48912,5.40142c-2.58486,8.11928-9.29686,8.0765-16.06621,4.76621l-.00437-.00052a50.343,50.343,0,0,1-4.58715,9.3737l-11.98671-1.50717c-.02628-.13861-.04764-.281-.06955-.419a13.7266,13.7266,0,0,0,3.34166.2195c-.75216-1.20311-1.50317-2.4148-2.25532-3.6179a.742.742,0,0,1-.04777-.07147c-.3809-.61518-.76662-1.22672-1.14753-1.84195l-.00015-.00072a20.10637,20.10637,0,0,1,1.22349-5.00838Z'
-                transform='translate(-444.41424 -108.71297)'
-                fill='#f2f2f2'
-            />
-        </svg>
-    )
-}
-
-export default PieChartSVG
