--- src/components/svg/TooltipTriangleSVG.jsx
@@ -1,28 +1,0 @@
-const TooltipTriangleSVG = () => {
-    return (
-        <svg
-            className='absolute left-0 -ml-2 bottom-0 top-0 h-full'
-            width='9px'
-            height='16px'
-            viewBox='0 0 9 16'
-            version='1.1'
-            xmlns='http://www.w3.org/2000/svg'
-        >
-            <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-                <g id='Tooltips-' transform='translate(-874.000000, -1029.000000)' fill='#FFFFFF'>
-                    <g id='Group-3-Copy-16' transform='translate(850.000000, 975.000000)'>
-                        <g id='Group-2' transform='translate(24.000000, 0.000000)'>
-                            <polygon
-                                id='Triangle'
-                                transform='translate(4.500000, 62.000000) rotate(-90.000000) translate(-4.500000, -62.000000) '
-                                points='4.5 57.5 12.5 66.5 -3.5 66.5'
-                            ></polygon>
-                        </g>
-                    </g>
-                </g>
-            </g>
-        </svg>
-    )
-}
-
-export default TooltipTriangleSVG
